@@ -9,11 +9,13 @@
         <DebugType>embedded</DebugType>
         <Platforms>AnyCPU;x64</Platforms>
         <RuntimeIdentifiers>win-x64;win-x86</RuntimeIdentifiers>
-        <LangVersion>9</LangVersion>
+        <LangVersion>10</LangVersion>
         <ApplicationIcon>icon.ico</ApplicationIcon>
         <GenerateDocumentationFile>True</GenerateDocumentationFile>
         <IncludeAllContentForSelfExtract>true</IncludeAllContentForSelfExtract>
         <StartupObject>UndertaleModTool.Program</StartupObject>
+        <!-- Suppress all missing XML comment warnings -->
+        <NoWarn>1591</NoWarn>
     </PropertyGroup>
     <ItemGroup>
       <None Remove="Resources\tabs_left_button.png" />
@@ -78,7 +80,7 @@
             <Version>1.3.1</Version>
         </PackageReference>
         <PackageReference Include="Newtonsoft.Json">
-            <Version>13.0.1</Version>
+            <Version>13.0.2</Version>
         </PackageReference>
         <PackageReference Include="Ookii.Dialogs.Wpf" Version="5.0.1" />
         <PackageReference Include="PropertyChanged.Fody" Version="3.3.3" />
