@@ -1,5 +1,6 @@
 using System;
 using System.Collections;
+using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.Diagnostics;
@@ -63,6 +64,8 @@ namespace UndertaleModLib.Decompiler
         public GlobalDecompileContext GlobalContext;
         public UndertaleCode TargetCode;
         public UndertaleGameObject Object;
+        public static bool GMS2_3;
+        public bool AssetResolutionEnabled => !GlobalContext.Data.IsVersionAtLeast(2023, 8);
 
         public DecompileContext(GlobalDecompileContext globalContext, UndertaleCode code, bool computeObject = true)
         {
@@ -72,21 +75,32 @@ namespace UndertaleModLib.Decompiler
             if (code.ParentEntry != null)
                 throw new InvalidOperationException("This code block represents a function nested inside " + code.ParentEntry.Name + " - decompile that instead");
 
-            if (computeObject && globalContext.Data != null)
+            if (computeObject && globalContext.Data is not null)
             {
                 // TODO: This is expensive, move it somewhere else as a dictionary
                 // and have it update when events/objects are modified.
-                foreach (var obj in globalContext.Data.GameObjects)
-                    foreach (var event_list in obj.Events)
-                        foreach (var subevent in event_list)
-                            foreach (var ev in subevent.Actions)
-                                if (ev.CodeId == code)
-                                {
-                                    Object = obj;
-                                    goto LoopEnd;
-                                }
+                
+                // Currently using for loops on purpose, as foreach has memory issues due to IEnumerable
+                for (int i = 0; i < globalContext.Data.GameObjects.Count; i++)
+                {
+                    UndertaleGameObject obj = globalContext.Data.GameObjects[i];
+                    for (int j = 0; j < obj.Events.Count; j++)
+                    {
+                        var eventList = obj.Events[j];
+                        for (int k = 0; k < eventList.Count; k++)
+                        {
+                            UndertaleGameObject.Event subEvent = eventList[k];
+                            for (int l = 0; l < subEvent.Actions.Count; l++)
+                            {
+                                UndertaleGameObject.EventAction ev = subEvent.Actions[l];
+                                if (ev.CodeId != code) continue;
+                                Object = obj;
+                                return;
+                            }
+                        }
+                    }
+                }
             }
-        LoopEnd: return;
         }
 
         #region Struct management
@@ -257,7 +271,7 @@ namespace UndertaleModLib.Decompiler
                     UndertaleInstruction.Opcode.Not => "~",
                     UndertaleInstruction.Opcode.Shl => "<<",
                     UndertaleInstruction.Opcode.Shr => ">>",
-                    _ => op.ToString().ToUpper(),
+                    _ => op.ToString().ToUpper(CultureInfo.InvariantCulture),
                 };
             }
 
@@ -272,7 +286,7 @@ namespace UndertaleModLib.Decompiler
                     UndertaleInstruction.ComparisonType.NEQ => "!=",
                     UndertaleInstruction.ComparisonType.GTE => ">=",
                     UndertaleInstruction.ComparisonType.GT => ">",
-                    _ => op.ToString().ToUpper(),
+                    _ => op.ToString().ToUpper(CultureInfo.InvariantCulture),
                 };
             }
 
@@ -379,7 +393,7 @@ namespace UndertaleModLib.Decompiler
                 {
                     int? val = ConvertToInt(Value);
                     if (val != null && val < 0 && val >= -16)
-                        return ((UndertaleInstruction.InstanceType)Value).ToString().ToLower();
+                        return ((UndertaleInstruction.InstanceType)Value).ToString().ToLower(CultureInfo.InvariantCulture);
                 }
                 else switch (AssetType) // Need to put else because otherwise it gets terribly unoptimized with GameObject type
                 {
@@ -488,7 +502,7 @@ namespace UndertaleModLib.Decompiler
                         break;*/
                 }
 
-                if (context.GlobalContext.Data != null && AssetType != AssetIDType.Other)
+                if ((context.AssetResolutionEnabled || AssetType == AssetIDType.Script) && context.GlobalContext.Data != null && AssetType != AssetIDType.Other)
                 {
                     IList assetList = null;
                     switch (AssetType)
@@ -573,6 +587,128 @@ namespace UndertaleModLib.Decompiler
             }
         }
 
+        // Represents a reference to an asset in the resource tree, used in 2023.8+ only
+        public class ExpressionAssetRef : Expression
+        {
+            // NOTE: Also see generalized "ResourceType" enum. This has slightly differing values, though
+            public enum RefType
+            {
+                Object = 0,
+                Sprite = 1,
+                Sound = 2,
+                Room = 3,
+                Background = 4,
+                Path = 5,
+                // missing 6
+                Font = 7,
+                Timeline = 8,
+                // missing 9
+                Shader = 10,
+                Sequence = 11,
+                AnimCurve = 12,
+                ParticleSystem = 13
+            }
+
+            public int AssetIndex;
+            public RefType AssetRefType;
+
+            public ExpressionAssetRef(int encodedResourceIndex)
+            {
+                Type = UndertaleInstruction.DataType.Variable;
+
+                // Break down index - first 24 bits are the ID, the rest is the ref type
+                AssetIndex = encodedResourceIndex & 0xffffff;
+                AssetRefType = (RefType)(encodedResourceIndex >> 24);
+            }
+
+            public ExpressionAssetRef(int resourceIndex, RefType resourceType)
+            {
+                Type = UndertaleInstruction.DataType.Variable;
+                AssetIndex = resourceIndex;
+                AssetRefType = resourceType;
+            }
+
+            internal override bool IsDuplicationSafe()
+            {
+                return true;
+            }
+
+            public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
+            {
+                return this;
+            }
+            public override string ToString(DecompileContext context)
+            {
+                if (context.GlobalContext.Data != null)
+                {
+                    IList assetList = null;
+                    switch (AssetRefType)
+                    {
+                        case RefType.Sprite:
+                            assetList = (IList)context.GlobalContext.Data.Sprites;
+                            break;
+                        case RefType.Background:
+                            assetList = (IList)context.GlobalContext.Data.Backgrounds;
+                            break;
+                        case RefType.Sound:
+                            assetList = (IList)context.GlobalContext.Data.Sounds;
+                            break;
+                        case RefType.Font:
+                            assetList = (IList)context.GlobalContext.Data.Fonts;
+                            break;
+                        case RefType.Path:
+                            assetList = (IList)context.GlobalContext.Data.Paths;
+                            break;
+                        case RefType.Timeline:
+                            assetList = (IList)context.GlobalContext.Data.Timelines;
+                            break;
+                        case RefType.Room:
+                            assetList = (IList)context.GlobalContext.Data.Rooms;
+                            break;
+                        case RefType.Object:
+                            assetList = (IList)context.GlobalContext.Data.GameObjects;
+                            break;
+                        case RefType.Shader:
+                            assetList = (IList)context.GlobalContext.Data.Shaders;
+                            break;
+                        case RefType.AnimCurve:
+                            assetList = (IList)context.GlobalContext.Data.AnimationCurves;
+                            break;
+                        case RefType.Sequence:
+                            assetList = (IList)context.GlobalContext.Data.Sequences;
+                            break;
+                        case RefType.ParticleSystem:
+                            assetList = (IList)context.GlobalContext.Data.ParticleSystems;
+                            break;
+                    }
+
+                    if (assetList != null && AssetIndex >= 0 && AssetIndex < assetList.Count)
+                        return ((UndertaleNamedResource)assetList[AssetIndex]).Name.Content;
+                }
+                return $"/* ERROR: missing {AssetRefType} asset, using ID instead */ {AssetIndex}";
+            }
+            internal override AssetIDType DoTypePropagation(DecompileContext context, AssetIDType suggestedType)
+            {
+                // Convert type to corresponding AssetIDType equivalent
+                return AssetRefType switch
+                {
+                    RefType.Object => AssetIDType.GameObject,
+                    RefType.Sprite => AssetIDType.Sprite,
+                    RefType.Sound => AssetIDType.Sound,
+                    RefType.Room => AssetIDType.Room,
+                    RefType.Background => AssetIDType.Background,
+                    RefType.Path => AssetIDType.Path,
+                    RefType.Font => AssetIDType.Font,
+                    RefType.Timeline => AssetIDType.Timeline,
+                    RefType.Shader => AssetIDType.Shader,
+                    RefType.Sequence => AssetIDType.Sequence,
+                    RefType.AnimCurve => AssetIDType.AnimCurve,
+                    RefType.ParticleSystem => AssetIDType.ParticleSystem,
+                    _ => throw new NotImplementedException($"Missing ref type {AssetRefType}")
+                };
+            }
+        }
+
         // Represents an expression converted to one of another data type - makes no difference on high-level code.
         public class ExpressionCast : Expression
         {
@@ -636,7 +772,7 @@ namespace UndertaleModLib.Decompiler
                 if (Opcode == UndertaleInstruction.Opcode.Not && Type == UndertaleInstruction.DataType.Boolean)
                     op = "!"; // This is a logical negation instead, see #93
                 string arg = Argument.ToString(context);
-                if (arg.Contains(" "))
+                if (arg.Contains(' ', StringComparison.InvariantCulture))
                     return String.Format("({0}({1}))", op, arg);
                 return String.Format("({0}{1})", op, arg);
             }
@@ -1078,7 +1214,7 @@ namespace UndertaleModLib.Decompiler
                 {
                     if (AssetTypeResolver.return_types.ContainsKey(context.TargetCode.Name.Content))
                         Value.DoTypePropagation(context, AssetTypeResolver.return_types[context.TargetCode.Name.Content]);
-                    if (context.GlobalContext.Data != null && !context.GlobalContext.Data.GMS2_3)
+                    if (context.GlobalContext.Data != null && !DecompileContext.GMS2_3)
                     {
                         // We might be decompiling a legacy script - resolve it's name
                         UndertaleScript script = context.GlobalContext.Data.Scripts.FirstOrDefault(x => x.Code == context.TargetCode);
@@ -1087,7 +1223,7 @@ namespace UndertaleModLib.Decompiler
                     }
 
                     string cleanVal = Value.ToString(context);
-                    if (cleanVal.EndsWith("\n"))
+                    if (cleanVal.EndsWith("\n", StringComparison.InvariantCulture))
                         cleanVal = cleanVal.Substring(0, cleanVal.Length - 1);
 
                     return "return " + cleanVal + ";";
@@ -1126,7 +1262,7 @@ namespace UndertaleModLib.Decompiler
                     {
                         try
                         {
-                            if (Destination.Var.Name.Content.StartsWith("___struct___"))
+                            if (Destination.Var.Name.Content.StartsWith("___struct___", StringComparison.InvariantCulture))
                             {
                                 Expression val = Value;
                                 while (val is ExpressionCast cast)
@@ -1409,7 +1545,21 @@ namespace UndertaleModLib.Decompiler
                             if (kvp.Key != null)
                                 sb.Append(kvp.Key);
                             else
-                                sb.Append((context.Statements[0].Last() as AssignmentStatement).Destination.Var.Name.Content);
+                            {
+                                //Attempt to find function names before going with the last functions' name
+                                bool gotFuncName = false;
+                                if (Function.Name.Content.StartsWith("gml_Script_"))
+                                {
+                                    string funcName = Function.Name.Content.Substring("gml_Script_".Length);
+                                    if (context.Statements[0].Any(x => x is AssignmentStatement && (x as AssignmentStatement).Destination.Var.Name.Content == funcName))
+                                    {
+                                        sb.Append(funcName);
+                                        gotFuncName = true;
+                                    }
+                                }
+                                if(!gotFuncName)
+                                    sb.Append((context.Statements[0].Last() as AssignmentStatement).Destination.Var.Name.Content);
+                            }
                         }
                         sb.Append("(");
                         for (int i = 0; i < FunctionBodyCodeEntry.ArgumentsCount; ++i)
@@ -1552,9 +1702,9 @@ namespace UndertaleModLib.Decompiler
                     else
                         constructor = Arguments[0].ToString(context);
 
-                    if (constructor.StartsWith("gml_Script_"))
+                    if (constructor.StartsWith("gml_Script_", StringComparison.InvariantCulture))
                         constructor = constructor.Substring(11);
-                    if (constructor.EndsWith(context.TargetCode.Name.Content))
+                    if (constructor.EndsWith(context.TargetCode.Name.Content, StringComparison.InvariantCulture))
                         constructor = constructor.Substring(0, constructor.Length - context.TargetCode.Name.Content.Length - 1);
 
                     if (AssetTypeResolver.builtin_funcs.TryGetValue(constructor, out AssetIDType[] types))
@@ -1616,8 +1766,6 @@ namespace UndertaleModLib.Decompiler
 
                     return ret;
                 }
-
-
             }
 
             public override Statement CleanStatement(DecompileContext context, BlockHLStatement block)
@@ -1808,7 +1956,7 @@ namespace UndertaleModLib.Decompiler
                 string name = Var.Name.Content;
                 if (ArrayIndices != null)
                 {
-                    if (context?.GlobalContext.Data?.GMS2_3 == true)
+                    if (DecompileContext.GMS2_3 == true)
                     {
                         if (name == "argument" && context.DecompilingStruct && context.ArgumentReplacements != null && ArrayIndices.Count == 1)
                         {
@@ -1854,7 +2002,7 @@ namespace UndertaleModLib.Decompiler
                             if (constant.AssetType == AssetIDType.GameObject && val < 0)
                             {
                                 UndertaleInstruction.InstanceType instanceType = (UndertaleInstruction.InstanceType)val;
-                                prefix = (instanceType == UndertaleInstruction.InstanceType.Global || instanceType == UndertaleInstruction.InstanceType.Other) ? prefix.ToLower() : "";
+                                prefix = (instanceType == UndertaleInstruction.InstanceType.Global || instanceType == UndertaleInstruction.InstanceType.Other) ? prefix.ToLower(CultureInfo.InvariantCulture) : "";
                             }
                         }
                     }
@@ -2100,17 +2248,38 @@ namespace UndertaleModLib.Decompiler
 
                     case UndertaleInstruction.Opcode.Ret:
                     case UndertaleInstruction.Opcode.Exit:
-                        ReturnStatement stmt = new ReturnStatement(instr.Kind == UndertaleInstruction.Opcode.Ret ? stack.Pop() : null);
-                        /*
-                            This shouldn't be necessary: all unused things on the stack get converted to tempvars at the end anyway, and this fixes decompilation of repeat()
-                            See #85
+                        // 2.3 scripts add exits to every script, even those that lack a return
+                        // This detects that type of exit using the next block.
+                        Block nextBlock = null;
+                        if (DecompileContext.GMS2_3 && instr.Kind == UndertaleInstruction.Opcode.Exit)
+                        {
+                            uint[] blockAddresses = blocks.Keys.ToArray();
+                            Array.Sort(blockAddresses);
+                            int nextBlockIndex = Array.IndexOf(blockAddresses, block.Address ?? 0) + 1;
+                            if (blockAddresses.Length > nextBlockIndex)
+                            {
+                                uint nextBlockAddress = blockAddresses[nextBlockIndex];
+                                nextBlock = blocks[nextBlockAddress];
+                            }
+                        }
 
-                            foreach (var expr in stack.Reverse())
-                                if (!(expr is ExpressionTempVar))
-                                    statements.Add(expr);
-                            stack.Clear();
-                        */
-                        statements.Add(stmt);
+                        if (!(nextBlock is not null
+                            && nextBlock.Instructions.Count > 0
+                            && nextBlock.Instructions[0].Kind == UndertaleInstruction.Opcode.Push
+                            && nextBlock.Instructions[0].Value.GetType() != typeof(int)))
+                        {
+                            ReturnStatement stmt = new ReturnStatement(instr.Kind == UndertaleInstruction.Opcode.Ret ? stack.Pop() : null);
+                            /*
+                                This shouldn't be necessary: all unused things on the stack get converted to tempvars at the end anyway, and this fixes decompilation of repeat()
+                                See #85
+
+                                foreach (var expr in stack.Reverse())
+                                    if (!(expr is ExpressionTempVar))
+                                        statements.Add(expr);
+                                stack.Clear();
+                            */
+                            statements.Add(stmt);
+                        }
                         end = true;
                         returned = true;
                         break;
@@ -2158,7 +2327,7 @@ namespace UndertaleModLib.Decompiler
                         break;
 
                     case UndertaleInstruction.Opcode.PushEnv:
-                        if (context.GlobalContext.Data?.GMS2_3 == true)
+                        if (DecompileContext.GMS2_3 == true)
                         {
                             Expression expr = stack.Pop();
 
@@ -2469,7 +2638,7 @@ namespace UndertaleModLib.Decompiler
                             if (callTargetBody != null && callTargetBody.ParentEntry != null && !context.DisableAnonymousFunctionNameResolution)
                             {
                                 // Special case: this is a direct reference to a method variable
-                                // Figure out what it's actual name is
+                                // Figure out what its actual name is
 
                                 static string FindActualNameForAnonymousCodeObject(DecompileContext context, UndertaleCode anonymousCodeObject)
                                 {
@@ -2481,6 +2650,7 @@ namespace UndertaleModLib.Decompiler
                                         Dictionary<uint, Block> blocks2 = PrepareDecompileFlow(anonymousCodeObject.ParentEntry, new List<uint>() { 0 });
                                         DecompileFromBlock(childContext, blocks2, blocks2[0]);
                                         // This hack handles decompilation of code entries getting shorter, but not longer or out of place.
+                                        // Probably is no longer needed since we now update Length mostly-correctly
                                         Block lastBlock;
                                         if (!blocks2.TryGetValue(anonymousCodeObject.Length / 4, out lastBlock))
                                             lastBlock = blocks2[blocks2.Keys.Max()];
@@ -2488,7 +2658,8 @@ namespace UndertaleModLib.Decompiler
                                         foreach (Statement stmt2 in statements)
                                         {
                                             if (stmt2 is AssignmentStatement assign &&
-                                                assign.Value is FunctionDefinition funcDef)
+                                                assign.Value is FunctionDefinition funcDef &&
+                                                funcDef.FunctionBodyCodeEntry == anonymousCodeObject)
                                             {
                                                 if (funcDef.FunctionBodyEntryBlock.Address == anonymousCodeObject.Offset / 4)
                                                     return assign.Destination.Var.Name.Content;
@@ -2535,7 +2706,7 @@ namespace UndertaleModLib.Decompiler
 
                     case UndertaleInstruction.Opcode.Break:
                         // GMS 2.3 sub-opcodes
-                        if (context.GlobalContext.Data?.GMS2_3 == true)
+                        if (DecompileContext.GMS2_3 == true)
                         {
                             switch ((short)instr.Value)
                             {
@@ -2632,6 +2803,9 @@ namespace UndertaleModLib.Decompiler
                                      */
 
                                     // Note that this operator peeks from the stack, it does not pop directly.
+                                    break;
+                                case -11: // GM 2023.8+, pushref
+                                    stack.Push(new ExpressionAssetRef(instr.IntArgument));
                                     break;
                             }
                         }
@@ -3269,7 +3443,7 @@ namespace UndertaleModLib.Decompiler
                 if (IsForLoop)
                 {
                     string conditionStr = Condition.ToString(context); // Cut off parenthesis for the condition.
-                    if (conditionStr.StartsWith("(") && conditionStr.EndsWith(")"))
+                    if (conditionStr.StartsWith("(", StringComparison.InvariantCulture) && conditionStr.EndsWith(")", StringComparison.InvariantCulture))
                         conditionStr = conditionStr.Substring(1, conditionStr.Length - 2);
 
                     return "for (" + InitializeStatement.ToString(context) + "; " + conditionStr + "; " + StepStatement.ToString(context) + ")\n" + context.Indentation + Block.ToString(context);
@@ -3463,7 +3637,7 @@ namespace UndertaleModLib.Decompiler
 
                     Block b = blockList[i];
 
-                    IEnumerable<Block> e;
+                    Block[] e;
                     if (b.conditionalExit)
                     {
                         reverseUse2[0] = b.nextBlockTrue;
@@ -3654,7 +3828,7 @@ namespace UndertaleModLib.Decompiler
                                 (!(cmp.Argument1 is ExpressionTempVar) || (cmp.Argument1 as ExpressionTempVar).Var.Var != switchTempVar))
                                 throw new Exception("Malformed switch statement: bad condition var (" + cmp.Argument1.ToString(context) + ")");
                             if (cmp.Opcode != UndertaleInstruction.ComparisonType.EQ)
-                                throw new Exception("Malformed switch statement: bad contition type (" + cmp.Opcode.ToString().ToUpper() + ")");
+                                throw new Exception("Malformed switch statement: bad contition type (" + cmp.Opcode.ToString().ToUpper(CultureInfo.InvariantCulture) + ")");
                             caseExpr = cmp.Argument2;
                         }
 
@@ -3907,11 +4081,17 @@ namespace UndertaleModLib.Decompiler
             return result;
         }
 
-        public static string Decompile(UndertaleCode code, GlobalDecompileContext globalContext)
+        public static string Decompile(UndertaleCode code, GlobalDecompileContext globalContext, Action<string> msgDelegate = null)
         {
             globalContext.DecompilerWarnings.Clear();
             DecompileContext context = new DecompileContext(globalContext, code);
+
+            if (msgDelegate is not null)
+                msgDelegate("Building the cache of all sub-functions...");
             BuildSubFunctionCache(globalContext.Data);
+            if (msgDelegate is not null)
+                msgDelegate("Decompiling, please wait... This can take a while on complex scripts.");
+
             try
             {
                 if (globalContext.Data != null && globalContext.Data.ToolInfo.ProfileMode)
@@ -3962,42 +4142,83 @@ namespace UndertaleModLib.Decompiler
         {
             // Find all functions defined in GlobalScripts
             // Use the cache so this only gets calculated once
-            if (data == null || !data.GMS2_3 || data.KnownSubFunctions != null)
+            if (data == null || !data.IsVersionAtLeast(2, 3) || data.KnownSubFunctions != null)
                 return;
-            foreach (var func in data.Functions)
+
+            // There's no "ConcurrentHashSet<>"; values aren't used.
+            ConcurrentDictionary<string, string> processingCodeList = new();
+            byte elapsedSec = 1;
+            Task mainTask = Task.Run(() =>
             {
-                if (data.Code.ByName(func.Name.Content) != null)
-                    func.Autogenerated = true;
-            }
-            data.KnownSubFunctions = new Dictionary<string, UndertaleFunction>();
-            GlobalDecompileContext globalDecompileContext = new GlobalDecompileContext(data, false);
-            Parallel.ForEach(data.GlobalInitScripts, globalScript =>
-            {
-                UndertaleCode scriptCode = globalScript.Code;
-                try
+                HashSet<string> codeNames = new(data.Code.Select(c => c.Name?.Content));
+                foreach (var func in data.Functions)
                 {
-                    DecompileContext childContext = new DecompileContext(globalDecompileContext, scriptCode, false);
-                    childContext.DisableAnonymousFunctionNameResolution = true;
-                    Dictionary<uint, Block> blocks2 = PrepareDecompileFlow(scriptCode, new List<uint>() { 0 });
-                    DecompileFromBlock(childContext, blocks2, blocks2[0]);
-                    List<Statement> statements = HLDecompile(childContext, blocks2, blocks2[0], blocks2[scriptCode.Length / 4]);
-                    foreach (Statement stmt2 in statements)
+                    if (codeNames.Contains(func.Name.Content))
+                        func.Autogenerated = true;
+                }
+                data.KnownSubFunctions = new Dictionary<string, UndertaleFunction>();
+                GlobalDecompileContext globalDecompileContext = new GlobalDecompileContext(data, false);
+
+                foreach (var func in data.Functions)
+                {
+                    if (func.Name.Content.StartsWith("gml_Script_"))
                     {
-                        if (stmt2 is AssignmentStatement assign &&
-                            assign.Value is FunctionDefinition funcDef)
+                        var funcName = func.Name.Content.Substring("gml_Script_".Length);
+                        data.KnownSubFunctions.TryAdd(funcName, func);
+                    }
+                }
+
+                Parallel.ForEach(data.GlobalInitScripts, globalScript =>
+                {
+                    UndertaleCode scriptCode = globalScript.Code;
+                    processingCodeList[scriptCode.Name.Content] = null;
+                    try
+                    {
+                        DecompileContext childContext = new DecompileContext(globalDecompileContext, scriptCode, false);
+                        childContext.DisableAnonymousFunctionNameResolution = true;
+                        Dictionary<uint, Block> blocks2 = PrepareDecompileFlow(scriptCode, new List<uint>() { 0 });
+                        DecompileFromBlock(childContext, blocks2, blocks2[0]);
+                        List<Statement> statements = HLDecompile(childContext, blocks2, blocks2[0], blocks2[scriptCode.Length / 4]);
+                        foreach (Statement stmt2 in statements)
                         {
-                            lock (data.KnownSubFunctions)
+                            if (stmt2 is AssignmentStatement assign &&
+                                assign.Value is FunctionDefinition funcDef)
                             {
-                                data.KnownSubFunctions.Add(assign.Destination.Var.Name.Content, funcDef.Function);
+                                lock (data.KnownSubFunctions)
+                                {
+                                    data.KnownSubFunctions.Add(assign.Destination.Var.Name.Content, funcDef.Function);
+                                }
                             }
                         }
                     }
-                }
-                catch (Exception e)
+                    catch (Exception e)
+                    {
+                        Debug.WriteLine(e.ToString());
+                    }
+                    processingCodeList.Remove(scriptCode.Name.Content, out _);
+                });
+                elapsedSec = 3 * 60;
+            });
+
+            Task timeoutTask = Task.Run(async () =>
+            {
+                while (true)
                 {
-                    Debug.WriteLine(e.ToString());
+                    await Task.Delay(1000);
+
+                    if (++elapsedSec > 3 * 60)
+                        return;
                 }
             });
+
+            // If the timeout task ended earlier than the main task
+            if (Task.WaitAny(mainTask, timeoutTask) == 1)
+            {
+                throw new TimeoutException("The building cache process hung.\n" +
+                                           "The function code entries that didn't manage to decompile:\n" +
+                                           String.Join('\n', processingCodeList.Keys) + "\n\n" +
+                                           "You should save the game data (if it's necessary) and re-open the app.\n");
+            }
         }
 
         private static void DoTypePropagation(DecompileContext context, Dictionary<uint, Block> blocks)
@@ -4018,8 +4239,13 @@ namespace UndertaleModLib.Decompiler
                 return meetPoint;
 
             Queue<Block> blocks = new Queue<Block>();
+            // Preventing the same block and its children from being queued repeatedly
+            // becomes increasingly important on large switches. The HashSet should give
+            // good performance while preventing this type of duplication.
+            HashSet<Block> usedBlocks = new HashSet<Block>(); 
 
             blocks.Enqueue(start);
+            usedBlocks.Add(start);
             while (blocks.Count > 0)
             {
                 Block test = blocks.Dequeue();
@@ -4028,10 +4254,16 @@ namespace UndertaleModLib.Decompiler
                     return end;
                 if (test == meetPoint)
                     return meetPoint;
-
-                blocks.Enqueue(test.nextBlockTrue);
-                if (test.nextBlockTrue != test.nextBlockFalse)
+                if (!usedBlocks.Contains(test.nextBlockTrue))
+                {
+                    blocks.Enqueue(test.nextBlockTrue);
+                    usedBlocks.Add(test.nextBlockTrue);
+                }
+                if (!usedBlocks.Contains(test.nextBlockFalse))
+                {
                     blocks.Enqueue(test.nextBlockFalse);
+                    usedBlocks.Add(test.nextBlockFalse);
+                }
 
             }
 
@@ -4048,7 +4280,7 @@ namespace UndertaleModLib.Decompiler
                 sb.Append("    block_" + pair.Key + " [label=\"");
                 sb.Append("[" + block.ToString() + ", Exit: " + block.conditionalExit + (block.nextBlockTrue != null ? ", T: " + block.nextBlockTrue.Address : "") + (block.nextBlockFalse != null ? ", F: " + block.nextBlockFalse.Address : "") + "]\n");
                 foreach (var instr in block.Instructions)
-                    sb.Append(instr.ToString().Replace("\"", "\\\"") + "\\n");
+                    sb.Append(instr.ToString().Replace("\"", "\\\"", StringComparison.InvariantCulture) + "\\n");
                 sb.Append('"');
                 sb.Append(pair.Key == 0 ? ", color=\"blue\"" : "");
                 sb.AppendLine(", shape=\"box\"];");
