@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -14,6 +15,7 @@ using System.Windows.Navigation;
 using System.Windows.Shapes;
 using UndertaleModLib;
 using UndertaleModLib.Models;
+using static UndertaleModLib.Models.UndertaleExtensionOption;
 
 namespace UndertaleModTool
 {
@@ -22,27 +24,147 @@ namespace UndertaleModTool
     /// </summary>
     public partial class UndertaleExtensionEditor : DataUserControl
     {
-        public int MyIndex { get => (Application.Current.MainWindow as MainWindow).Data.Extensions.IndexOf((UndertaleExtension)this.DataContext); }
+        private static readonly MainWindow mainWindow = Application.Current.MainWindow as MainWindow;
+
+        public int MyIndex
+        {
+            get
+            {
+                if (DataContext is not UndertaleExtension ext)
+                    return -1;
+                
+                return mainWindow.Data.Extensions.IndexOf(ext);
+            }
+        }
         // God this is so ugly, if there's a better way, please, put in a pull request
-        public byte[] ProductIdData { get => (((((Application.Current.MainWindow as MainWindow).Data?.GeneralInfo?.Major ?? 0) >= 2) || ((((Application.Current.MainWindow as MainWindow).Data?.GeneralInfo?.Major ?? 0) == 1) && ((((Application.Current.MainWindow as MainWindow).Data?.GeneralInfo?.Build ?? 0) >= 1773) || (((Application.Current.MainWindow as MainWindow).Data?.GeneralInfo?.Build ?? 0) == 1539)))) ? (Application.Current.MainWindow as MainWindow).Data.FORM.EXTN.productIdData[MyIndex] : null); set => (Application.Current.MainWindow as MainWindow).Data.FORM.EXTN.productIdData[MyIndex] = value; }
+        public byte[] ProductIdData { get => ((((mainWindow.Data?.GeneralInfo?.Major ?? 0) >= 2) || (((mainWindow.Data?.GeneralInfo?.Major ?? 0) == 1) && (((mainWindow.Data?.GeneralInfo?.Build ?? 0) >= 1773) || ((mainWindow.Data?.GeneralInfo?.Build ?? 0) == 1539)))) ? mainWindow.Data.FORM.EXTN.productIdData[MyIndex] : null); set => mainWindow.Data.FORM.EXTN.productIdData[MyIndex] = value; }
 
         public UndertaleExtensionEditor()
         {
             InitializeComponent();
         }
 
-        private void Button_Click(object sender, RoutedEventArgs e)
+        private void NewFileButton_Click(object sender, RoutedEventArgs e)
         {
-            int lastItem = ((Application.Current.MainWindow as MainWindow).Data.Extensions[MyIndex]).Files.Count;
+            if (DataContext is not UndertaleExtension extension)
+                return;
+            
+            int lastItem = extension.Files.Count;
 
-            UndertaleExtensionFile obj = new UndertaleExtensionFile()
+            UndertaleExtensionFile obj = new()
             {
                 Kind = UndertaleExtensionKind.Dll,
-                Filename = (Application.Current.MainWindow as MainWindow).Data.Strings.MakeString($"NewExtensionFile{lastItem}.dll"),
+                Filename = mainWindow.Data.Strings.MakeString($"NewExtensionFile{lastItem}.dll"),
                 Functions = new UndertalePointerList<UndertaleExtensionFunction>()
             };
+            extension.Files.Add(obj);
+        }
+        private void NewOptionButton_Click(object sender, RoutedEventArgs e)
+        {
+            if (DataContext is not UndertaleExtension extension)
+                return;
+            
+            int lastItem = extension.Options.Count;
 
-            ((Application.Current.MainWindow as MainWindow).Data.Extensions[MyIndex]).Files.Add(obj);
+            UndertaleExtensionOption obj = new()
+            {
+                Name = mainWindow.Data.Strings.MakeString($"extensionOption{lastItem}"),
+                Value = mainWindow.Data.Strings.MakeString("", true)
+            };
+            extension.Options.Add(obj);
+        }
+
+        private void KindComboBox_SelectionChanged(object sender, SelectionChangedEventArgs e)
+        {
+            var comboBox = sender as ComboBox;
+            var option = comboBox?.DataContext as UndertaleExtensionOption;
+            if (option?.Value is null)
+                return;
+            
+            switch (comboBox.SelectedItem)
+            {
+                case OptionKind.String:
+                case null:
+                    break;
+                
+                case OptionKind.Boolean:
+                    if (option.Value.Content.ToLowerInvariant() == "true")
+                        option.Value.Content = "True";
+                    else
+                        option.Value.Content = "False";
+                    break;
+                
+                case OptionKind.Number:
+                    if (!Double.TryParse(option.Value.Content, NumberStyles.Any, CultureInfo.InvariantCulture, out double _))
+                        option.Value.Content = "0";
+                    break;
+            };
+        }
+    }
+
+    public class OptionValueTemplateSelector : DataTemplateSelector
+    {
+        public DataTemplate StringTemplate { get; set; }
+        public DataTemplate BooleanTemplate { get; set; }
+        public DataTemplate NumberTemplate { get; set; }
+
+        public override DataTemplate SelectTemplate(object item, DependencyObject container)
+        {
+            if (item is not OptionKind kind)
+                return null;
+
+            return kind switch
+            {
+                OptionKind.String => StringTemplate,
+                OptionKind.Boolean => BooleanTemplate,
+                OptionKind.Number => NumberTemplate,
+                _ => null
+            };
+        }
+    }
+
+    public class OptionValueConverter : IValueConverter
+    {
+        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (value is not string str)
+                return null;
+            
+            switch (parameter)
+            {
+                case "boolean":
+                    return str.ToLowerInvariant() == "true";
+                
+                case "number":
+                    if (Double.TryParse(str, NumberStyles.Any, CultureInfo.InvariantCulture, out double _))
+                        return str;
+                    return "0";
+                
+                default:
+                    return str;
+            }
+        }
+
+        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
+        {
+            if (parameter is not string par)
+                return null;
+            
+            switch (par)
+            {
+                case "boolean":
+                    if (value is not bool b)
+                        return new ValidationResult(false, "Invalid boolean value");
+                    return (b ? "True" : "False");
+                
+                case "number":
+                    if (value is string s && Double.TryParse(s, NumberStyles.Any, CultureInfo.InvariantCulture, out double _))
+                        return s;
+                    return new ValidationResult(false, "Invalid number string");
+                
+                default:
+                    return null;
+            }
         }
     }
 }
