@@ -35,6 +35,8 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
     public UndertaleResourceById<UndertaleGameObject, UndertaleChunkOBJT> _parentId = new();
     public UndertaleResourceById<UndertaleSprite, UndertaleChunkSPRT> _textureMaskId = new();
 
+    public static readonly int EventTypeCount = Enum.GetValues(typeof(EventType)).Length;
+
     /// <summary>
     /// The name of the game object.
     /// </summary>
@@ -148,6 +150,10 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
 
     /// <inheritdoc />
     public event PropertyChangedEventHandler PropertyChanged;
+    
+    /// <summary>
+    /// Invoked whenever the effective value of any dependency property has been updated.
+    /// </summary>
     protected void OnPropertyChanged([CallerMemberName] string name = null)
     {
         PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(name));
@@ -158,8 +164,9 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
     /// </summary>
     public UndertaleGameObject()
     {
-        for (int i = 0; i < Enum.GetValues(typeof(EventType)).Length; i++)
-            Events.Add(new UndertalePointerList<Event>());
+        Events.SetCapacity(EventTypeCount);
+        for (int i = 0; i < EventTypeCount; i++)
+            Events.InternalAdd(new UndertalePointerList<Event>());
     }
 
     /// <inheritdoc />
@@ -168,7 +175,7 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
         writer.WriteUndertaleString(Name);
         writer.WriteUndertaleObject(_sprite);
         writer.Write(Visible);
-        if (writer.undertaleData.GM2022_5)
+        if (writer.undertaleData.IsVersionAtLeast(2022, 5))
             writer.Write(Managed);
         writer.Write(Solid);
         writer.Write(Depth);
@@ -209,7 +216,7 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
         Name = reader.ReadUndertaleString();
         _sprite = reader.ReadUndertaleObject<UndertaleResourceById<UndertaleSprite, UndertaleChunkSPRT>>();
         Visible = reader.ReadBoolean();
-        if (reader.undertaleData.GM2022_5)
+        if (reader.undertaleData.IsVersionAtLeast(2022, 5))
             Managed = reader.ReadBoolean();
         Solid = reader.ReadBoolean();
         Depth = reader.ReadInt32();
@@ -240,6 +247,7 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
         Awake = reader.ReadBoolean();
         Kinematic = reader.ReadBoolean();
         // Needs to be done manually because count is separated
+        PhysicsVertices.Capacity = physicsShapeVertexCount;
         for (int i = 0; i < physicsShapeVertexCount; i++)
         {
             UndertalePhysicsVertex v = new UndertalePhysicsVertex();
@@ -249,12 +257,31 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
         Events = reader.ReadUndertaleObject<UndertalePointerList<UndertalePointerList<Event>>>();
     }
 
+    /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+    public static uint UnserializeChildObjectCount(UndertaleReader reader)
+    {
+        uint count = 0;
+
+        if (reader.undertaleData.IsVersionAtLeast(2022, 5))
+            reader.Position += 64 + 4; // + "Managed"
+        else
+            reader.Position += 64;
+
+        int physicsShapeVertexCount = reader.ReadInt32();
+        reader.Position += 12 + (uint)physicsShapeVertexCount * UndertalePhysicsVertex.ChildObjectsSize;
+
+        count += 2 + 1 + UndertalePointerList<UndertalePointerList<Event>>.UnserializeChildObjectCount(reader);
+
+        return count;
+    }
+
     #region EventHandlerFor() overloads
-    //TODO: what do all these eventhandlers do? can't find any references right now.
+    // TODO: Add documentation for these methods.
+    // These methods are used by scripts for getting a code entry for a certain event of the game object.
 
     public UndertaleCode EventHandlerFor(EventType type, uint subtype, IList<UndertaleString> strg, IList<UndertaleCode> codelist, IList<UndertaleCodeLocals> localslist)
     {
-        Event subtypeObj = Events[(int)type].Where((x) => x.EventSubtype == subtype).FirstOrDefault();
+        Event subtypeObj = Events[(int)type].FirstOrDefault(x => x.EventSubtype == subtype);
         if (subtypeObj == null)
             Events[(int)type].Add(subtypeObj = new Event() { EventSubtype = subtype });
         EventAction action = subtypeObj.Actions.FirstOrDefault();
@@ -266,7 +293,7 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
         UndertaleCode code = action.CodeId;
         if (code == null)
         {
-            var name = strg.MakeString("gml_Object_" + Name.Content + "_" + type.ToString() + "_" + subtype);
+            var name = strg.MakeString("gml_Object_" + Name.Content + "_" + type + "_" + subtype);
             code = new UndertaleCode()
             {
                 Name = name,
@@ -384,7 +411,7 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
         {
             foreach (var subEv in ev)
                 subEv?.Dispose();
-         }
+        }
         Name = null;
         Events = new();
     }
@@ -458,6 +485,14 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
             Actions = reader.ReadUndertaleObject<UndertalePointerList<EventAction>>();
         }
 
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            reader.Position += 4; // "EventSubtype"
+
+            return 1 + UndertalePointerList<EventAction>.UnserializeChildObjectCount(reader);
+        }
+
         /// <inheritdoc/>
         public void Dispose()
         {
@@ -472,8 +507,14 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
     /// <summary>
     /// An action in an event.
     /// </summary>
-    public class EventAction : UndertaleObject, INotifyPropertyChanged, IDisposable
+    public class EventAction : UndertaleObject, INotifyPropertyChanged, IDisposable,
+                               IStaticChildObjectsSize, IStaticChildObjCount
     {
+        /// <inheritdoc cref="IStaticChildObjCount.ChildObjectCount" />
+        public static readonly uint ChildObjectCount = 1;
+        /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
+        public static readonly uint ChildObjectsSize = 56;
+
         // All the unknown values seem to be provided for compatibility only - in older versions of GM:S they stored the drag and drop blocks,
         // but newer versions compile them down to GML bytecode anyway
         // Possible meaning of values: https://github.com/WarlockD/GMdsam/blob/26aefe3e90a7a7a1891cb83f468079546f32b4b7/GMdsam/GameMaker/ChunkTypes.cs#L466
@@ -500,6 +541,7 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
         public bool IsNot { get; set; } // always 0
         public uint UnknownAlwaysZero { get; set; } // always 0
 
+        /// <inheritdoc />
         public event PropertyChangedEventHandler PropertyChanged;
 
         /// <inheritdoc />
@@ -554,8 +596,11 @@ public class UndertaleGameObject : UndertaleNamedResource, INotifyPropertyChange
     /// Class representing a physics vertex used for a <see cref="CollisionShape"/> of type <see cref="CollisionShapeFlags.Custom"/>.
     /// </summary>
     [PropertyChanged.AddINotifyPropertyChangedInterface]
-    public class UndertalePhysicsVertex : UndertaleObject
+    public class UndertalePhysicsVertex : UndertaleObject, IStaticChildObjectsSize
     {
+        /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
+        public static readonly uint ChildObjectsSize = 8;
+
         /// <summary>
         /// The x position of the vertex.
         /// </summary>
