@@ -80,6 +80,31 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
         Moments = reader.ReadUndertaleObject<UndertaleSimpleList<Keyframe<Moment>>>();
     }
 
+    /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+    public static uint UnserializeChildObjectCount(UndertaleReader reader)
+    {
+        uint count = 0;
+
+        reader.Position += 32;
+
+        count += 1 + UndertaleSimpleList<Keyframe<BroadcastMessage>>.UnserializeChildObjectCount(reader);
+
+        count += 1 + UndertaleSimpleList<Track>.UnserializeChildObjectCount(reader);
+
+        int funcIDsCount = reader.ReadInt32();
+        reader.Position += (uint)funcIDsCount * 8;
+
+        count += 1 + UndertaleSimpleList<Keyframe<Moment>>.UnserializeChildObjectCount(reader);
+
+        return count;
+    }
+
+    /// <inheritdoc/>
+    public override string ToString()
+    {
+        return $"Sequence \"{Name.Content}\"";
+    }
+
     /// <inheritdoc/>
     public void Dispose()
     {
@@ -132,6 +157,38 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
                 Channels[channel] = data;
             }
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            uint count = 0;
+
+            reader.Position += 16;
+            int chCount = reader.ReadInt32();
+
+            Type t = typeof(T);
+            if (t.IsAssignableTo(typeof(IStaticChildObjectsSize)))
+            {
+                uint subSize = reader.GetStaticChildObjectsSize(t);
+                uint subCount = 0;
+
+                if (t.IsAssignableTo(typeof(IStaticChildObjCount)))
+                    subCount = reader.GetStaticChildCount(t);
+
+                reader.Position += (uint)(chCount * 4 + chCount * subSize);
+
+                return (uint)chCount * subCount;
+            }
+
+            var unserializeFunc = reader.GetUnserializeCountFunc(t);
+            for (int i = 0; i < chCount; i++)
+            {
+                reader.Position += 4; // channel ID
+                count += unserializeFunc(reader);
+            }
+
+            return count;
+        }
     }
 
     public class BroadcastMessage : UndertaleObject
@@ -149,6 +206,12 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
         {
             Messages = new UndertaleSimpleListString();
             Messages.Unserialize(reader);
+        }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            return UndertaleSimpleListString.UnserializeChildObjectCount(reader);
         }
     }
 
@@ -172,18 +235,58 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
             if (InternalCount > 0)
                 Event = reader.ReadUndertaleString();
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            int internalCount = reader.ReadInt32();
+            if (internalCount > 0)
+                reader.Position += 4; // "Event"
+
+            return 0;
+        }
     }
 
     public class Track : UndertaleObject
     {
+        public enum TrackBuiltinName
+        {
+            Gain = 5,
+            Pitch = 6,
+            Falloff = 7,
+            Rotation = 8,
+            BlendAdd = 9,
+            BlendMultiply = 10,
+            Mask = 12,
+            Subject = 13,
+            Position = 14,
+            Scale = 15,
+            Origin = 16,
+            ImageSpeed = 17,
+            ImageIndex = 18,
+            ImageAngle = Rotation,
+            ImageBlend = BlendMultiply,
+            FrameSize = 20,
+            CharacterSpacing = 21,
+            LineSpacing = 22,
+            ParagraphSpacing = 23
+        }
+
+        [Flags]
+        public enum TrackTraits
+        {
+            None,
+            ChildrenIgnoreOrigin
+        }
+
         public UndertaleString ModelName { get; set; } // Such as GMInstanceTrack, GMRealTrack, etc.
         public UndertaleString Name { get; set; } // An asset or property name
-        public int BuiltinName { get; set; }
-        public int Traits { get; set; }
+        public TrackBuiltinName BuiltinName { get; set; }
+        public TrackTraits Traits { get; set; }
         public bool IsCreationTrack { get; set; }
         public List<int> Tags { get; set; }
         public List<Track> Tracks { get; set; } // Sub-tracks
-        public TrackKeyframes Keyframes { get; set; }
+        public ITrackKeyframes Keyframes { get; set; }
         public List<UndertaleResource> OwnedResources { get; set; }
 
         public UndertaleString GMAnimCurveString;
@@ -193,8 +296,8 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
         {
             writer.WriteUndertaleString(ModelName);
             writer.WriteUndertaleString(Name);
-            writer.Write(BuiltinName);
-            writer.Write(Traits);
+            writer.Write((int)BuiltinName);
+            writer.Write((int)Traits);
             writer.Write(IsCreationTrack);
 
             writer.Write(Tags.Count);
@@ -252,6 +355,12 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
                 case "GMColourTrack":
                     writer.WriteUndertaleObject(Keyframes as RealKeyframes);
                     break;
+                case "GMTextTrack":     // Introduced in GM 2022.2
+                    writer.WriteUndertaleObject(Keyframes as TextKeyframes);
+                    break;
+                case "GMParticleTrack": // Introduced in GM 2023.2
+                    writer.WriteUndertaleObject(Keyframes as ParticleKeyframes);
+                    break;
             }
         }
 
@@ -262,28 +371,34 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
             UndertaleString ForceReadString()
             {
                 UndertaleString res = reader.ReadUndertaleString();
-                uint returnTo = reader.Position;
+                if (res.Content is not null)
+                    return res;
+
+                reader.SwitchReaderType(false);
+                long returnTo = reader.Position;
                 reader.Position = reader.GetOffsetMapRev()[res];
                 reader.ReadUndertaleObject<UndertaleString>();
                 reader.Position = returnTo;
+                reader.SwitchReaderType(true);
+
                 return res;
             }
 
             ModelName = ForceReadString();
             Name = reader.ReadUndertaleString();
-            BuiltinName = reader.ReadInt32();
-            Traits = reader.ReadInt32();
+            BuiltinName = (TrackBuiltinName)reader.ReadInt32();
+            Traits = (TrackTraits)reader.ReadInt32();
             IsCreationTrack = reader.ReadBoolean();
 
             int tagCount = reader.ReadInt32();
             int ownedResCount = reader.ReadInt32();
             int trackCount = reader.ReadInt32();
 
-            Tags = new List<int>();
+            Tags = new List<int>(tagCount);
             for (int i = 0; i < tagCount; i++)
                 Tags.Add(reader.ReadInt32());
 
-            OwnedResources = new List<UndertaleResource>();
+            OwnedResources = new List<UndertaleResource>(ownedResCount);
             for (int i = 0; i < ownedResCount; i++)
             {
                 GMAnimCurveString = ForceReadString();
@@ -294,7 +409,7 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
                 OwnedResources.Add(res);
             }
 
-            Tracks = new List<Track>();
+            Tracks = new List<Track>(trackCount);
             for (int i = 0; i < trackCount; i++)
                 Tracks.Add(reader.ReadUndertaleObject<Track>());
 
@@ -332,27 +447,154 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
                 case "GMColourTrack":
                     Keyframes = reader.ReadUndertaleObject<RealKeyframes>();
                     break;
+                case "GMTextTrack":     // Introduced in GM 2022.2
+                    if (!reader.undertaleData.IsVersionAtLeast(2022, 2))
+                        reader.undertaleData.SetGMS2Version(2022, 2);
+                    Keyframes = reader.ReadUndertaleObject<TextKeyframes>();
+                    break;
+                case "GMParticleTrack": // Introduced in GM 2023.2
+                    Keyframes = reader.ReadUndertaleObject<ParticleKeyframes>();
+                    break;
+                
+                // "GMGroupTrack" and "GMClipMaskTrack" have null keyframes
             }
+        }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            uint count = 0;
+
+            string ForceReadString()
+            {
+                uint strPtr = reader.ReadUInt32();
+
+                reader.SwitchReaderType(false);
+                long returnTo = reader.Position;
+                reader.Position = strPtr - 4;
+                string res = reader.ReadGMString();
+                reader.Position = returnTo;
+                reader.SwitchReaderType(true);
+
+                return res;
+            }
+
+            string modelName = ForceReadString();
+            reader.Position += 16;
+
+            int tagCount = reader.ReadInt32();
+            int ownedResCount = reader.ReadInt32();
+            int trackCount = reader.ReadInt32();
+
+            reader.Position += (uint)tagCount * 4; // "Tags"
+
+            for (int i = 0; i < ownedResCount; i++)
+            {
+                reader.Position += 4; // "GMAnimCurveString"
+                count += UndertaleAnimationCurve.UnserializeChildObjectCount(reader);
+            }
+
+            // "Tracks"
+            for (int i = 0; i < trackCount; i++)
+                count += 1 + UnserializeChildObjectCount(reader);
+
+            switch (modelName)
+            {
+                case "GMAudioTrack":
+                    count += 1 + AudioKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                case "GMInstanceTrack":
+                    count += 1 + InstanceKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                case "GMGraphicTrack":
+                    count += 1 + GraphicKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                case "GMSequenceTrack":
+                    count += 1 + SequenceKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                case "GMSpriteFramesTrack":
+                    count += 1 + SpriteFramesKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                case "GMAssetTrack": // TODO?
+                    throw new NotImplementedException("GMAssetTrack not implemented, report this");
+                case "GMBoolTrack":
+                    count += 1 + BoolKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                case "GMStringTrack":
+                    count += 1 + StringKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                // TODO?
+                //case "GMIntTrack":
+                //  count += 1 + IntKeyframes.UnserializeChildObjectCount(reader);
+                //  break;
+                case "GMRealTrack":
+                case "GMColourTrack":
+                    count += 1 + RealKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                case "GMTextTrack":     // Introduced in GM 2022.2
+                    count += 1 + TextKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+                case "GMParticleTrack": // Introduced in GM 2023.2
+                    count += 1 + ParticleKeyframes.UnserializeChildObjectCount(reader);
+                    break;
+
+                // "GMGroupTrack" and "GMClipMaskTrack" have null keyframes
+            }
+
+            return count;
+        }
+
+        /// <inheritdoc/>
+        public override string ToString()
+        {
+            if (BuiltinName != 0)
+            {
+                if (ModelName.Content == "GMColourTrack")
+                    return $"Sequence sub-track - \"Color\" (mode \"{BuiltinName}\")";
+                else
+                    return $"Sequence sub-track - \"{BuiltinName}\"";
+            }
+            else
+                return $"Sequence track \"{Name.Content}\"";
         }
     }
 
     /// Here begins all of the keyframe data classes. Some generics used to shorten sections, but some verbosity maintained
 
-    public class TrackKeyframes : UndertaleObject
+    public interface ITrackKeyframes : UndertaleObject
     {
+    }
+    public class TrackKeyframes<T> : ITrackKeyframes where T : UndertaleObject, new()
+    {
+        public UndertaleSimpleList<Keyframe<T>> List;
+
         /// <inheritdoc />
         public virtual void Serialize(UndertaleWriter writer)
         {
             while (writer.Position % 4 != 0)
                 writer.Write((byte)0);
+
+            List.Serialize(writer);
         }
 
         /// <inheritdoc />
         public virtual void Unserialize(UndertaleReader reader)
         {
-            while (reader.Position % 4 != 0)
+            while (reader.AbsPosition % 4 != 0)
                 if (reader.ReadByte() != 0)
                     throw new IOException("Padding error!");
+
+            List = new UndertaleSimpleList<Keyframe<T>>();
+            List.Unserialize(reader);
+        }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            while (reader.AbsPosition % 4 != 0)
+                reader.Position++;
+
+            return UndertaleSimpleList<Keyframe<T>>.UnserializeChildObjectCount(reader);
         }
     }
 
@@ -372,9 +614,40 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
             Resource = new T();
             Resource.Unserialize(reader);
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            // At this moment, T could be only "UndertaleResourceById<>".
+            // If that changes, you should replace the contents with the following:
+            // return reader.GetChildObjectCount<T>();
+
+            reader.Position += 4;
+            return 0;
+        }
     }
 
-    public class AudioKeyframes : TrackKeyframes
+    public class SimpleIntData : IStaticChildObjectsSize, UndertaleObject
+    {
+        /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
+        public static readonly uint ChildObjectsSize = 4;
+
+        public int Value { get; set; }
+
+        /// <inheritdoc />
+        public void Serialize(UndertaleWriter writer)
+        {
+            writer.Write(Value);
+        }
+
+        /// <inheritdoc />
+        public void Unserialize(UndertaleReader reader)
+        {
+            Value = reader.ReadInt32();
+        }
+    }
+
+    public class AudioKeyframes : TrackKeyframes<AudioKeyframes.Data>
     {
         public class Data : ResourceData<UndertaleResourceById<UndertaleSound, UndertaleChunkSOND>>
         {
@@ -384,7 +657,7 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
             public override void Serialize(UndertaleWriter writer)
             {
                 base.Serialize(writer);
-                writer.Write((int)0);
+                writer.Write(0);
                 writer.Write(Mode);
             }
 
@@ -396,197 +669,64 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
                     throw new IOException("Expected 0 in Audio keyframe");
                 Mode = reader.ReadInt32();
             }
-        }
 
-        public UndertaleSimpleList<Keyframe<Data>> List;
+            /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+            public new static uint UnserializeChildObjectCount(UndertaleReader reader)
+            {
+                uint count = ResourceData<UndertaleResourceById<UndertaleSound, UndertaleChunkSOND>>.UnserializeChildObjectCount(reader);
 
-        /// <inheritdoc />
-        public override void Serialize(UndertaleWriter writer)
-        {
-            base.Serialize(writer);
-            List.Serialize(writer);
-        }
+                reader.Position += 8;
 
-        /// <inheritdoc />
-        public override void Unserialize(UndertaleReader reader)
-        {
-            base.Unserialize(reader);
-            List = new UndertaleSimpleList<Keyframe<Data>>();
-            List.Unserialize(reader);
+                return count;
+            }
         }
     }
 
-    public class InstanceKeyframes : TrackKeyframes
+    public class InstanceKeyframes : TrackKeyframes<InstanceKeyframes.Data>
     {
         public class Data : ResourceData<UndertaleResourceById<UndertaleGameObject, UndertaleChunkOBJT>> { }
-        public UndertaleSimpleList<Keyframe<Data>> List;
-
-        /// <inheritdoc />
-        public override void Serialize(UndertaleWriter writer)
-        {
-            base.Serialize(writer);
-            List.Serialize(writer);
-        }
-
-        /// <inheritdoc />
-        public override void Unserialize(UndertaleReader reader)
-        {
-            base.Unserialize(reader);
-            List = new UndertaleSimpleList<Keyframe<Data>>();
-            List.Unserialize(reader);
-        }
     }
 
-    public class GraphicKeyframes : TrackKeyframes
+    public class GraphicKeyframes : TrackKeyframes<GraphicKeyframes.Data>
     {
         public class Data : ResourceData<UndertaleResourceById<UndertaleSprite, UndertaleChunkSPRT>> { }
-        public UndertaleSimpleList<Keyframe<Data>> List;
-
-        /// <inheritdoc />
-        public override void Serialize(UndertaleWriter writer)
-        {
-            base.Serialize(writer);
-            List.Serialize(writer);
-        }
-
-        /// <inheritdoc />
-        public override void Unserialize(UndertaleReader reader)
-        {
-            base.Unserialize(reader);
-            List = new UndertaleSimpleList<Keyframe<Data>>();
-            List.Unserialize(reader);
-        }
     }
 
-    public class SequenceKeyframes : TrackKeyframes
+    public class SequenceKeyframes : TrackKeyframes<SequenceKeyframes.Data>
     {
         public class Data : ResourceData<UndertaleResourceById<UndertaleSequence, UndertaleChunkSEQN>> { }
-        public UndertaleSimpleList<Keyframe<Data>> List;
-
-        /// <inheritdoc />
-        public override void Serialize(UndertaleWriter writer)
-        {
-            base.Serialize(writer);
-            List.Serialize(writer);
-        }
-
-        /// <inheritdoc />
-        public override void Unserialize(UndertaleReader reader)
-        {
-            base.Unserialize(reader);
-            List = new UndertaleSimpleList<Keyframe<Data>>();
-            List.Unserialize(reader);
-        }
     }
 
-    public class SpriteFramesData : UndertaleObject
+    public class SpriteFramesKeyframes : TrackKeyframes<SpriteFramesKeyframes.Data>
     {
-        public int Value { get; set; }
-
-        /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
-        {
-            writer.Write(Value);
-        }
-
-        /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
-        {
-            Value = reader.ReadInt32();
-        }
+        public class Data : SimpleIntData { }
     }
 
-    public class SpriteFramesKeyframes : TrackKeyframes
+    public class BoolKeyframes : TrackKeyframes<BoolKeyframes.Data>
     {
-        public UndertaleSimpleList<Keyframe<SpriteFramesData>> List;
-
-        /// <inheritdoc />
-        public override void Serialize(UndertaleWriter writer)
-        {
-            base.Serialize(writer);
-            List.Serialize(writer);
-        }
-
-        /// <inheritdoc />
-        public override void Unserialize(UndertaleReader reader)
-        {
-            base.Unserialize(reader);
-            List = new UndertaleSimpleList<Keyframe<SpriteFramesData>>();
-            List.Unserialize(reader);
-        }
+        public class Data : SimpleIntData { }
     }
 
-    public class BoolData : UndertaleObject
+    public class StringKeyframes : TrackKeyframes<StringKeyframes.Data>
     {
-        public int Value { get; set; }
-
-        /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public class Data : UndertaleObject, IStaticChildObjectsSize
         {
-            writer.Write(Value);
-        }
+            /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
+            public static readonly uint ChildObjectsSize = 4;
 
-        /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
-        {
-            Value = reader.ReadInt32();
-        }
-    }
+            public UndertaleString Value { get; set; }
 
-    public class BoolKeyframes : TrackKeyframes
-    {
-        public UndertaleSimpleList<Keyframe<BoolData>> List;
+            /// <inheritdoc />
+            public void Serialize(UndertaleWriter writer)
+            {
+                writer.WriteUndertaleString(Value);
+            }
 
-        /// <inheritdoc />
-        public override void Serialize(UndertaleWriter writer)
-        {
-            base.Serialize(writer);
-            List.Serialize(writer);
-        }
-
-        /// <inheritdoc />
-        public override void Unserialize(UndertaleReader reader)
-        {
-            base.Unserialize(reader);
-            List = new UndertaleSimpleList<Keyframe<BoolData>>();
-            List.Unserialize(reader);
-        }
-    }
-
-    public class StringData : UndertaleObject
-    {
-        public UndertaleString Value { get; set; }
-
-        /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
-        {
-            writer.WriteUndertaleString(Value);
-        }
-
-        /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
-        {
-            Value = reader.ReadUndertaleString();
-        }
-    }
-
-    public class StringKeyframes : TrackKeyframes
-    {
-        public UndertaleSimpleList<Keyframe<StringData>> List;
-
-        /// <inheritdoc />
-        public override void Serialize(UndertaleWriter writer)
-        {
-            base.Serialize(writer);
-            List.Serialize(writer);
-        }
-
-        /// <inheritdoc />
-        public override void Unserialize(UndertaleReader reader)
-        {
-            base.Unserialize(reader);
-            List = new UndertaleSimpleList<Keyframe<StringData>>();
-            List.Unserialize(reader);
+            /// <inheritdoc />
+            public void Unserialize(UndertaleReader reader)
+            {
+                Value = reader.ReadUndertaleString();
+            }
         }
     }
 
@@ -631,6 +771,24 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
                 AssetAnimCurve.Unserialize(reader);
             }
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            uint count = 0;
+            
+            // "IsCurveEmbedded"
+            if (reader.ReadBoolean())
+            {
+                reader.Position += 4;
+
+                count += UndertaleAnimationCurve.UnserializeChildObjectCount(reader, false);
+            }
+            else
+                reader.Position += 4;
+
+            return count;
+        }
     }
 
     public class IntData : CurveData
@@ -650,17 +808,25 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
             Value = reader.ReadInt32();
             base.Unserialize(reader);
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public new static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            reader.Position += 4; // "Value"
+
+            return CurveData.UnserializeChildObjectCount(reader);
+        }
     }
 
-    public class IntKeyframes : TrackKeyframes
+    public class IntKeyframes : TrackKeyframes<IntData>
     {
-        public UndertaleSimpleList<Keyframe<IntData>> List;
         public int Interpolation;
 
         /// <inheritdoc />
         public override void Serialize(UndertaleWriter writer)
         {
-            base.Serialize(writer);
+            while (writer.Position % 4 != 0)
+                writer.Write((byte)0);
 
             writer.Write(Interpolation);
 
@@ -670,12 +836,25 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
         /// <inheritdoc />
         public override void Unserialize(UndertaleReader reader)
         {
-            base.Unserialize(reader);
+            while (reader.AbsPosition % 4 != 0)
+                if (reader.ReadByte() != 0)
+                    throw new IOException("Padding error!");
 
             Interpolation = reader.ReadInt32();
 
             List = new UndertaleSimpleList<Keyframe<IntData>>();
             List.Unserialize(reader);
+        }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public new static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            while (reader.AbsPosition % 4 != 0)
+                reader.Position++;
+
+            reader.Position += 4; // "Interpolation"
+
+            return UndertaleSimpleList<Keyframe<IntData>>.UnserializeChildObjectCount(reader);
         }
     }
 
@@ -696,17 +875,25 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
             Value = reader.ReadSingle();
             base.Unserialize(reader);
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public new static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            reader.Position += 4; // "Value"
+
+            return CurveData.UnserializeChildObjectCount(reader);
+        }
     }
 
-    public class RealKeyframes : TrackKeyframes
+    public class RealKeyframes : TrackKeyframes<RealData>
     {
-        public UndertaleSimpleList<Keyframe<RealData>> List;
         public int Interpolation;
 
         /// <inheritdoc />
         public override void Serialize(UndertaleWriter writer)
         {
-            base.Serialize(writer);
+            while (writer.Position % 4 != 0)
+                writer.Write((byte)0);
 
             writer.Write(Interpolation);
 
@@ -716,13 +903,75 @@ public class UndertaleSequence : UndertaleNamedResource, IDisposable
         /// <inheritdoc />
         public override void Unserialize(UndertaleReader reader)
         {
-            base.Unserialize(reader);
+            while (reader.AbsPosition % 4 != 0)
+                if (reader.ReadByte() != 0)
+                    throw new IOException("Padding error!");
 
             Interpolation = reader.ReadInt32();
 
             List = new UndertaleSimpleList<Keyframe<RealData>>();
             List.Unserialize(reader);
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public new static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            while (reader.AbsPosition % 4 != 0)
+                reader.Position++;
+
+            reader.Position += 4; // "Interpolation"
+
+            return UndertaleSimpleList<Keyframe<RealData>>.UnserializeChildObjectCount(reader);
+        }
+    }
+    
+    public class TextKeyframes : TrackKeyframes<TextKeyframes.Data>
+    {
+        // Source - https://github.com/YoYoGames/GameMaker-HTML5/blob/develop/scripts/yySequence.js#L2227
+        // ("yyTextTrackKey")
+        public class Data : UndertaleObject, IStaticChildObjectsSize
+        {
+            /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
+            public static readonly uint ChildObjectsSize = 16;
+
+            private int _alignment;
+
+            public UndertaleString Text { get; set; }
+            public bool Wrap { get; set; }
+            public int AlignmentV
+            {
+                get => (_alignment >> 8) & 0xff;
+                set => _alignment = (_alignment & 0xff) | (value & 0xff) << 8;
+            }
+            public int AlignmentH
+            {
+                get => _alignment & 0xff;
+                set => _alignment = (_alignment & ~0xff) | (value & 0xff);
+            }
+            public int FontIndex { get; set; }
+
+            /// <inheritdoc />
+            public void Serialize(UndertaleWriter writer)
+            {
+                writer.WriteUndertaleString(Text);
+                writer.Write(Wrap);
+                writer.Write(_alignment);
+                writer.Write(FontIndex);
+            }
+
+            /// <inheritdoc />
+            public void Unserialize(UndertaleReader reader)
+            {
+                Text = reader.ReadUndertaleString();
+                Wrap = reader.ReadBoolean();
+                _alignment = reader.ReadInt32();
+                FontIndex = reader.ReadInt32();
+            }
+        }
     }
 
+    public class ParticleKeyframes : TrackKeyframes<ParticleKeyframes.Data>
+    {
+        public class Data : ResourceData<UndertaleResourceById<UndertaleParticleSystem, UndertaleChunkPSYS>> { }
+    }
 }
