@@ -29,9 +29,10 @@ public class UndertaleTags : UndertaleObject, IDisposable
             UndertaleShader => ResourceType.Shader,
             UndertaleSequence => ResourceType.Sequence,
             UndertaleAnimationCurve => ResourceType.AnimCurve,
+            UndertaleParticleSystem => ResourceType.ParticleSystem,
             _ => throw new ArgumentException("Invalid resource type!")
         };
-        IList list = data[resource.GetType()] as IList;
+        IList list = data[resource.GetType()];
 
         int offset = resource is UndertaleScript ? 100000 : 0;
 
@@ -60,6 +61,17 @@ public class UndertaleTags : UndertaleObject, IDisposable
             AssetTags[t.ID] = t.Tags;
     }
 
+    /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+    public static uint UnserializeChildObjectCount(UndertaleReader reader)
+    {
+        uint count = 0;
+
+        count += UndertaleSimpleListString.UnserializeChildObjectCount(reader);
+        count += 1 + UndertalePointerList<TempAssetTags>.UnserializeChildObjectCount(reader);
+
+        return count;
+    }
+
     /// <inheritdoc/>
     public void Dispose()
     {
@@ -86,6 +98,13 @@ public class UndertaleTags : UndertaleObject, IDisposable
         {
             ID = reader.ReadInt32();
             Tags = reader.ReadUndertaleObject<UndertaleSimpleListString>();
+        }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            reader.Position += 4;
+            return 1 + UndertaleSimpleListString.UnserializeChildObjectCount(reader);
         }
 
         /// <inheritdoc/>
