@@ -136,12 +136,29 @@ public class UndertaleTimeline : UndertaleNamedResource, IDisposable
         // Read the actions for each moment
         for (int i = 0; i < momentCount; i++)
         {
-            if (reader.Position != unnecessaryPointers[i])
+            if (reader.AbsPosition != unnecessaryPointers[i])
                 throw new UndertaleSerializationException("Invalid action list pointer");
 
             // Read action list and assign time point (put into list)
             var timeEvent = reader.ReadUndertaleObject<UndertalePointerList<UndertaleGameObject.EventAction>>();
             Moments.Add(new UndertaleTimelineMoment(timePoints[i], timeEvent));
         }
+    }
+
+    /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+    public static uint UnserializeChildObjectCount(UndertaleReader reader)
+    {
+        uint count = 0;
+
+        reader.Position += 4; // "Name"
+
+        int momentCount = reader.ReadInt32();
+
+        reader.Position += (uint)momentCount * 8;
+
+        for (int i = 0; i < momentCount; i++)
+            count += 1 + UndertalePointerList<UndertaleGameObject.EventAction>.UnserializeChildObjectCount(reader);
+
+        return count;
     }
 }
