@@ -1,9 +1,10 @@
-﻿using System;
+using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Diagnostics;
 using System.IO;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using UndertaleModLib.Models;
@@ -46,11 +47,16 @@ namespace UndertaleModLib
         public UndertaleChunkSTRG STRG => Chunks.GetValueOrDefault("STRG") as UndertaleChunkSTRG;
         public UndertaleChunkTXTR TXTR => Chunks.GetValueOrDefault("TXTR") as UndertaleChunkTXTR;
         public UndertaleChunkAUDO AUDO => Chunks.GetValueOrDefault("AUDO") as UndertaleChunkAUDO;
-        // GMS2.3+ for the below chunks
+        // GMS 2.3+ for the below chunks
         public UndertaleChunkACRV ACRV => Chunks.GetValueOrDefault("ACRV") as UndertaleChunkACRV;
         public UndertaleChunkSEQN SEQN => Chunks.GetValueOrDefault("SEQN") as UndertaleChunkSEQN;
         public UndertaleChunkTAGS TAGS => Chunks.GetValueOrDefault("TAGS") as UndertaleChunkTAGS;
         public UndertaleChunkFEAT FEAT => Chunks.GetValueOrDefault("FEAT") as UndertaleChunkFEAT;
+        // GMS 2.3.6+
+        public UndertaleChunkFEDS FEDS => Chunks.GetValueOrDefault("FEDS") as UndertaleChunkFEDS;
+        // GM 2023.2+
+        public UndertaleChunkPSEM PSEM => Chunks.GetValueOrDefault("PSEM") as UndertaleChunkPSEM;
+        public UndertaleChunkPSYS PSYS => Chunks.GetValueOrDefault("PSYS") as UndertaleChunkPSYS;
 
         internal override void SerializeChunk(UndertaleWriter writer)
         {
@@ -62,9 +68,10 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            Chunks.Clear();
+            if (Chunks.Count != 1 || Chunks.Keys.First() != "GEN8")
+                Chunks.Clear();
             ChunksTypeDict.Clear();
-            uint startPos = reader.Position;
+            long startPos = reader.Position;
 
             // First, find the last chunk in the file because of padding changes
             // (also, calculate all present chunks while we're at it)
@@ -80,9 +87,6 @@ namespace UndertaleModLib
             reader.LastChunkName = lastChunk;
             reader.Position = startPos;
 
-            reader.GMS2_3 = reader.AllChunkNames.Contains("SEQN");
-            reader.undertaleData.GMS2_3 = reader.GMS2_3;
-
             // Now, parse the chunks
             while (reader.Position < startPos + Length)
             {
@@ -90,17 +94,80 @@ namespace UndertaleModLib
                 if (chunk != null)
                 {
                     if (Chunks.ContainsKey(chunk.Name))
-                        throw new IOException("Duplicate chunk " + chunk.Name);
+                    {
+                        if (Chunks.Count == 1 && chunk.Name == "GEN8")
+                            Chunks.Clear();
+                        else
+                            throw new IOException("Duplicate chunk " + chunk.Name);
+                    }
+
                     Chunks.Add(chunk.Name, chunk);
                     ChunksTypeDict.Add(chunk.GetType(), chunk);
+
+                    if (reader.ReadOnlyGEN8 && chunk.Name == "GEN8")
+                        return;
                 }
             }
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            uint totalCount = 0;
+
+            long startPos = reader.Position;
+            reader.AllChunkNames = new List<string>();
+            while (reader.Position < reader.Length)
+            {
+                string chunkName = reader.ReadChars(4);
+                reader.AllChunkNames.Add(chunkName);
+                uint length = reader.ReadUInt32();
+                reader.Position += length;
+            }
+            reader.Position = startPos;
+
+            if (reader.AllChunkNames[0] == "GEN8")
+            {
+                UndertaleChunkGEN8 gen8Chunk = new();
+                gen8Chunk.UnserializeGeneralData(reader);
+                Chunks.Add("GEN8", gen8Chunk);
+
+                reader.Position = startPos;
+            }
+
+            while (reader.Position < startPos + Length)
+                totalCount += reader.CountChunkChildObjects();
+
+            return totalCount;
         }
     }
 
     public class UndertaleChunkGEN8 : UndertaleSingleChunk<UndertaleGeneralInfo>
     {
         public override string Name => "GEN8";
+
+        public void UnserializeGeneralData(UndertaleReader reader)
+        {
+            Object = new UndertaleGeneralInfo();
+
+            reader.Position += 8; // Chunk name + length
+
+            reader.Position++; // "IsDebuggerDisabled"
+            Object.BytecodeVersion = reader.ReadByte();
+            reader.undertaleData.UnsupportedBytecodeVersion
+                = Object.BytecodeVersion < 13 || Object.BytecodeVersion > 17;
+            reader.Bytecode14OrLower = Object.BytecodeVersion <= 14;
+
+            reader.Position += 42;
+
+            Object.Major = reader.ReadUInt32();
+            Object.Minor = reader.ReadUInt32();
+            Object.Release = reader.ReadUInt32();
+            Object.Build = reader.ReadUInt32();
+
+            var readVer = (Object.Major, Object.Minor, Object.Release, Object.Build);
+            var detectedVer = UndertaleGeneralInfo.TestForCommonGMSVersions(reader, readVer);
+            (Object.Major, Object.Minor, Object.Release, Object.Build) = detectedVer;
+        }
     }
 
     public class UndertaleChunkOPTN : UndertaleSingleChunk<UndertaleOptions>
@@ -111,6 +178,11 @@ namespace UndertaleModLib
     public class UndertaleChunkLANG : UndertaleSingleChunk<UndertaleLanguage>
     {
         public override string Name => "LANG";
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            return 1;
+        }
     }
 
     public class UndertaleChunkEXTN : UndertaleListChunk<UndertaleExtension>
@@ -118,77 +190,122 @@ namespace UndertaleModLib
         public override string Name => "EXTN";
         public List<byte[]> productIdData = new List<byte[]>();
 
-        internal override void UnserializeChunk(UndertaleReader reader)
+        private static bool checkedFor2022_6, checkedFor2023_4;
+        private void CheckFor2022_6(UndertaleReader reader)
         {
-            if (reader.undertaleData.GMS2_3)
+            if (!reader.undertaleData.IsVersionAtLeast(2, 3) || reader.undertaleData.IsVersionAtLeast(2022, 6))
             {
-                // Check for 2022.6, if possible
-                bool definitely2022_6 = true;
-                uint returnPosition = reader.Position;
+                checkedFor2022_6 = true;
+                return;
+            }
 
-                int extCount = reader.ReadInt32();
-                if (extCount > 0)
+            bool definitely2022_6 = true;
+            long returnPosition = reader.AbsPosition;
+
+            int extCount = reader.ReadInt32();
+            if (extCount > 0)
+            {
+                uint firstExtPtr = reader.ReadUInt32();
+                uint firstExtEndPtr = (extCount >= 2) ? reader.ReadUInt32() /* second ptr */ : (uint)(returnPosition + this.Length);
+
+                reader.AbsPosition = firstExtPtr + 12;
+                uint newPointer1 = reader.ReadUInt32();
+                uint newPointer2 = reader.ReadUInt32();
+
+                if (newPointer1 != reader.AbsPosition)
+                    definitely2022_6 = false; // first pointer mismatch
+                else if (newPointer2 <= reader.AbsPosition || newPointer2 >= (returnPosition + this.Length))
+                    definitely2022_6 = false; // second pointer out of bounds
+                else
                 {
-                    uint firstExtPtr = reader.ReadUInt32();
-                    uint firstExtEndPtr = (extCount >= 2) ? reader.ReadUInt32() /* second ptr */ : (returnPosition + this.Length);
-
-                    reader.Position = firstExtPtr + 12;
-                    uint newPointer1 = reader.ReadUInt32();
-                    uint newPointer2 = reader.ReadUInt32();
-
-                    if (newPointer1 != reader.Position)
-                        definitely2022_6 = false; // first pointer mismatch
-                    else if (newPointer2 <= reader.Position || newPointer2 >= (returnPosition + this.Length))
-                        definitely2022_6 = false; // second pointer out of bounds
-                    else
+                    // Check ending position
+                    reader.AbsPosition = newPointer2;
+                    uint optionCount = reader.ReadUInt32();
+                    if (optionCount > 0)
                     {
-                        // Check ending position
-                        reader.Position = newPointer2;
-                        uint optionCount = reader.ReadUInt32();
-                        if (optionCount > 0)
+                        long newOffsetCheck = reader.AbsPosition + (4 * (optionCount - 1));
+                        if (newOffsetCheck >= (returnPosition + this.Length))
                         {
-                            long newOffsetCheck = reader.Position + (4 * (optionCount - 1));
+                            // Option count would place us out of bounds
+                            definitely2022_6 = false;
+                        }
+                        else
+                        {
+                            reader.Position += (4 * (optionCount - 1));
+                            newOffsetCheck = reader.ReadUInt32() + 12; // jump past last option
                             if (newOffsetCheck >= (returnPosition + this.Length))
                             {
-                                // Option count would place us out of bounds
+                                // Pointer list element would place us out of bounds
                                 definitely2022_6 = false;
                             }
                             else
                             {
-                                reader.Position += (4 * (optionCount - 1));
-                                newOffsetCheck = reader.ReadUInt32() + 12; // jump past last option
-                                if (newOffsetCheck >= (returnPosition + this.Length))
-                                {
-                                    // Pointer list element would place us out of bounds
-                                    definitely2022_6 = false;
-                                }
-                                else
-                                {
-                                    reader.Position = (uint)newOffsetCheck;
-                                }
+                                reader.AbsPosition = (uint)newOffsetCheck;
                             }
-                        }
-                        if (definitely2022_6)
-                        {
-                            if (extCount == 1)
-                            {
-                                reader.Position += 16; // skip GUID data (only one of them)
-                                if (reader.Position % 16 != 0)
-                                    reader.Position += 16 - (reader.Position % 16); // align to chunk end
-                            }
-                            if (reader.Position != firstExtEndPtr)
-                                definitely2022_6 = false;
                         }
                     }
+                    if (definitely2022_6)
+                    {
+                        if (extCount == 1)
+                        {
+                            reader.Position += 16; // skip GUID data (only one of them)
+                            if (reader.AbsPosition % 16 != 0)
+                                reader.Position += 16 - (reader.AbsPosition % 16); // align to chunk end
+                        }
+                        if (reader.AbsPosition != firstExtEndPtr)
+                            definitely2022_6 = false;
+                    }
                 }
-                else
-                    definitely2022_6 = false;
-
-                reader.Position = returnPosition;
-
-                if (definitely2022_6)
-                    reader.undertaleData.GM2022_6 = true;
             }
+            else
+                definitely2022_6 = false;
+
+            reader.AbsPosition = returnPosition;
+
+            if (definitely2022_6)
+                reader.undertaleData.SetGMS2Version(2022, 6);
+
+            checkedFor2022_6 = true;
+        }
+        private void CheckFor2023_4(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsVersionAtLeast(2022, 6) || reader.undertaleData.IsVersionAtLeast(2023, 4))
+            {
+                checkedFor2023_4 = true;
+                return;
+            }
+
+            long returnPosition = reader.Position;
+
+            int extCount = reader.ReadInt32();
+            if (extCount > 0)
+            {
+                // Go to the first extension
+                reader.AbsPosition = reader.ReadUInt32();
+
+                // Skip the miminal amount of strings
+                reader.Position += 4 * 3;
+
+                uint filesPtr = reader.ReadUInt32();
+                uint optionsPtr = reader.ReadUInt32();
+
+                // The file list pointer should be less than the option list pointer.
+                // If it's not true, then "filesPtr" is actually a string pointer, so it's GM 2023.4+.
+                if (filesPtr > optionsPtr)
+                    reader.undertaleData.SetGMS2Version(2023, 4);
+            }
+
+            reader.Position = returnPosition;
+
+            checkedFor2023_4 = true;
+        }
+
+        internal override void UnserializeChunk(UndertaleReader reader)
+        {
+            if (!checkedFor2022_6)
+                CheckFor2022_6(reader);
+            if (!checkedFor2023_4)
+                CheckFor2023_4(reader);
 
             base.UnserializeChunk(reader);
 
@@ -220,6 +337,16 @@ namespace UndertaleModLib
 
                 writer.Write(data);
             }
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            checkedFor2022_6 = false;
+
+            CheckFor2022_6(reader);
+            CheckFor2023_4(reader);
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -288,9 +415,9 @@ namespace UndertaleModLib
         {
             reader.Position -= 4;
             int chunkLength = reader.ReadInt32();
-            uint chunkEnd = reader.Position + (uint)chunkLength;
+            long chunkEnd = reader.AbsPosition + chunkLength;
 
-            uint beginPosition = reader.Position;
+            long beginPosition = reader.Position;
 
             // Figure out where the starts/ends of each shader object are
             int count = reader.ReadInt32();
@@ -299,7 +426,7 @@ namespace UndertaleModLib
             {
                 objectLocations[i] = (uint)reader.ReadInt32();
             }
-            objectLocations[count] = chunkEnd;
+            objectLocations[count] = (uint)chunkEnd;
 
             Dictionary<uint, UndertaleObject> objPool = reader.GetOffsetMap();
             Dictionary<UndertaleObject, uint> objPoolRev = reader.GetOffsetMapRev();
@@ -323,6 +450,117 @@ namespace UndertaleModLib
         public override string Name => "FONT";
         public byte[] Padding;
 
+        private static bool checkedFor2022_2;
+        private static bool checkedFor2023_6;
+        private void CheckForGM2022_2(UndertaleReader reader)
+        {
+           /* This code performs four checks to identify GM2022.2.
+            * First, as you've seen, is the bytecode version.
+            * Second, we assume it is. If there are no Glyphs, we are vindicated by the impossibility of null values there.
+            * Third, we check that the Glyph Length is less than the chunk length. If it's going outside the chunk, that means
+            * that the length was misinterpreted.
+            * Fourth, in case of a terrible fluke causing this to appear valid erroneously, we verify that each pointer leads into the next.
+            * And if someone builds their game so the first pointer is absolutely valid length data and the next font is valid glyph data-
+            * screw it, call Jacky720 when someone constructs that and you want to mod it.
+            * Maybe try..catch on the whole shebang?
+            */
+            if (reader.undertaleData.GeneralInfo?.BytecodeVersion < 17 || reader.undertaleData.IsVersionAtLeast(2022, 2))
+            {
+                checkedFor2022_2 = true;
+                return;
+            }
+
+            long positionToReturn = reader.Position;
+            bool GMS2022_2 = false;
+
+            if (reader.ReadUInt32() > 0) // Font count
+            {
+                uint firstFontPointer = reader.ReadUInt32();
+                reader.AbsPosition = firstFontPointer + 48; // There are 48 bytes of existing metadata.
+                uint glyphsLength = reader.ReadUInt32();
+                GMS2022_2 = true;
+                if ((glyphsLength * 4) > this.Length)
+                {
+                    GMS2022_2 = false;
+                }
+                else if (glyphsLength != 0)
+                {
+                    List<uint> glyphPointers = new List<uint>((int)glyphsLength);
+                    for (uint i = 0; i < glyphsLength; i++)
+                        glyphPointers.Add(reader.ReadUInt32());
+                    foreach (uint pointer in glyphPointers)
+                    {
+                        if (reader.AbsPosition != pointer)
+                        {
+                            GMS2022_2 = false;
+                            break;
+                        }
+
+                        reader.Position += 14;
+                        ushort kerningLength = reader.ReadUInt16();
+                        reader.Position += (uint)4 * kerningLength; // combining read/write would apparently break
+                    }
+                }
+
+            }
+            if (GMS2022_2)
+                reader.undertaleData.SetGMS2Version(2022, 2);
+            reader.Position = positionToReturn;
+
+            checkedFor2022_2 = true;
+        }
+
+        private void CheckForGM2023_6(UndertaleReader reader)
+        {
+            // This is basically the same as the 2022.2 check, but adapted for the LineHeight value instead of Ascender.
+            
+            // We already know whether the version is more or less than 2023.2 due to PSEM. Checking a shorter range narrows possibility of error.
+            if (!reader.undertaleData.IsVersionAtLeast(2023, 2) || reader.undertaleData.IsVersionAtLeast(2023, 6))
+            {
+                checkedFor2023_6 = true;
+                return;
+            }
+
+            long positionToReturn = reader.Position;
+            bool GMS2023_6 = false;
+
+            if (reader.ReadUInt32() > 0) // Font count
+            {
+                uint firstFontPointer = reader.ReadUInt32();
+                reader.AbsPosition = firstFontPointer + 56; // Two more values: SDFSpread and LineHeight. 48 + 4 + 4 = 56.
+                uint glyphsLength = reader.ReadUInt32();
+                GMS2023_6 = true;
+                if ((glyphsLength * 4) > this.Length)
+                {
+                    GMS2023_6 = false;
+                }
+                else if (glyphsLength != 0)
+                {
+                    List<uint> glyphPointers = new List<uint>((int)glyphsLength);
+                    for (uint i = 0; i < glyphsLength; i++)
+                        glyphPointers.Add(reader.ReadUInt32());
+                    foreach (uint pointer in glyphPointers)
+                    {
+                        if (reader.AbsPosition != pointer)
+                        {
+                            GMS2023_6 = false;
+                            break;
+                        }
+
+                        reader.Position += 14;
+                        ushort kerningLength = reader.ReadUInt16();
+                        reader.Position += (uint)4 * kerningLength; // combining read/write would apparently break
+                    }
+                }
+
+            }
+            if (GMS2023_6)
+                reader.undertaleData.SetGMS2Version(2023, 6);
+            reader.Position = positionToReturn;
+
+            checkedFor2023_6 = true;
+        }
+
         internal override void SerializeChunk(UndertaleWriter writer)
         {
             base.SerializeChunk(writer);
@@ -339,55 +577,26 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            if (reader.undertaleData.GeneralInfo?.BytecodeVersion >= 17)
-            {
-                /* This code performs four checks to identify GM2022.2.
-                 * First, as you've seen, is the bytecode version.
-                 * Second, we assume it is. If there are no Glyphs, we are vindicated by the impossibility of null values there.
-                 * Third, we check that the Glyph Length is less than the chunk length. If it's going outside the chunk, that means
-                 * that the length was misinterpreted.
-                 * Fourth, in case of a terrible fluke causing this to appear valid erroneously, we verify that each pointer leads into the next.
-                 * And if someone builds their game so the first pointer is absolutely valid length data and the next font is valid glyph data-
-                 * screw it, call Jacky720 when someone constructs that and you want to mod it.
-                 * Maybe try..catch on the whole shebang?
-                 */
-                uint positionToReturn = reader.Position;
-                if (reader.ReadUInt32() > 0) // Font count
-                {
-                    uint firstFontPointer = reader.ReadUInt32();
-                    reader.Position = firstFontPointer + 48; // There are 48 bytes of existing metadata.
-                    uint glyphsLength = reader.ReadUInt32();
-                    reader.undertaleData.GMS2022_2 = true;
-                    if ((glyphsLength * 4) > this.Length)
-                    {
-                        reader.undertaleData.GMS2022_2 = false;
-                    }
-                    else if (glyphsLength != 0)
-                    {
-                        List<uint> glyphPointers = new List<uint>();
-                        for (uint i = 0; i < glyphsLength; i++)
-                            glyphPointers.Add(reader.ReadUInt32());
-                        foreach (uint pointer in glyphPointers)
-                        {
-                            if (reader.Position != pointer)
-                            {
-                                reader.undertaleData.GMS2022_2 = false;
-                                break;
-                            }
+            if (!checkedFor2022_2)
+                CheckForGM2022_2(reader);
 
-                            reader.Position += 14;
-                            ushort kerningLength = reader.ReadUInt16();
-                            reader.Position += (uint) 4 * kerningLength; // combining read/write would apparently break
-                        }
-                    }
-
-                }
-                reader.Position = positionToReturn;
-            }
+            if (!checkedFor2023_6)
+                CheckForGM2023_6(reader);
 
             base.UnserializeChunk(reader);
 
             Padding = reader.ReadBytes(512);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            checkedFor2022_2 = false;
+            checkedFor2023_6 = false;
+
+            CheckForGM2022_2(reader);
+            CheckForGM2023_6(reader);
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -400,6 +609,50 @@ namespace UndertaleModLib
     {
         public override string Name => "OBJT";
 
+        private static bool checkedFor2022_5;
+
+        // Simple chunk parser to check for 2022.5, assumes old format until shown otherwise
+        private void CheckFor2022_5(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsVersionAtLeast(2, 3) || reader.undertaleData.IsVersionAtLeast(2022, 5))
+            {
+                checkedFor2022_5 = true;
+                return;
+            }
+
+            long positionToReturn = reader.Position;
+            bool GM2022_5 = false;
+
+            if (reader.ReadUInt32() > 0) // Object count
+            {
+                uint firstObjectPointer = reader.ReadUInt32();
+                reader.AbsPosition = firstObjectPointer + 64;
+                uint vertexCount = reader.ReadUInt32();
+
+                // If any of these checks fail, it's 2022.5
+                GM2022_5 = true;
+                // Bounds check on vertex data
+                if (reader.Position + 12 + vertexCount * 8 < positionToReturn + this.Length)
+                {
+                    reader.Position += 12 + vertexCount * 8;
+                    // A pointer list of events
+                    if (reader.ReadUInt32() == UndertaleGameObject.EventTypeCount)
+                    {
+                        uint subEventPointer = reader.ReadUInt32();
+                        // Should start right after the list
+                        if (reader.AbsPosition + 56 == subEventPointer)
+                            GM2022_5 = false;
+                    }
+                }
+            }
+            if (GM2022_5)
+                reader.undertaleData.SetGMS2Version(2022, 5);
+
+            reader.Position = positionToReturn;
+
+            checkedFor2022_5 = true;
+        }
+
         internal override void SerializeChunk(UndertaleWriter writer)
         {
             base.SerializeChunk(writer);
@@ -407,36 +660,19 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            // Simple chunk parser to check for 2022.5, assumes old format until shown otherwise
-            if (reader.undertaleData.GMS2_3)
-            {
-                uint positionToReturn = reader.Position;
-                if (reader.ReadUInt32() > 0) // Object count
-                {
-                    uint firstObjectPointer = reader.ReadUInt32();
-                    reader.Position = firstObjectPointer + 64;
-                    uint vertexCount = reader.ReadUInt32();
-
-                    // If any of these checks fail, it's 2022.5
-                    reader.undertaleData.GM2022_5 = true;
-                    // Bounds check on vertex data
-                    if (reader.Position + 12 + vertexCount * 8 < positionToReturn + this.Length)
-                    {
-                        reader.Position += (uint)(12 + vertexCount * 8);
-                        // 15 events as a pointer list
-                        if (reader.ReadUInt32() == 15)
-                        {
-                            uint subEventPointer = reader.ReadUInt32();
-                            // Should start right after the list
-                            if (reader.Position + 56 == subEventPointer)
-                                reader.undertaleData.GM2022_5 = false;
-                        }
-                    }
-                }
-                reader.Position = positionToReturn;
-            }
+            if (!checkedFor2022_5)
+                CheckFor2022_5(reader);
 
             base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            checkedFor2022_5 = false;
+
+            CheckFor2022_5(reader);
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -446,17 +682,41 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            CheckForEffectData(reader);
+            if (!checkedFor2022_1)
+                CheckForEffectData(reader);
 
             base.UnserializeChunk(reader);
         }
 
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            checkedFor2022_1 = false;
+            UndertaleRoom.CheckedForGMS2_2_2_302 = false;
+
+            CheckForEffectData(reader);
+
+            if (reader.undertaleData.GeneralInfo?.BytecodeVersion >= 16)
+            {
+                // "GameObject._preCreateCode"
+
+                Type gameObjType = typeof(GameObject);
+
+                uint newValue = GameObject.ChildObjectCount + 1;
+                reader.SetStaticChildCount(gameObjType, newValue);
+                newValue = GameObject.ChildObjectsSize + 4;
+                reader.SetStaticChildObjectsSize(gameObjType, newValue);
+            }
+
+            return base.UnserializeObjectCount(reader);
+        }
+
+        private static bool checkedFor2022_1;
         private void CheckForEffectData(UndertaleReader reader)
         {
             // Do a length check on room layers to see if this is 2022.1 or higher
-            if (!reader.undertaleData.GMS2022_1 && reader.undertaleData.GMS2_3)
+            if (reader.undertaleData.IsVersionAtLeast(2, 3) && !reader.undertaleData.IsVersionAtLeast(2022, 1))
             {
-                uint returnTo = reader.Position;
+                long returnTo = reader.Position;
 
                 // Iterate over all rooms until a length check is performed
                 int roomCount = reader.ReadInt32();
@@ -466,17 +726,17 @@ namespace UndertaleModLib
                     // Advance to room data we're interested in (and grab pointer for next room)
                     reader.Position = returnTo + 4 + (4 * roomIndex);
                     uint roomPtr = (uint)reader.ReadInt32();
-                    reader.Position = roomPtr + (22 * 4);
+                    reader.AbsPosition = roomPtr + (22 * 4);
 
                     // Get the pointer for this room's layer list, as well as pointer to sequence list
                     uint layerListPtr = (uint)reader.ReadInt32();
                     int seqnPtr = reader.ReadInt32();
-                    reader.Position = layerListPtr;
+                    reader.AbsPosition = layerListPtr;
                     int layerCount = reader.ReadInt32();
                     if (layerCount >= 1)
                     {
                         // Get pointer into the individual layer data (plus 8 bytes) for the first layer in the room
-                        uint jumpOffset = (uint)(reader.ReadInt32() + 8);
+                        int jumpOffset = reader.ReadInt32() + 8;
 
                         // Find the offset for the end of this layer
                         int nextOffset;
@@ -486,41 +746,41 @@ namespace UndertaleModLib
                             nextOffset = reader.ReadInt32(); // (pointer to next element in the layer list)
 
                         // Actually perform the length checks, depending on layer data
-                        reader.Position = jumpOffset;
+                        reader.AbsPosition = jumpOffset;
                         switch ((LayerType)reader.ReadInt32())
                         {
                             case LayerType.Background:
-                                if (nextOffset - reader.Position > 16 * 4)
-                                    reader.undertaleData.GMS2022_1 = true;
+                                if (nextOffset - reader.AbsPosition > 16 * 4)
+                                    reader.undertaleData.SetGMS2Version(2022, 1);
                                 finished = true;
                                 break;
                             case LayerType.Instances:
                                 reader.Position += 6 * 4;
                                 int instanceCount = reader.ReadInt32();
-                                if (nextOffset - reader.Position != (instanceCount * 4))
-                                    reader.undertaleData.GMS2022_1 = true;
+                                if (nextOffset - reader.AbsPosition != (instanceCount * 4))
+                                    reader.undertaleData.SetGMS2Version(2022, 1);
                                 finished = true;
                                 break;
                             case LayerType.Assets:
                                 reader.Position += 6 * 4;
                                 int tileOffset = reader.ReadInt32();
-                                if (tileOffset != reader.Position + 8)
-                                    reader.undertaleData.GMS2022_1 = true;
+                                if (tileOffset != reader.AbsPosition + 8)
+                                    reader.undertaleData.SetGMS2Version(2022, 1);
                                 finished = true;
                                 break;
                             case LayerType.Tiles:
                                 reader.Position += 7 * 4;
                                 int tileMapWidth = reader.ReadInt32();
                                 int tileMapHeight = reader.ReadInt32();
-                                if (nextOffset - reader.Position != (tileMapWidth * tileMapHeight * 4))
-                                    reader.undertaleData.GMS2022_1 = true;
+                                if (nextOffset - reader.AbsPosition != (tileMapWidth * tileMapHeight * 4))
+                                    reader.undertaleData.SetGMS2Version(2022, 1);
                                 finished = true;
                                 break;
                             case LayerType.Effect:
                                 reader.Position += 7 * 4;
                                 int propertyCount = reader.ReadInt32();
-                                if (nextOffset - reader.Position != (propertyCount * 3 * 4))
-                                    reader.undertaleData.GMS2022_1 = true;
+                                if (nextOffset - reader.AbsPosition != (propertyCount * 3 * 4))
+                                    reader.undertaleData.SetGMS2Version(2022, 1);
                                 finished = true;
                                 break;
                         }
@@ -529,6 +789,8 @@ namespace UndertaleModLib
 
                 reader.Position = returnTo;
             }
+
+            checkedFor2022_1 = true;
         }
     }
 
@@ -587,7 +849,32 @@ namespace UndertaleModLib
                 List = null;
                 return;
             }
+
+            UndertaleCode.CurrCodeIndex = 0;
             base.UnserializeChunk(reader);
+
+            reader.InstructionArraysLengths = null;
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (Length == 0)
+                return 0;
+
+            if (reader.undertaleData.UnsupportedBytecodeVersion)
+                return reader.ReadUInt32();
+
+            int codeCount = (int)reader.ReadUInt32();
+            reader.Position -= 4;
+
+            reader.GMS2BytecodeAddresses = new(codeCount);
+            reader.InstructionArraysLengths = new int[codeCount];
+            UndertaleCode.CurrCodeIndex = 0;
+
+            uint count = base.UnserializeObjectCount(reader);
+            reader.GMS2BytecodeAddresses.Clear();
+
+            return count;
         }
     }
 
@@ -638,7 +925,7 @@ namespace UndertaleModLib
 
             if (reader.undertaleData.UnsupportedBytecodeVersion)
                 return;
-            uint startPosition = reader.Position;
+            long startPosition = reader.Position;
             uint varLength;
             if (!reader.Bytecode14OrLower)
             {
@@ -651,8 +938,26 @@ namespace UndertaleModLib
             else
                 varLength = 12;
             List.Clear();
+            List.Capacity = (int)(Length / varLength);
             while (reader.Position + varLength <= startPosition + Length)
                 List.Add(reader.ReadUndertaleObject<UndertaleVariable>());
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (Length == 0)
+                return 0;
+
+            if (reader.undertaleData.UnsupportedBytecodeVersion)
+                return 0;
+
+            if (!reader.Bytecode14OrLower)
+            {
+                reader.Position += 12;
+                return (Length - 12) / 20;
+            }
+            else
+                return Length / 12;
         }
     }
 
@@ -695,8 +1000,9 @@ namespace UndertaleModLib
                 return;
             if (reader.Bytecode14OrLower)
             {
-                uint startPosition = reader.Position;
+                long startPosition = reader.Position;
                 Functions.Clear();
+                Functions.SetCapacity(Length / 12);
                 while (reader.Position + 12 <= startPosition + Length)
                     Functions.Add(reader.ReadUndertaleObject<UndertaleFunction>());
             }
@@ -705,6 +1011,24 @@ namespace UndertaleModLib
                 Functions = reader.ReadUndertaleObject<UndertaleSimpleList<UndertaleFunction>>();
                 CodeLocals = reader.ReadUndertaleObject<UndertaleSimpleList<UndertaleCodeLocals>>();
             }
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (Length == 0 && reader.undertaleData.GeneralInfo?.BytecodeVersion > 14)
+                return 0;
+
+            uint count = 0;
+            
+            if (!reader.Bytecode14OrLower)
+            {
+                count += 1 + UndertaleSimpleList<UndertaleFunction>.UnserializeChildObjectCount(reader);
+                count += 1 + UndertaleSimpleList<UndertaleCodeLocals>.UnserializeChildObjectCount(reader);
+            }
+            else
+                count = Length / 12;
+
+            return count;
         }
     }
 
@@ -726,15 +1050,87 @@ namespace UndertaleModLib
             base.UnserializeChunk(reader);
 
             // padding
-            while (reader.Position % 0x80 != 0)
+            while (reader.AbsPosition % 0x80 != 0)
                 if (reader.ReadByte() != 0)
                     throw new IOException("Padding error in STRG");
         }
+
+        // There's no need to check padding in "UnserializeObjectCount()"
     }
 
     public class UndertaleChunkTXTR : UndertaleListChunk<UndertaleEmbeddedTexture>
     {
         public override string Name => "TXTR";
+
+        private static bool checkedFor2022_3;
+        private void CheckFor2022_3And5(UndertaleReader reader)
+        {
+            // Detect GM2022.3
+            if (reader.undertaleData.IsVersionAtLeast(2, 3) && !reader.undertaleData.IsVersionAtLeast(2022, 3))
+            {
+                long positionToReturn = reader.Position;
+
+                // Check for 2022.3 format
+                uint texCount = reader.ReadUInt32();
+                if (texCount == 1) // If no textures exist, this could false positive.
+                {
+                    reader.Position += 16; // Jump to either padding or length, depending on version
+                    if (reader.ReadUInt32() > 0) // Check whether it's padding or length
+                        reader.undertaleData.SetGMS2Version(2022, 3);
+                }
+                else if (texCount > 1)
+                {
+                    uint firstTex = reader.ReadUInt32();
+                    uint secondTex = reader.ReadUInt32();
+                    if (firstTex + 16 == secondTex)
+                        reader.undertaleData.SetGMS2Version(2022, 3);
+                }
+
+                if (reader.undertaleData.IsVersionAtLeast(2022, 3))
+                {
+                    // Also check for 2022.5 format
+                    reader.Position = positionToReturn + 4;
+                    for (uint i = 0; i < texCount; i++)
+                    {
+                        // Go to each texture, and then to each texture's data
+                        reader.Position = positionToReturn + 4 + (i * 4);
+                        reader.AbsPosition = reader.ReadUInt32() + 12; // go to texture, at an offset
+                        reader.AbsPosition = reader.ReadUInt32(); // go to texture data
+                        byte[] header = reader.ReadBytes(4);
+                        if (header.SequenceEqual(UndertaleEmbeddedTexture.TexData.QOIAndBZip2Header))
+                        {
+                            reader.Position += 4; // skip width/height
+                            bool is2022_5 = false;
+                            // Now check the actual BZ2 headers
+                            if (reader.ReadByte() != (byte)'B')
+                                is2022_5 = true;
+                            else if (reader.ReadByte() != (byte)'Z')
+                                is2022_5 = true;
+                            else if (reader.ReadByte() != (byte)'h')
+                                is2022_5 = true;
+                            else
+                            {
+                                reader.Position++;
+                                if (reader.ReadUInt24() != 0x594131) // digits of pi... (block header)
+                                    is2022_5 = true;
+                                else if (reader.ReadUInt24() != 0x595326)
+                                    is2022_5 = true;
+                            }
+
+                            if (is2022_5)
+                                reader.undertaleData.SetGMS2Version(2022, 5);
+
+                            // Checked one QOI+BZ2 texture. No need to check any more
+                            break;
+                        }
+                    }
+                }
+
+                reader.Position = positionToReturn;
+            }
+
+            checkedFor2022_3 = true;
+        }
 
         internal override void SerializeChunk(UndertaleWriter writer)
         {
@@ -762,7 +1158,7 @@ namespace UndertaleModLib
                 {
                     // Calculate maximum size of QOI converter buffer
                     maxSize = List.Select(x => x.TextureData.Width * x.TextureData.Height).Max()
-                              * QoiConverter.MaxChunkSize + QoiConverter.HeaderSize + (writer.undertaleData.GM2022_3 ? 0 : 4);
+                              * QoiConverter.MaxChunkSize + QoiConverter.HeaderSize + (writer.undertaleData.IsVersionAtLeast(2022, 3) ? 0 : 4);
                     QoiConverter.InitSharedBuffer(maxSize);
                 }
             }
@@ -775,84 +1171,98 @@ namespace UndertaleModLib
                 writer.Write((byte)0);
         }
 
-        internal override void UnserializeChunk(UndertaleReader reader)
+        // GMS 2.0.6.96 is the oldest available runtime version,
+        // so this actually could be some other version between GMS 2.0 - 2.0.6.
+        // (the oldest copy of "Zeus-Runtime.rss" on web.archive.org has this version as first one)
+        private void CheckForGMS2_0_6(UndertaleReader reader)
         {
-            // Detect GM2022.3
-            if (reader.undertaleData.GM2022_5)
-                reader.undertaleData.GM2022_3 = true;
-            else if (reader.undertaleData.GMS2_3)
+            bool atLeastGMS2_0 = reader.undertaleData.IsGameMaker2();
+            bool atLeastGMS2_2 = reader.undertaleData.IsVersionAtLeast(2, 2);
+            if (!atLeastGMS2_0 || atLeastGMS2_2 || reader.undertaleData.IsVersionAtLeast(2, 0, 6))
+                return;
+
+            long returnPos = reader.Position;
+            bool noGeneratedMips = false;
+
+            uint count = reader.ReadUInt32();
+            if (count == 0)
             {
-                uint positionToReturn = reader.Position;
-
-                // Check for 2022.3 format
-                uint texCount = reader.ReadUInt32();
-                if (texCount == 1) // If no textures exist, this could false positive.
-                {
-                    reader.Position += 16; // Jump to either padding or length, depending on version
-                    if (reader.ReadUInt32() > 0) // Check whether it's padding or length
-                        reader.undertaleData.GM2022_3 = true;
-                }
-                else if (texCount > 1)
-                {
-                    uint firstTex = reader.ReadUInt32();
-                    uint secondTex = reader.ReadUInt32();
-                    if (firstTex + 16 == secondTex)
-                        reader.undertaleData.GM2022_3 = true;
-                }
-
-                if (reader.undertaleData.GM2022_3)
-                {
-                    // Also check for 2022.5 format
-                    reader.Position = positionToReturn + 4;
-                    for (uint i = 0; i < texCount; i++)
-                    {
-                        // Go to each texture, and then to each texture's data
-                        reader.Position = positionToReturn + 4 + (i * 4);
-                        reader.Position = reader.ReadUInt32() + 12; // go to texture, at an offset
-                        reader.Position = reader.ReadUInt32(); // go to texture data
-                        byte[] header = reader.ReadBytes(4);
-                        if (header.SequenceEqual(UndertaleEmbeddedTexture.TexData.QOIAndBZip2Header))
-                        {
-                            reader.Position += 4; // skip width/height
-
-                            // Now check the actual BZ2 headers
-                            if (reader.ReadByte() != (byte)'B')
-                                reader.undertaleData.GM2022_5 = true;
-                            else if (reader.ReadByte() != (byte)'Z')
-                                reader.undertaleData.GM2022_5 = true;
-                            else if (reader.ReadByte() != (byte)'h')
-                                reader.undertaleData.GM2022_5 = true;
-                            else
-                            {
-                                reader.ReadByte();
-                                if (reader.ReadUInt24() != 0x594131) // digits of pi... (block header)
-                                    reader.undertaleData.GM2022_5 = true;
-                                else if (reader.ReadUInt24() != 0x595326)
-                                    reader.undertaleData.GM2022_5 = true;
-                            }
-
-                            // Checked one QOI+BZ2 texture. No need to check any more
-                            break;
-                        }
-                    }
-                }
-
-                reader.Position = positionToReturn;
+                reader.Position = returnPos;
+                return;
             }
 
+            if (count >= 2)
+            {
+                uint firstPtr = reader.ReadUInt32();
+                uint secondPtr = reader.ReadUInt32();
+
+                if (atLeastGMS2_0 && !atLeastGMS2_2) // 2 < version < 2.2
+                {
+                    if (secondPtr - firstPtr == 8) // "Scaled" + "_textureData" -> 8
+                        noGeneratedMips = true;
+                }
+            }
+            else
+            {
+                // Go to the first texture pointer (+ minimal texture entry size)
+                reader.AbsPosition = reader.ReadUInt32() + 8;
+
+                // If there is a zero instead of texture data pointer (which cannot be zero)
+                if (reader.ReadUInt32() == 0)
+                    noGeneratedMips = true;
+            }
+
+            if (!noGeneratedMips)
+                reader.undertaleData.SetGMS2Version(2, 0, 6);
+
+            reader.Position = returnPos;
+        }
+
+        internal override void UnserializeChunk(UndertaleReader reader)
+        {
+            if (!checkedFor2022_3)
+                CheckFor2022_3And5(reader);
+
+            CheckForGMS2_0_6(reader);
+
             base.UnserializeChunk(reader);
+            reader.SwitchReaderType(false);
 
             // texture blobs
-            foreach (UndertaleEmbeddedTexture obj in List)
+            for (int index = 0; index < List.Count; index++)
             {
+                UndertaleEmbeddedTexture obj = List[index];
+
                 obj.UnserializeBlob(reader);
-                obj.Name = new UndertaleString("Texture " + List.IndexOf(obj).ToString());
+                obj.Name = new UndertaleString("Texture " + index.ToString());
             }
 
             // padding
+            // (not "AbsPosition" because of "reader.SwitchReaderType(false)")
             while (reader.Position % 4 != 0)
                 if (reader.ReadByte() != 0)
                     throw new IOException("Padding error!");
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            checkedFor2022_3 = false;
+
+            CheckFor2022_3And5(reader);
+
+            uint txtrSize = UndertaleEmbeddedTexture.ChildObjectsSize;
+            if (reader.undertaleData.IsVersionAtLeast(2, 0, 6))
+                txtrSize += 4; // "GeneratedMips"
+            if (reader.undertaleData.IsVersionAtLeast(2022, 3))
+                txtrSize += 4; // "TextureBlockSize"
+            if (reader.undertaleData.IsVersionAtLeast(2022, 9))
+                txtrSize += 12;
+
+            if (txtrSize != UndertaleEmbeddedTexture.ChildObjectsSize)
+                reader.SetStaticChildObjectsSize(typeof(UndertaleEmbeddedTexture), txtrSize);
+
+            // Texture blobs are already included in the count
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -874,6 +1284,13 @@ namespace UndertaleModLib
                 List[index].Name = new UndertaleString("EmbeddedSound " + index.ToString());
             }
         }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            // Though "UndertaleEmbeddedAudio" has dynamic child objects size, 
+            // there's still no need to unserialize the count for each object.
+            return reader.ReadUInt32();
+        }
     }
 
     // GMS2 only
@@ -883,7 +1300,7 @@ namespace UndertaleModLib
 
         internal override void SerializeChunk(UndertaleWriter writer)
         {
-            if (writer.undertaleData.GeneralInfo.Major < 2)
+            if (!writer.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
             writer.Write((uint)1); // apparently hardcoded 1, see https://github.com/krzys-h/UndertaleModTool/issues/4#issuecomment-421844420
             base.SerializeChunk(writer);
@@ -891,11 +1308,22 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            if (reader.undertaleData.GeneralInfo.Major < 2)
+            if (!reader.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
             if (reader.ReadUInt32() != 1)
                 throw new Exception("Expected EMBI version 1");
             base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsGameMaker2())
+                throw new InvalidOperationException();
+
+            if (reader.ReadUInt32() != 1)
+                throw new Exception("Expected EMBI version 1");
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -904,42 +1332,99 @@ namespace UndertaleModLib
     {
         public override string Name => "TGIN";
 
+        private static bool checkedFor2022_9;
+        private void CheckFor2022_9And2023(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsVersionAtLeast(2, 3)
+                || reader.undertaleData.IsVersionAtLeast(2022, 9) || reader.undertaleData.IsVersionAtLeast(2023, 1))
+            {
+                checkedFor2022_9 = true;
+                return;
+            }
+
+            // Check for 2022.9
+            long returnPosition = reader.AbsPosition;
+
+            bool isGM2022_9 = false;
+            uint tginCount = reader.ReadUInt32();
+            if (tginCount > 0)
+            {
+                uint tginPtr = reader.ReadUInt32();
+                uint secondTginPtr = (tginCount >= 2) ? reader.ReadUInt32() : (uint)(returnPosition + this.Length);
+                reader.AbsPosition = tginPtr + 4;
+
+                // Check to see if the pointer located at this address points within this object
+                // If not, then we know we're using a new format!
+                uint ptr = reader.ReadUInt32();
+                if (ptr < tginPtr || ptr >= secondTginPtr)
+                {
+                    isGM2022_9 = true;
+                    reader.undertaleData.SetGMS2Version(2022, 9);
+                }
+            }
+
+            reader.AbsPosition = returnPosition;
+
+            // Check for 2023.1
+            if (isGM2022_9)
+            {
+                reader.Position += 4; // Skip "tginCount"
+
+                uint firstEntryPtr = reader.ReadUInt32();
+                // Go to the the 4th list pointer of the first TGIN entry.
+                // (either to "Fonts" or "SpineTextures" depending on the version)
+                reader.AbsPosition = firstEntryPtr + 16 + (sizeof(uint) * 3); // +16 = "TexturePages" pointer
+                uint fourthPtr = reader.ReadUInt32();
+
+                // If there's a "TexturePages" count instead of the 5th list pointer.
+                // The count can't be greater than the pointer.
+                // (the list could be either "Tilesets" or "Fonts").
+                if (reader.ReadUInt32() <= fourthPtr)
+                    reader.undertaleData.SetGMS2Version(2023, 1);
+            }
+
+            reader.AbsPosition = returnPosition;
+
+            checkedFor2022_9 = true;
+        }
+
         internal override void SerializeChunk(UndertaleWriter writer)
         {
-            if (writer.undertaleData.GeneralInfo.Major < 2)
+            if (!writer.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
+
             writer.Write((uint)1); // Version
+
             base.SerializeChunk(writer);
         }
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            if (reader.undertaleData.GeneralInfo.Major < 2)
+            if (!reader.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
+
             if (reader.ReadUInt32() != 1)
                 throw new IOException("Expected TGIN version 1");
-            if (reader.undertaleData.GMS2_3)
-            {
-                // Check for 2022.8
-                uint returnPosition = reader.Position;
 
-                uint tginCount = reader.ReadUInt32();
-                if (tginCount > 0)
-                {
-                    uint tginPtr = reader.ReadUInt32();
-                    uint secondTginPtr = (tginCount >= 2) ? reader.ReadUInt32() : (returnPosition + this.Length);
-                    reader.Position = tginPtr + 4;
+            if (!checkedFor2022_9)
+                CheckFor2022_9And2023(reader);
 
-                    // Check to see if the pointer located at this address points within this object
-                    // If not, then we know we're using a new format!
-                    uint ptr = reader.ReadUInt32();
-                    if (ptr < tginPtr || ptr >= secondTginPtr)
-                        reader.undertaleData.GM2022_9 = true;
-                }
-
-                reader.Position = returnPosition;
-            }
             base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            checkedFor2022_9 = false;
+
+            if (!reader.undertaleData.IsGameMaker2())
+                throw new InvalidOperationException();
+
+            if (reader.ReadUInt32() != 1)
+                throw new IOException("Expected TGIN version 1");
+
+            CheckFor2022_9And2023(reader);
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -948,9 +1433,48 @@ namespace UndertaleModLib
     {
         public override string Name => "ACRV";
 
+        private static bool checkedForGMS2_3_1;
+        private void CheckForGMS2_3_1(UndertaleReader reader)
+        {
+            if (reader.undertaleData.IsVersionAtLeast(2, 3, 1))
+            {
+                checkedForGMS2_3_1 = true;
+                return;
+            }
+
+            long returnTo = reader.Position;
+
+            uint count = reader.ReadUInt32();
+            if (count == 0)
+            {
+                reader.Position = returnTo;
+                checkedForGMS2_3_1 = true;
+                return;
+            }
+
+            reader.AbsPosition = reader.ReadUInt32(); // go to the first "Point"
+            reader.Position += 8;
+
+            if (reader.ReadUInt32() != 0) // in 2.3 a int with the value of 0 would be set here,
+            {                             // it cannot be version 2.3 if this value isn't 0
+                reader.undertaleData.SetGMS2Version(2, 3, 1);
+                reader.Position -= 4;
+            }
+            else
+            {
+                if (reader.ReadUInt32() == 0)                      // At all points (besides the first one)
+                    reader.undertaleData.SetGMS2Version(2, 3, 1);  // if BezierX0 equals to 0 (the above check)
+                reader.Position -= 8;                              // then BezierY0 equals to 0 as well (the current check)
+            }
+
+            reader.Position = returnTo;
+
+            checkedForGMS2_3_1 = true;
+        }
+
         internal override void SerializeChunk(UndertaleWriter writer)
         {
-            if (writer.undertaleData.GeneralInfo.Major < 2)
+            if (!writer.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             while (writer.Position % 4 != 0)
@@ -963,18 +1487,38 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            if (reader.undertaleData.GeneralInfo.Major < 2)
+            if (!reader.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             // Padding
-            while (reader.Position % 4 != 0)
+            while (reader.AbsPosition % 4 != 0)
                 if (reader.ReadByte() != 0)
                     throw new IOException("Padding error!");
 
             if (reader.ReadUInt32() != 1)
                 throw new IOException("Expected ACRV version 1");
 
+            if (!checkedForGMS2_3_1)
+                CheckForGMS2_3_1(reader);
+
             base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            checkedForGMS2_3_1 = false;
+
+            // Padding
+            while (reader.AbsPosition % 4 != 0)
+                if (reader.ReadByte() != 0)
+                    throw new IOException("Padding error!");
+
+            if (reader.ReadUInt32() != 1)
+                throw new IOException("Expected ACRV version 1");
+
+            CheckForGMS2_3_1(reader);
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -985,7 +1529,7 @@ namespace UndertaleModLib
 
         internal override void SerializeChunk(UndertaleWriter writer)
         {
-            if (writer.undertaleData.GeneralInfo.Major < 2)
+            if (!writer.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             while (writer.Position % 4 != 0)
@@ -998,7 +1542,7 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            if (reader.undertaleData.GeneralInfo.Major < 2)
+            if (!reader.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             // Apparently SEQN can be empty
@@ -1006,7 +1550,7 @@ namespace UndertaleModLib
                 return;
 
             // Padding
-            while (reader.Position % 4 != 0)
+            while (reader.AbsPosition % 4 != 0)
                 if (reader.ReadByte() != 0)
                     throw new IOException("Padding error!");
 
@@ -1015,6 +1559,27 @@ namespace UndertaleModLib
                 throw new IOException("Expected SEQN version 1, got " + version.ToString());
 
             base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsGameMaker2())
+                throw new InvalidOperationException();
+
+            // Apparently SEQN can be empty
+            if (Length == 0)
+                return 0;
+
+            // Padding
+            while (reader.AbsPosition % 4 != 0)
+                if (reader.ReadByte() != 0)
+                    throw new IOException("Padding error!");
+
+            uint version = reader.ReadUInt32();
+            if (version != 1)
+                throw new IOException("Expected SEQN version 1, got " + version.ToString());
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -1025,7 +1590,7 @@ namespace UndertaleModLib
 
         internal override void SerializeChunk(UndertaleWriter writer)
         {
-            if (writer.undertaleData.GeneralInfo.Major < 2)
+            if (!writer.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             while (writer.Position % 4 != 0)
@@ -1038,11 +1603,11 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            if (reader.undertaleData.GeneralInfo.Major < 2)
+            if (!reader.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             // Padding
-            while (reader.Position % 4 != 0)
+            while (reader.AbsPosition % 4 != 0)
                 if (reader.ReadByte() != 0)
                     throw new IOException("Padding error!");
 
@@ -1050,6 +1615,22 @@ namespace UndertaleModLib
                 throw new IOException("Expected TAGS version 1");
 
             base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsGameMaker2())
+                throw new InvalidOperationException();
+
+            // Padding
+            while (reader.AbsPosition % 4 != 0)
+                if (reader.ReadByte() != 0)
+                    throw new IOException("Padding error!");
+
+            if (reader.ReadUInt32() != 1)
+                throw new IOException("Expected TAGS version 1");
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 
@@ -1060,7 +1641,7 @@ namespace UndertaleModLib
 
         internal override void SerializeChunk(UndertaleWriter writer)
         {
-            if (writer.undertaleData.GeneralInfo.Major < 2)
+            if (!writer.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             while (writer.Position % 4 != 0)
@@ -1073,11 +1654,11 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            if (reader.undertaleData.GeneralInfo.Major < 2)
+            if (!reader.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             // Padding
-            while (reader.Position % 4 != 0)
+            while (reader.AbsPosition % 4 != 0)
                 if (reader.ReadByte() != 0)
                     throw new IOException("Padding error!");
 
@@ -1086,16 +1667,33 @@ namespace UndertaleModLib
 
             base.UnserializeChunk(reader);
         }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsGameMaker2())
+                throw new InvalidOperationException();
+
+            // Padding
+            while (reader.AbsPosition % 4 != 0)
+                if (reader.ReadByte() != 0)
+                    throw new IOException("Padding error!");
+
+            uint version = reader.ReadUInt32();
+            if (version != 1)
+                throw new IOException("Expected FEDS version 1, got " + version.ToString());
+
+            return base.UnserializeObjectCount(reader);
+        }
     }
 
-    // GMS2022.8+ only
+    // GM2022.8+ only
     public class UndertaleChunkFEAT : UndertaleSingleChunk<UndertaleFeatureFlags>
     {
         public override string Name => "FEAT";
 
         internal override void SerializeChunk(UndertaleWriter writer)
         {
-            if (writer.undertaleData.GeneralInfo.Major < 2)
+            if (!writer.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             while (writer.Position % 4 != 0)
@@ -1106,15 +1704,120 @@ namespace UndertaleModLib
 
         internal override void UnserializeChunk(UndertaleReader reader)
         {
-            if (reader.undertaleData.GeneralInfo.Major < 2)
+            if (!reader.undertaleData.IsGameMaker2())
                 throw new InvalidOperationException();
 
             // Padding
-            while (reader.Position % 4 != 0)
+            while (reader.AbsPosition % 4 != 0)
                 if (reader.ReadByte() != 0)
                     throw new IOException("Padding error!");
 
             base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsGameMaker2())
+                throw new InvalidOperationException();
+
+            // Padding
+            while (reader.AbsPosition % 4 != 0)
+                if (reader.ReadByte() != 0)
+                    throw new IOException("Padding error!");
+
+            return base.UnserializeObjectCount(reader);
+        }
+    }
+
+    // GM2023.2+ only
+    public class UndertaleChunkPSEM : UndertaleListChunk<UndertaleParticleSystemEmitter>
+    {
+        public override string Name => "PSEM";
+
+        internal override void SerializeChunk(UndertaleWriter writer)
+        {
+            if (!writer.undertaleData.IsVersionAtLeast(2023, 2))
+                throw new InvalidOperationException();
+
+            writer.Align(4);
+
+            writer.Write((uint)1); // Version
+
+            base.SerializeChunk(writer);
+        }
+
+        internal override void UnserializeChunk(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsVersionAtLeast(2023, 2))
+                throw new InvalidOperationException();
+
+            // Padding
+            reader.Align(4);
+
+            if (reader.ReadUInt32() != 1)
+                throw new IOException("Expected PSEM version 1");
+
+            base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsVersionAtLeast(2023, 2))
+                throw new InvalidOperationException();
+
+            // Padding
+            reader.Align(4);
+
+            uint version = reader.ReadUInt32();
+            if (version != 1)
+                throw new IOException("Expected PSEM version 1, got " + version.ToString());
+
+            return base.UnserializeObjectCount(reader);
+        }
+    }
+    public class UndertaleChunkPSYS : UndertaleListChunk<UndertaleParticleSystem>
+    {
+        public override string Name => "PSYS";
+
+        internal override void SerializeChunk(UndertaleWriter writer)
+        {
+            if (!writer.undertaleData.IsVersionAtLeast(2023, 2))
+                throw new InvalidOperationException();
+
+            writer.Align(4);
+
+            writer.Write((uint)1); // Version
+
+            base.SerializeChunk(writer);
+        }
+
+        internal override void UnserializeChunk(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsVersionAtLeast(2023, 2))
+                throw new InvalidOperationException();
+
+            // Padding
+            reader.Align(4);
+
+            if (reader.ReadUInt32() != 1)
+                throw new IOException("Expected PSYS version 1");
+
+            base.UnserializeChunk(reader);
+        }
+
+        internal override uint UnserializeObjectCount(UndertaleReader reader)
+        {
+            if (!reader.undertaleData.IsVersionAtLeast(2023, 2))
+                throw new InvalidOperationException();
+
+            // Padding
+            reader.Align(4);
+
+            uint version = reader.ReadUInt32();
+            if (version != 1)
+                throw new IOException("Expected PSYS version 1, got " + version.ToString());
+
+            return base.UnserializeObjectCount(reader);
         }
     }
 }
