@@ -1,9 +1,11 @@
 ﻿using System;
+using System.Buffers;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.IO;
 using System.IO.Ports;
 using System.Linq;
+using System.Reflection;
 using System.Text;
 using System.Threading.Tasks;
 using UndertaleModLib.Compiler;
@@ -24,8 +26,11 @@ namespace UndertaleModLib
         int SerializeById(UndertaleWriter writer);
     }
 
-    public class UndertaleResourceById<T, ChunkT> : UndertaleResourceRef, IDisposable where T : UndertaleResource, new() where ChunkT : UndertaleListChunk<T>
+    public class UndertaleResourceById<T, ChunkT> : UndertaleResourceRef, IStaticChildObjectsSize, IDisposable where T : UndertaleResource, new() where ChunkT : UndertaleListChunk<T>
     {
+        /// <inheritdoc cref="IStaticChildObjectsSize.ChildObjectsSize" />
+        public static readonly uint ChildObjectsSize = 4;
+
         public int CachedId { get; set; } = -1;
         public T Resource { get; set; }
 
@@ -140,7 +145,7 @@ namespace UndertaleModLib
         }
     }
 
-    public class UndertaleReader : Util.FileBinaryReader
+    public class UndertaleReader : AdaptiveBinaryReader
     {
         /// <summary>
         /// function to delegate warning messages to
@@ -155,6 +160,8 @@ namespace UndertaleModLib
         private WarningHandlerDelegate WarningHandler;
         private MessageHandlerDelegate MessageHandler;
 
+        public bool ReadOnlyGEN8 { get; set; }
+
         /// <summary>
         /// The detected absolute path of the data file, if a FileStream is passed in, or null otherwise (by default).
         /// Can also be manually changed.
@@ -168,15 +175,19 @@ namespace UndertaleModLib
         public string Directory { get; set; } = null;
 
         public UndertaleReader(Stream input,
-                               WarningHandlerDelegate warningHandler = null, MessageHandlerDelegate messageHandler = null) : base(input)
+                               WarningHandlerDelegate warningHandler = null, MessageHandlerDelegate messageHandler = null,
+                               bool onlyGeneralInfo = false) : base(input)
         {
             WarningHandler = warningHandler;
             MessageHandler = messageHandler;
+            ReadOnlyGEN8 = onlyGeneralInfo;
             if (input is FileStream fs)
             {
                 FilePath = fs.Name;
                 Directory = Path.GetDirectoryName(FilePath);
             }
+
+            FillUnserializeCountDictionaries();
         }
 
         // TODO: This would be more useful if it reported location like the exceptions did
@@ -198,12 +209,16 @@ namespace UndertaleModLib
 
         public string LastChunkName;
         public List<string> AllChunkNames;
-        public bool GMS2_3 = false;
+        //public bool GMS2_3 = false;
         public bool Bytecode14OrLower = false;
 
         public UndertaleChunk ReadUndertaleChunk()
         {
             return UndertaleChunk.Unserialize(this);
+        }
+        public uint CountChunkChildObjects()
+        {
+            return UndertaleChunk.CountChunkChildObjects(this);
         }
 
         private List<UndertaleResourceRef> resUpdate = new List<UndertaleResourceRef>();
@@ -224,6 +239,29 @@ namespace UndertaleModLib
             DebugUtil.Assert(data.FORM.Name == name);
             data.FORM.Length = length;
 
+            long startPos = Position;
+            uint poolSize = 0;
+            if (!ProcessCountExc()) // process an exception from "FillUnserializeCountDictionaries()"
+            {
+                try
+                {
+                    if (!ReadOnlyGEN8)
+                        poolSize = data.FORM.UnserializeObjectCount(this);
+                }
+                catch (Exception e)
+                {
+                    countUnserializeExc = e;
+                    Debug.WriteLine(e);
+
+                    SwitchReaderType(false);
+                }
+            }
+            utListPtrsPool = null;
+
+            InitializePools(poolSize);
+
+            Position = startPos;
+
             var lenReader = EnsureLengthFromHere(data.FORM.Length);
             data.FORM.UnserializeChunk(this);
             lenReader.ToHere();
@@ -233,9 +271,15 @@ namespace UndertaleModLib
                 res.PostUnserialize(this);
             resUpdate.Clear();
 
-            data.BuiltinList = new BuiltinList(data);
-            Decompiler.AssetTypeResolver.InitializeTypes(data);
-            UndertaleEmbeddedTexture.FindAllTextureInfo(data);
+            // Skip if it's "audiogroup*.dat" file
+            if (!FilePath.EndsWith(".dat"))
+            {
+                data.BuiltinList = new BuiltinList(data);
+                Decompiler.AssetTypeResolver.InitializeTypes(data);
+                UndertaleEmbeddedTexture.FindAllTextureInfo(data);
+            }
+
+            ProcessCountExc(poolSize);
 
             return data;
         }
@@ -255,9 +299,198 @@ namespace UndertaleModLib
             throw new IOException("Invalid boolean value: " + a);
         }
 
-        private Dictionary<uint, UndertaleObject> objectPool = new Dictionary<uint, UndertaleObject>();
-        private Dictionary<UndertaleObject, uint> objectPoolRev = new Dictionary<UndertaleObject, uint>();
+        private Dictionary<uint, UndertaleObject> objectPool;
+        private Dictionary<UndertaleObject, uint> objectPoolRev;
         private HashSet<uint> unreadObjects = new HashSet<uint>();
+
+        private Exception countUnserializeExc = null;
+        private readonly Dictionary<Type, Func<UndertaleReader, uint>> unserializeFuncDict = new();
+        private readonly Dictionary<Type, uint> staticObjCountDict = new();
+        private readonly Dictionary<Type, uint> staticObjSizeDict = new();
+        public HashSet<uint> GMS2BytecodeAddresses;
+        public int[] InstructionArraysLengths;
+
+        private readonly BindingFlags publicStaticFlags
+            = BindingFlags.Public | BindingFlags.Static | BindingFlags.FlattenHierarchy;
+        private readonly Type[] readerArgType = { typeof(UndertaleReader) };
+        private readonly Type delegateType = typeof(Func<UndertaleReader, uint>);
+        private readonly Func<UndertaleReader, uint> blankCountFunc = new(_ => { return 0; });
+
+        public ArrayPool<uint> utListPtrsPool = ArrayPool<uint>.Create(100000, 17);
+
+        private bool ProcessCountExc(uint poolSize = 0)
+        {
+            if (countUnserializeExc is not null)
+            {
+                try
+                {
+                    string fileDir = Path.GetDirectoryName(Environment.ProcessPath);
+                    File.WriteAllText(Path.Combine(fileDir, "unserializeCountError.txt"),
+                                      countUnserializeExc.ToString() + "\n"
+                                      + countUnserializeExc.Message + "\n"
+                                      + countUnserializeExc.StackTrace);
+
+                    SubmitWarning("Warning - there was an error while trying to unserialize total object count.\n" +
+                                  "The error log is saved to \"unserializeCountError.txt\"." +
+                                  "Please report that error to UndertaleModTool GitHub.");
+                }
+                catch { }
+
+                countUnserializeExc = null;
+
+                return true;
+            }
+
+            if (poolSize != 0 && poolSize != objectPool.Count)
+            {
+                SubmitWarning("Warning - the estimated object pool size differs from the actual size.\n" +
+                              "Please report this on UndertaleModTool GitHub.");
+            }
+
+            return false;
+        }
+        private void FillUnserializeCountDictionaries()
+        {
+            try
+            {
+                Assembly currAssem = Assembly.GetExecutingAssembly();
+                Type[] allTypes = currAssem.GetTypes();
+
+                Type utObjectType = typeof(UndertaleObject);
+                Type staticObjCountType = typeof(IStaticChildObjCount);
+                Type staticObjSizeType = typeof(IStaticChildObjectsSize);
+
+                allTypes = allTypes.Where(t => t.IsAssignableTo(utObjectType)).ToArray();
+                foreach (Type t in allTypes)
+                {
+                    // It's not possible to call a static method of generic classes without present type argument.
+                    if (t.ContainsGenericParameters)
+                        continue;
+
+                    MethodInfo mi = t.GetMethod("UnserializeChildObjectCount", publicStaticFlags, readerArgType);
+                    if (mi is null)
+                        continue;
+
+                    var func = Delegate.CreateDelegate(delegateType, mi) as Func<UndertaleReader, uint>;
+                    if (func is null)
+                    {
+                        Debug.WriteLine($"Can't create a delegate from MethodInfo of type \"{t.FullName}\"");
+                        continue;
+                    }
+
+                    unserializeFuncDict[t] = func;
+                }
+
+                for (int i = 0; i < allTypes.Length; i++)
+                {
+                    Type t = allTypes[i];
+                    FieldInfo fi;
+                    object res;
+
+                    // It's not supported to get a static field from generic classes without present type argument.
+                    if (t.ContainsGenericParameters)
+                        continue;
+
+                    if (t.IsAssignableTo(staticObjCountType))
+                    {
+                        fi = t.GetField("ChildObjectCount", publicStaticFlags);
+                        if (fi is null)
+                        {
+                            Debug.WriteLine($"Can't get \"ChildObjectCount\" field of \"{t.FullName}\"");
+                            continue;
+                        }
+
+                        res = fi.GetValue(null);
+                        if (res is null)
+                        {
+                            Debug.WriteLine($"Can't get value of \"ChildObjectCount\" of \"{t.FullName}\"");
+                            continue;
+                        }
+
+                        staticObjCountDict[t] = (uint)res;
+                    }
+
+                    if (t.IsAssignableTo(staticObjSizeType))
+                    {
+                        fi = t.GetField("ChildObjectsSize", publicStaticFlags);
+                        if (fi is null)
+                        {
+                            Debug.WriteLine($"Can't get \"ChildObjectsSize\" field of \"{t.FullName}\"");
+                            continue;
+                        }
+
+                        res = fi.GetValue(null);
+                        if (res is null)
+                        {
+                            Debug.WriteLine($"Can't get value of \"ChildObjectsSize\" of \"{t.FullName}\"");
+                            continue;
+                        }
+
+                        staticObjSizeDict[t] = (uint)res;
+                    }
+                }
+            }
+            catch (Exception e)
+            {
+                Debug.WriteLine(e);
+                countUnserializeExc = e;
+            }
+        }
+        public Func<UndertaleReader, uint> GetUnserializeCountFunc(Type objType)
+        {
+            if (!unserializeFuncDict.TryGetValue(objType, out var res))
+            {
+                MethodInfo mi = objType.GetMethod("UnserializeChildObjectCount", publicStaticFlags, readerArgType);
+                if (mi is null)
+                {
+                    Debug.WriteLine($"\"UndertaleReader.unserializeFuncDict\" doesn't contain a method for \"{objType.FullName}\".");
+                    return blankCountFunc;
+                }
+
+                //Debug.WriteLine($"Adding a generic class method for \"{objType.FullName}\" to \"UndertaleReader.unserializeFuncDict\".");
+
+                var func = Delegate.CreateDelegate(delegateType, mi) as Func<UndertaleReader, uint>;
+                if (func is null)
+                {
+                    Debug.WriteLine($"Can't create a delegate from MethodInfo of type \"{objType.FullName}\"");
+                    return blankCountFunc;
+                }
+
+                unserializeFuncDict[objType] = func;
+
+                res = func;
+            }
+
+            return res;
+        }
+        public uint GetStaticChildCount(Type objType)
+        {
+            if (!staticObjCountDict.TryGetValue(objType, out uint res))
+            {
+                Debug.WriteLine($"\"UndertaleReader.staticObjCountDict\" doesn't contain type \"{objType.FullName}\".");
+                return 0;
+            }
+
+            return res;
+        }
+        public uint GetStaticChildObjectsSize(Type objType)
+        {
+            if (!staticObjSizeDict.TryGetValue(objType, out uint res))
+            {
+                Debug.WriteLine($"\"UndertaleReader.staticObjSizeDict\" doesn't contain type \"{objType.FullName}\".");
+                return 0;
+            }
+
+            return res;
+        }
+        public void SetStaticChildCount(Type objType, uint count)
+        {
+            staticObjCountDict[objType] = count;
+        }
+        public void SetStaticChildObjectsSize(Type objType, uint size)
+        {
+            staticObjSizeDict[objType] = size;
+        }
 
         public Dictionary<uint, UndertaleObject> GetOffsetMap()
         {
@@ -268,6 +501,48 @@ namespace UndertaleModLib
         {
             return objectPoolRev;
         }
+
+        public void InitializePools(uint objCount = 0)
+        {
+            if (objCount == 0)
+            {
+                objectPool = new();
+                objectPoolRev = new();
+            }
+            else
+            {
+                int objCountInt = (int)objCount;
+                objectPool = new(objCountInt);
+                objectPoolRev = new(objCountInt);
+            }
+        }
+
+        public uint GetChildObjectCount(Type t)
+        {
+            if (!unserializeFuncDict.TryGetValue(t, out var func))
+            {
+                if (staticObjSizeDict.TryGetValue(t, out uint size))
+                {
+                    Position += size;
+
+                    staticObjCountDict.TryGetValue(t, out uint subCount);
+
+                    return subCount;
+                }
+
+                throw new UndertaleSerializationException(
+                    $"\"UndertaleReader.unserializeFuncDict\" doesn't contain a method for \"{t.FullName}\".");
+            }
+
+            return func(this);
+        }
+        public uint GetChildObjectCount<T>() where T : UndertaleObject
+        {
+            Type t = typeof(T);
+
+            return GetChildObjectCount(t);
+        }
+        
 
         public T GetUndertaleObjectAtAddress<T>(uint address) where T : UndertaleObject, new()
         {
@@ -296,61 +571,23 @@ namespace UndertaleModLib
             try
             {
                 var expectedAddress = GetAddressForUndertaleObject(obj);
-                if (expectedAddress != Position)
+                if (expectedAddress != AbsPosition)
                 {
-                    SubmitWarning("Reading misaligned at " + Position.ToString("X8") + ", realigning back to " + expectedAddress.ToString("X8") + "\nHIGH RISK OF DATA LOSS! The file is probably corrupted, or uses unsupported features\nProceed at your own risk");
-                    Position = expectedAddress;
+                    SubmitWarning("Reading misaligned at " + AbsPosition.ToString("X8") + ", realigning back to " + expectedAddress.ToString("X8") + "\nHIGH RISK OF DATA LOSS! The file is probably corrupted, or uses unsupported features\nProceed at your own risk");
+                    AbsPosition = expectedAddress;
                 }
-                unreadObjects.Remove(Position);
+                unreadObjects.Remove((uint)AbsPosition);
                 obj.Unserialize(this);
             }
             catch (Exception e)
             {
-                throw new UndertaleSerializationException(e.Message + "\nat " + Position.ToString("X8") + " while reading object " + typeof(T).FullName, e);
-            }
-        }
-
-        public void ReadUndertaleObject<T>(T obj, uint endPosition) where T : UndertaleObjectEndPos, new()
-        {
-            try
-            {
-                var expectedAddress = GetAddressForUndertaleObject(obj);
-                if (expectedAddress != Position)
-                {
-                    SubmitWarning("Reading misaligned at " + Position.ToString("X8") + ", realigning back to " + expectedAddress.ToString("X8") + "\nHIGH RISK OF DATA LOSS! The file is probably corrupted, or uses unsupported features\nProceed at your own risk");
-                    Position = expectedAddress;
-                }
-                unreadObjects.Remove(Position);
-                obj.Unserialize(this, endPosition);
-            }
-            catch (Exception e)
-            {
-                throw new UndertaleSerializationException(e.Message + "\nat " + Position.ToString("X8") + " while reading object " + typeof(T).FullName, e);
-            }
-        }
-
-        public void ReadUndertaleObject<T>(T obj, int length) where T : UndertaleObjectLenCheck, new()
-        {
-            try
-            {
-                var expectedAddress = GetAddressForUndertaleObject(obj);
-                if (expectedAddress != Position)
-                {
-                    SubmitWarning("Reading misaligned at " + Position.ToString("X8") + ", realigning back to " + expectedAddress.ToString("X8") + "\nHIGH RISK OF DATA LOSS! The file is probably corrupted, or uses unsupported features\nProceed at your own risk");
-                    Position = expectedAddress;
-                }
-                unreadObjects.Remove(Position);
-                obj.Unserialize(this, length);
-            }
-            catch (Exception e)
-            {
-                throw new UndertaleSerializationException(e.Message + "\nat " + Position.ToString("X8") + " while reading object " + typeof(T).FullName, e);
+                throw new UndertaleSerializationException(e.Message + "\nat " + AbsPosition.ToString("X8") + " while reading object " + typeof(T).FullName, e);
             }
         }
 
         public T ReadUndertaleObject<T>() where T : UndertaleObject, new()
         {
-            T obj = GetUndertaleObjectAtAddress<T>(Position);
+            T obj = GetUndertaleObjectAtAddress<T>((uint)AbsPosition);
             ReadUndertaleObject(obj);
             return obj;
         }
@@ -374,6 +611,9 @@ namespace UndertaleModLib
 
         public void ThrowIfUnreadObjects()
         {
+            if (ReadOnlyGEN8)
+                return;
+
             if (unreadObjects.Count > 0)
             {
                 throw new IOException("Found pointer targets that were never read:\n" + String.Join("\n", unreadObjects.Take(10).Select((x) => "0x" + x.ToString("X8") + " (" + objectPool[x].GetType().Name + ")")) + (unreadObjects.Count > 10 ? "\n(and more, " + unreadObjects.Count + " total)" : ""));
@@ -398,9 +638,9 @@ namespace UndertaleModLib
                 if (length != expectedLength)
                 {
                     int diff = (int)expectedLength - (int)length;
-                    Console.WriteLine("WARNING: File specified length " + expectedLength + ", but read only " + length + " (" + diff + " padding?)");
+                    reader.SubmitMessage("WARNING: File specified length " + expectedLength + ", but read only " + length + " (" + diff + " padding?)");
                     if (diff > 0)
-                        reader.Position = reader.Position + (uint)diff;
+                        reader.Position += (uint)diff;
                     else
                         throw new IOException("Read underflow");
                 }
@@ -409,7 +649,7 @@ namespace UndertaleModLib
 
         public void Align(int alignment, byte paddingbyte = 0x00)
         {
-            while ((Position & (alignment - 1)) != paddingbyte)
+            while ((AbsPosition & (alignment - 1)) != paddingbyte)
             {
                 DebugUtil.Assert(ReadByte() == paddingbyte, "Invalid alignment padding");
             }
@@ -421,7 +661,7 @@ namespace UndertaleModLib
         }
     }
 
-    public class UndertaleWriter : Util.FileBinaryWriter
+    public class UndertaleWriter : FileBinaryWriter
     {
         internal UndertaleData undertaleData;
 
@@ -665,9 +905,10 @@ namespace UndertaleModLib
         public static bool IsDictionaryCleared { get; set; } = true;
 
         public static UndertaleData Read(Stream stream, UndertaleReader.WarningHandlerDelegate warningHandler = null,
-                                                        UndertaleReader.MessageHandlerDelegate messageHandler = null)
+                                                        UndertaleReader.MessageHandlerDelegate messageHandler = null,
+                                                        bool onlyGeneralInfo = false)
         {
-            UndertaleReader reader = new UndertaleReader(stream, warningHandler, messageHandler);
+            UndertaleReader reader = new UndertaleReader(stream, warningHandler, messageHandler, onlyGeneralInfo);
             var data = reader.ReadUndertaleData();
             reader.ThrowIfUnreadObjects();
             return data;
