@@ -3,14 +3,59 @@ using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Collections.Specialized;
 using System.IO;
+using System.Reflection;
 using UndertaleModLib.Models;
 
 namespace UndertaleModLib
 {
-    public class UndertaleSimpleList<T> : ObservableCollection<T>, UndertaleObject where T : UndertaleObject, new()
+    public abstract class UndertaleListBase<T> : ObservableCollection<T>, UndertaleObject
+    {
+        private readonly List<T> internalList;
+
+        public UndertaleListBase()
+        {
+            try
+            {
+                FieldInfo itemsField = typeof(Collection<T>)
+                                       .GetField("items", BindingFlags.NonPublic | BindingFlags.Instance);
+                internalList = (List<T>)itemsField.GetValue(this);
+
+            }
+            catch (Exception e)
+            {
+                throw new UndertaleSerializationException($"{e.Message}\nwhile trying to initialize \"UndertalePointerList<{typeof(T).FullName}>\".");
+            }
+        }
+
+        /// <inheritdoc />
+        public abstract void Serialize(UndertaleWriter writer);
+
+        /// <inheritdoc />
+        public abstract void Unserialize(UndertaleReader reader);
+
+        public void SetCapacity(int capacity)
+        {
+            try
+            {
+                internalList.Capacity = capacity;
+            }
+            catch (Exception e)
+            {
+                throw new UndertaleSerializationException($"{e.Message}\nwhile trying to \"SetCapacity()\" of \"UndertalePointerList<{typeof(T).FullName}>\".");
+            }
+        }
+        public void SetCapacity(uint capacity) => SetCapacity((int)capacity);
+
+        public void InternalAdd(T item)
+        {
+            internalList.Add(item);
+        }
+    }
+
+    public class UndertaleSimpleList<T> : UndertaleListBase<T> where T : UndertaleObject, new()
     {
         /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public override void Serialize(UndertaleWriter writer)
         {
             writer.Write((uint)Count);
             for (int i = 0; i < Count; i++)
@@ -27,15 +72,16 @@ namespace UndertaleModLib
         }
 
         /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
+        public override void Unserialize(UndertaleReader reader)
         {
             uint count = reader.ReadUInt32();
             Clear();
+            SetCapacity(count);
             for (uint i = 0; i < count; i++)
             {
                 try
                 {
-                    Add(reader.ReadUndertaleObject<T>());
+                    InternalAdd(reader.ReadUndertaleObject<T>());
                 }
                 catch (UndertaleSerializationException e)
                 {
@@ -43,12 +89,59 @@ namespace UndertaleModLib
                 }
             }
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            uint count = reader.ReadUInt32();
+            if (count == 0)
+                return 0;
+
+            uint totalCount = 0;
+
+            Type t = typeof(T);
+            if (t.IsAssignableTo(typeof(UndertaleResourceRef)))
+            {
+                // UndertaleResourceById<T, ChunkT> = 4 bytes
+                reader.Position += count * 4;
+
+                return count;
+            }
+
+            if (t.IsAssignableTo(typeof(IStaticChildObjectsSize)))
+            {
+                uint subSize = reader.GetStaticChildObjectsSize(t);
+                uint subCount = 0;
+
+                if (t.IsAssignableTo(typeof(IStaticChildObjCount)))
+                    subCount = reader.GetStaticChildCount(t);
+
+                reader.Position += count * subSize;
+
+                return count + count * subCount;
+            }
+
+            var unserializeFunc = reader.GetUnserializeCountFunc(t);
+            for (uint i = 0; i < count; i++)
+            {
+                try
+                {
+                    totalCount += 1 + unserializeFunc(reader);
+                }
+                catch (UndertaleSerializationException e)
+                {
+                    throw new UndertaleSerializationException(e.Message + "\nwhile reading child object count of item " + (i + 1) + " of " + count + " in a list of " + typeof(T).FullName, e);
+                }
+            }
+
+            return totalCount;
+        }
     }
 
-    public class UndertaleSimpleListString : ObservableCollection<UndertaleString>, UndertaleObject
+    public class UndertaleSimpleListString : UndertaleListBase<UndertaleString>
     {
         /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public override void Serialize(UndertaleWriter writer)
         {
             writer.Write((uint)Count);
             for (int i = 0; i < Count; i++)
@@ -65,15 +158,16 @@ namespace UndertaleModLib
         }
 
         /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
+        public override void Unserialize(UndertaleReader reader)
         {
             uint count = reader.ReadUInt32();
             Clear();
+            SetCapacity(count);
             for (uint i = 0; i < count; i++)
             {
                 try
                 {
-                    Add(reader.ReadUndertaleString());
+                    InternalAdd(reader.ReadUndertaleString());
                 }
                 catch (UndertaleSerializationException e)
                 {
@@ -81,23 +175,33 @@ namespace UndertaleModLib
                 }
             }
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            uint count = reader.ReadUInt32();
+            reader.Position += count * 4;
+            return 0;
+        }
     }
 
-    public class UndertaleSimpleListShort<T> : ObservableCollection<T>, UndertaleObject where T : UndertaleObject, new()
+    public class UndertaleSimpleListShort<T> : UndertaleListBase<T> where T : UndertaleObject, new()
     {
-        public UndertaleSimpleListShort()
+        private void EnsureShortCount()
         {
-            base.CollectionChanged += EnsureShortCount;
-        }
-
-        private void EnsureShortCount(object sender, NotifyCollectionChangedEventArgs e)
-        {
-            if (e.NewItems != null && e.NewItems.Count > Int16.MaxValue)
+            if (Count > Int16.MaxValue)
                 throw new InvalidOperationException("Count of short SimpleList exceeds maximum number allowed.");
         }
 
+        /// <inheritdoc cref="Collection{T}.Add(T)"/>
+        public new void Add(T item)
+        {
+            base.Add(item);
+            EnsureShortCount();
+        }
+
         /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public override void Serialize(UndertaleWriter writer)
         {
             writer.Write((ushort)Count);
             for (int i = 0; i < Count; i++)
@@ -114,15 +218,16 @@ namespace UndertaleModLib
         }
 
         /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
+        public override void Unserialize(UndertaleReader reader)
         {
             ushort count = reader.ReadUInt16();
             Clear();
+            SetCapacity(count);
             for (ushort i = 0; i < count; i++)
             {
                 try
                 {
-                    Add(reader.ReadUndertaleObject<T>());
+                    InternalAdd(reader.ReadUndertaleObject<T>());
                 }
                 catch (UndertaleSerializationException e)
                 {
@@ -130,12 +235,51 @@ namespace UndertaleModLib
                 }
             }
         }
+
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
+        {
+            ushort count = reader.ReadUInt16();
+            if (count == 0)
+                return 0;
+
+            uint totalCount = 0;
+
+            Type t = typeof(T);
+            if (t.IsAssignableTo(typeof(IStaticChildObjectsSize)))
+            {
+                uint subSize = reader.GetStaticChildObjectsSize(t);
+                uint subCount = 0;
+
+                if (t.IsAssignableTo(typeof(IStaticChildObjCount)))
+                    subCount = reader.GetStaticChildCount(t);
+
+                reader.Position += count * subSize;
+
+                return count + count * subCount;
+            }
+
+            var unserializeFunc = reader.GetUnserializeCountFunc(t);
+            for (uint i = 0; i < count; i++)
+            {
+                try
+                {
+                    totalCount += 1 + unserializeFunc(reader);
+                }
+                catch (UndertaleSerializationException e)
+                {
+                    throw new UndertaleSerializationException(e.Message + "\nwhile reading child object count of item " + (i + 1) + " of " + count + " in a list of " + typeof(T).FullName, e);
+                }
+            }
+
+            return totalCount;
+        }
     }
 
-    public class UndertalePointerList<T> : ObservableCollection<T>, UndertaleObject where T : UndertaleObject, new()
+    public class UndertalePointerList<T> : UndertaleListBase<T> where T : UndertaleObject, new()
     {
         /// <inheritdoc />
-        public void Serialize(UndertaleWriter writer)
+        public override void Serialize(UndertaleWriter writer)
         {
             writer.Write((uint)Count);
             foreach (T obj in this)
@@ -155,13 +299,15 @@ namespace UndertaleModLib
             {
                 try
                 {
-                    (this[i] as PrePaddedObject)?.SerializePrePadding(writer);
+                    T obj = this[i];
 
-                    writer.WriteUndertaleObject<T>(this[i]);
+                    (obj as PrePaddedObject)?.SerializePrePadding(writer);
+
+                    writer.WriteUndertaleObject<T>(obj);
 
                     // The last object does NOT get padding (TODO: at least in AUDO)
                     if (i != Count - 1)
-                        (this[i] as PaddedObject)?.SerializePadding(writer);
+                        (obj as PaddedObject)?.SerializePadding(writer);
                 }
                 catch (UndertaleSerializationException e)
                 {
@@ -171,40 +317,49 @@ namespace UndertaleModLib
         }
 
         /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader)
+        public override void Unserialize(UndertaleReader reader)
         {
             uint count = reader.ReadUInt32();
             Clear();
+            SetCapacity(count);
             for (uint i = 0; i < count; i++)
             {
                 try
                 {
-                    Add(reader.ReadUndertaleObjectPointer<T>());
+                    InternalAdd(reader.ReadUndertaleObjectPointer<T>());
                 }
                 catch (UndertaleSerializationException e)
                 {
                     throw new UndertaleSerializationException(e.Message + "\nwhile reading pointer to item " + (i + 1) + " of " + count + " in a list of " + typeof(T).FullName, e);
                 }
             }
-            if (Count > 0 && reader.Position != reader.GetAddressForUndertaleObject(this[0]))
+            if (Count > 0)
             {
-                int skip = (int)reader.GetAddressForUndertaleObject(this[0]) - (int)reader.Position;
-                if (skip > 0)
+                uint pos = reader.GetAddressForUndertaleObject(this[0]);
+                if (reader.AbsPosition != pos)
                 {
-                    //Console.WriteLine("Skip " + skip + " bytes of blobs");
-                    reader.Position = reader.Position + (uint)skip;
+                    long skip = pos - reader.AbsPosition;
+                    if (skip > 0)
+                    {
+                        //Console.WriteLine("Skip " + skip + " bytes of blobs");
+                        reader.AbsPosition += skip;
+                    }
+                    else
+                        throw new IOException("First list item starts inside the pointer list?!?!");
                 }
-                else
-                    throw new IOException("First list item starts inside the pointer list?!?!");
             }
             for (uint i = 0; i < count; i++)
             {
                 try
                 {
-                    (this[(int)i] as PrePaddedObject)?.UnserializePrePadding(reader);
-                    reader.ReadUndertaleObject(this[(int)i]);
+                    T obj = this[(int)i];
+
+                    (obj as PrePaddedObject)?.UnserializePrePadding(reader);
+
+                    reader.ReadUndertaleObject(obj);
+
                     if (i != count - 1)
-                        (this[(int)i] as PaddedObject)?.UnserializePadding(reader);
+                        (obj as PaddedObject)?.UnserializePadding(reader);
                 }
                 catch (UndertaleSerializationException e)
                 {
@@ -212,57 +367,62 @@ namespace UndertaleModLib
                 }
             }
         }
-    }
 
-    public class UndertalePointerListLenCheck<T> : UndertalePointerList<T>, UndertaleObjectEndPos where T : UndertaleObjectLenCheck, new()
-    {
-        /// <inheritdoc />
-        public void Unserialize(UndertaleReader reader, uint endPosition)
+        /// <inheritdoc cref="UndertaleObject.UnserializeChildObjectCount(UndertaleReader)"/>
+        public static uint UnserializeChildObjectCount(UndertaleReader reader)
         {
             uint count = reader.ReadUInt32();
-            Clear();
-            List<uint> pointers = new List<uint>();
-            for (uint i = 0; i < count; i++)
+            if (count == 0)
+                return 0;
+
+            uint totalCount = 0;
+
+            Type t = typeof(T);
+            if (t.IsAssignableTo(typeof(IStaticChildObjectsSize)))
             {
-                try
-                {
-                    uint ptr = reader.ReadUInt32();
-                    pointers.Add(ptr);
-                    Add(reader.GetUndertaleObjectAtAddress<T>(ptr));
-                }
-                catch (UndertaleSerializationException e)
-                {
-                    throw new UndertaleSerializationException(e.Message + "\nwhile reading pointer to item " + (i + 1) + " of " + count + " in a list of " + typeof(T).FullName, e);
-                }
+                uint subSize = reader.GetStaticChildObjectsSize(t);
+                uint subCount = 0;
+
+                if (t.IsAssignableTo(typeof(IStaticChildObjCount)))
+                    subCount = reader.GetStaticChildCount(t);
+
+                reader.Position += count * 4 + count * subSize;
+
+                return count + count * subCount;
             }
-            if (Count > 0 && reader.Position != reader.GetAddressForUndertaleObject(this[0]))
+
+            uint[] pointers = reader.utListPtrsPool.Rent((int)count);
+            for (uint i = 0; i < count; i++)
+                pointers[i] = reader.ReadUInt32();
+
+            uint pos = pointers[0];
+            if (reader.AbsPosition != pos)
             {
-                int skip = (int)reader.GetAddressForUndertaleObject(this[0]) - (int)reader.Position;
+                long skip = pos - reader.AbsPosition;
                 if (skip > 0)
-                {
-                    //Console.WriteLine("Skip " + skip + " bytes of blobs");
-                    reader.Position = reader.Position + (uint)skip;
-                }
+                    reader.AbsPosition += skip;
                 else
                     throw new IOException("First list item starts inside the pointer list?!?!");
             }
+
+            var unserializeFunc = reader.GetUnserializeCountFunc(t);
             for (uint i = 0; i < count; i++)
             {
                 try
                 {
-                    (this[(int)i] as PrePaddedObject)?.UnserializePrePadding(reader);
-                    if ((i + 1) < count)
-                        reader.ReadUndertaleObject(this[(int)i], (int)(pointers[(int)i + 1] - reader.Position));
-                    else
-                        reader.ReadUndertaleObject(this[(int)i], (int)(endPosition - reader.Position));
-                    if (i != count - 1)
-                        (this[(int)i] as PaddedObject)?.UnserializePadding(reader);
+                    reader.AbsPosition = pointers[i];
+                    totalCount += 1 + unserializeFunc(reader);
                 }
                 catch (UndertaleSerializationException e)
                 {
-                    throw new UndertaleSerializationException(e.Message + "\nwhile reading item " + (i + 1) + " of " + count + " in a list of " + typeof(T).FullName, e);
+                    reader.utListPtrsPool.Return(pointers);
+                    throw new UndertaleSerializationException(e.Message + "\nwhile reading child object count of item " + (i + 1) + " of " + count + " in a list of " + typeof(T).FullName, e);
                 }
             }
+
+            reader.utListPtrsPool.Return(pointers);
+
+            return totalCount;
         }
     }
 
