@@ -504,10 +504,10 @@ namespace UndertaleModTool
                                 resBMP.RotateFlip(RotateFlipType.Rotate90FlipNone);
                                 break;
                             case 5:
-                                resBMP.RotateFlip(RotateFlipType.Rotate90FlipY);
+                                resBMP.RotateFlip(RotateFlipType.Rotate270FlipY);
                                 break;
                             case 6:
-                                resBMP.RotateFlip(RotateFlipType.Rotate270FlipY);
+                                resBMP.RotateFlip(RotateFlipType.Rotate90FlipY);
                                 break;
                             case 7:
                                 resBMP.RotateFlip(RotateFlipType.Rotate270FlipNone);
