@@ -4,6 +4,7 @@ using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 using UndertaleModLib.Models;
+using UndertaleModTool.Windows;
 
 namespace UndertaleModTool
 {
@@ -12,9 +13,43 @@ namespace UndertaleModTool
     /// </summary>
     public partial class UndertaleBackgroundEditor : DataUserControl
     {
+        private static readonly MainWindow mainWindow = Application.Current.MainWindow as MainWindow;
+        private readonly ContextMenuDark tileContextMenu = new();
+
         public UndertaleBackgroundEditor()
         {
             InitializeComponent();
+
+            var item = new MenuItem()
+            {
+                Header = "Find all references of this tile"
+            };
+            item.Click += FindAllTileReferencesItem_Click;
+            tileContextMenu.Items.Add(item);
+        }
+
+        private void FindAllTileReferencesItem_Click(object sender, RoutedEventArgs e)
+        {
+            var tileSet = DataContext as UndertaleBackground;
+            var selectedID = TileIdList.SelectedItem as UndertaleBackground.TileID;
+            if (tileSet is null || selectedID is null)
+                return;
+
+            FindReferencesResults dialog = null;
+            try
+            {
+                var typeList = new HashSetTypesOverride() { typeof(UndertaleRoom.Layer) };
+                var tuple = (tileSet, selectedID);
+                var results = UndertaleResourceReferenceMethodsMap.GetReferencesOfObject(tuple, mainWindow.Data, typeList);
+                dialog = new(tuple, mainWindow.Data, results);
+                dialog.Show();
+            }
+            catch (Exception ex)
+            {
+                mainWindow.ShowError("An error occured in the object references related window.\n" +
+                                    $"Please report this on GitHub.\n\n{ex}");
+                dialog?.Close();
+            }
         }
 
         private void DataGrid_SelectionChanged(object sender, SelectionChangedEventArgs e)
@@ -30,10 +65,32 @@ namespace UndertaleModTool
             }
         }
 
-        private void BGTexture_MouseLeftButtonDown(object sender, MouseButtonEventArgs e)
+        private bool ScrollTileIntoView(int tileIndex)
+        {
+            ScrollViewer tileListViewer = MainWindow.FindVisualChild<ScrollViewer>(TileIdList);
+            if (tileListViewer is null)
+            {
+                mainWindow.ShowError("Cannot find the tile ID list scroll viewer.");
+                return false;
+            }
+            tileListViewer.ScrollToVerticalOffset(tileIndex + 1 - (tileListViewer.ViewportHeight / 2)); // DataGrid offset is logical
+            tileListViewer.UpdateLayout();
+
+            ScrollViewer dataEditorViewer = mainWindow.DataEditor.Parent as ScrollViewer;
+            double initOffset = dataEditorViewer?.VerticalOffset ?? 0;
+
+            TileIdList.SelectedIndex = tileIndex;
+            (TileIdList.ItemContainerGenerator.ContainerFromIndex(tileIndex) as DataGridRow)?.MoveFocus(new TraversalRequest(FocusNavigationDirection.Next));
+
+            dataEditorViewer.UpdateLayout();
+            dataEditorViewer.ScrollToVerticalOffset(initOffset);
+
+            return true;
+        }
+        private bool SelectTileRegion(object sender, MouseButtonEventArgs e)
         {
             if (!TileRectangle.IsVisible) // MainWindow.IsGMS2
-                return;
+                return false;
 
             Point pos = e.GetPosition((IInputElement)sender);
             UndertaleBackground bg = DataContext as UndertaleBackground;
@@ -41,30 +98,34 @@ namespace UndertaleModTool
             int y = (int)((int)pos.Y / (bg.GMS2TileHeight + (2 * bg.GMS2OutputBorderY)));
             int tileID = (int)((bg.GMS2TileColumns * y) + x);
             if (tileID > bg.GMS2TileCount - 1)
-                return;
+                return false;
 
             e.Handled = true;
 
             int tileIndex = bg.GMS2TileIds.FindIndex(x => x.ID == tileID);
+            if (tileIndex == -1)
+                return false;
 
-            ScrollViewer tileListViewer = MainWindow.FindVisualChild<ScrollViewer>(TileIdList);
-            tileListViewer.ScrollToVerticalOffset(tileIndex + 1 - (tileListViewer.ViewportHeight / 2)); // DataGrid offset is logical
-            tileListViewer.UpdateLayout();
+            return ScrollTileIntoView(tileIndex);
+        }
+        private void BGTexture_MouseLeftButtonDown(object sender, MouseButtonEventArgs e)
+        {
+            SelectTileRegion(sender, e);
+        }
+        private void BGTexture_PreviewMouseRightButtonDown(object sender, MouseButtonEventArgs e)
+        {
+            if (!SelectTileRegion(sender, e))
+                return;
 
-            ScrollViewer dataEditorViewer = (Application.Current.MainWindow as MainWindow).DataEditor.Parent as ScrollViewer;
-            double initOffset = dataEditorViewer.VerticalOffset;
+            UpdateLayout();
 
-            TileIdList.SelectedIndex = tileIndex;
-            (TileIdList.ItemContainerGenerator.ContainerFromIndex(tileIndex) as DataGridRow)?.MoveFocus(new TraversalRequest(FocusNavigationDirection.Next));
-
-            dataEditorViewer.UpdateLayout();
-            dataEditorViewer.ScrollToVerticalOffset(initOffset);
+            tileContextMenu.IsOpen = true;
         }
 
         private void DataUserControl_DataContextChanged(object sender, DependencyPropertyChangedEventArgs e)
         {
             if (IsLoaded)
-                MainWindow.FindVisualChild<ScrollViewer>(TileIdList).ScrollToVerticalOffset(0);
+                MainWindow.FindVisualChild<ScrollViewer>(TileIdList)?.ScrollToTop();
         }
     }
 }
