@@ -8,6 +8,7 @@ using ICSharpCode.AvalonEdit.Rendering;
 using ICSharpCode.AvalonEdit.Search;
 using Newtonsoft.Json;
 using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Globalization;
@@ -20,6 +21,7 @@ using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 using System.Windows;
 using System.Windows.Controls;
+using System.Windows.Controls.Primitives;
 using System.Windows.Data;
 using System.Windows.Documents;
 using System.Windows.Input;
@@ -33,6 +35,7 @@ using UndertaleModLib.Compiler;
 using UndertaleModLib.Decompiler;
 using UndertaleModLib.Models;
 using static UndertaleModTool.MainWindow.CodeEditorMode;
+using Input = System.Windows.Input;
 
 namespace UndertaleModTool
 {
@@ -56,14 +59,21 @@ namespace UndertaleModTool
         public bool DecompiledYet = false;
         public bool DecompiledSkipped = false;
         public SearchPanel DecompiledSearchPanel;
+        public static (int Line, int Column, double ScrollPos) OverriddenDecompPos;
 
         public bool DisassemblyFocused = false;
         public bool DisassemblyChanged = false;
         public bool DisassembledYet = false;
         public bool DisassemblySkipped = false;
         public SearchPanel DisassemblySearchPanel;
+        public static (int Line, int Column, double ScrollPos) OverriddenDisasmPos;
 
         public static RoutedUICommand Compile = new RoutedUICommand("Compile code", "Compile", typeof(UndertaleCodeEditor));
+
+        private static readonly Dictionary<string, UndertaleNamedResource> NamedObjDict = new();
+        private static readonly Dictionary<string, UndertaleNamedResource> ScriptsDict = new();
+        private static readonly Dictionary<string, UndertaleNamedResource> FunctionsDict = new();
+        private static readonly Dictionary<string, UndertaleNamedResource> CodeDict = new();
 
         public UndertaleCodeEditor()
         {
@@ -98,7 +108,7 @@ namespace UndertaleModTool
                     // says, this isn't working so it's just
                     // hardcoded in the XML for now
                     /*
-                    if(mainWindow.Data.GMS2_3)
+                    if(mainWindow.Data.IsVersionAtLeast(2, 3))
                     {
                         HighlightingColor color = null;
                         foreach (var rule in def.MainRuleSet.Rules)
@@ -131,12 +141,13 @@ namespace UndertaleModTool
 
             DecompiledEditor.Options.ConvertTabsToSpaces = true;
 
-            DecompiledEditor.TextArea.TextView.ElementGenerators.Add(new NumberGenerator());
-            DecompiledEditor.TextArea.TextView.ElementGenerators.Add(new NameGenerator());
+            TextArea textArea = DecompiledEditor.TextArea;
+            textArea.TextView.ElementGenerators.Add(new NumberGenerator(this, textArea));
+            textArea.TextView.ElementGenerators.Add(new NameGenerator(this, textArea));
 
-            DecompiledEditor.TextArea.TextView.Options.HighlightCurrentLine = true;
-            DecompiledEditor.TextArea.TextView.CurrentLineBackground = new SolidColorBrush(Color.FromRgb(60, 60, 60));
-            DecompiledEditor.TextArea.TextView.CurrentLineBorder = new Pen() { Thickness = 0 };
+            textArea.TextView.Options.HighlightCurrentLine = true;
+            textArea.TextView.CurrentLineBackground = new SolidColorBrush(Color.FromRgb(60, 60, 60));
+            textArea.TextView.CurrentLineBorder = new Pen() { Thickness = 0 };
 
             DecompiledEditor.Document.TextChanged += (s, e) =>
             {
@@ -144,10 +155,10 @@ namespace UndertaleModTool
                 DecompiledChanged = true;
             };
 
-            DecompiledEditor.TextArea.SelectionBrush = new SolidColorBrush(Color.FromRgb(100, 100, 100));
-            DecompiledEditor.TextArea.SelectionForeground = null;
-            DecompiledEditor.TextArea.SelectionBorder = null;
-            DecompiledEditor.TextArea.SelectionCornerRadius = 0;
+            textArea.SelectionBrush = new SolidColorBrush(Color.FromRgb(100, 100, 100));
+            textArea.SelectionForeground = null;
+            textArea.SelectionBorder = null;
+            textArea.SelectionCornerRadius = 0;
 
             // Disassembly editor styling and functionality
             DisassemblySearchPanel = SearchPanel.Install(DisassemblyEditor.TextArea);
@@ -162,18 +173,25 @@ namespace UndertaleModTool
                 }
             }
 
-            DisassemblyEditor.TextArea.TextView.ElementGenerators.Add(new NameGenerator());
+            textArea = DisassemblyEditor.TextArea;
+            textArea.TextView.ElementGenerators.Add(new NameGenerator(this, textArea));
 
-            DisassemblyEditor.TextArea.TextView.Options.HighlightCurrentLine = true;
-            DisassemblyEditor.TextArea.TextView.CurrentLineBackground = new SolidColorBrush(Color.FromRgb(60, 60, 60));
-            DisassemblyEditor.TextArea.TextView.CurrentLineBorder = new Pen() { Thickness = 0 };
+            textArea.TextView.Options.HighlightCurrentLine = true;
+            textArea.TextView.CurrentLineBackground = new SolidColorBrush(Color.FromRgb(60, 60, 60));
+            textArea.TextView.CurrentLineBorder = new Pen() { Thickness = 0 };
 
             DisassemblyEditor.Document.TextChanged += (s, e) => DisassemblyChanged = true;
 
-            DisassemblyEditor.TextArea.SelectionBrush = new SolidColorBrush(Color.FromRgb(100, 100, 100));
-            DisassemblyEditor.TextArea.SelectionForeground = null;
-            DisassemblyEditor.TextArea.SelectionBorder = null;
-            DisassemblyEditor.TextArea.SelectionCornerRadius = 0;
+            textArea.SelectionBrush = new SolidColorBrush(Color.FromRgb(100, 100, 100));
+            textArea.SelectionForeground = null;
+            textArea.SelectionBorder = null;
+            textArea.SelectionCornerRadius = 0;
+        }
+
+        private void UndertaleCodeEditor_Unloaded(object sender, RoutedEventArgs e)
+        {
+            OverriddenDecompPos = default;
+            OverriddenDisasmPos = default;
         }
 
         private void SearchPanel_LostFocus(object sender, RoutedEventArgs e)
@@ -197,6 +215,35 @@ namespace UndertaleModTool
             noMatchesTT.IsOpen = false;
         }
 
+        private void UndertaleCodeEditor_Loaded(object sender, RoutedEventArgs e)
+        {
+            FillInCodeViewer();
+        }
+        private void FillInCodeViewer(bool overrideFirst = false)
+        {
+            UndertaleCode code = DataContext as UndertaleCode;
+            if (DisassemblyTab.IsSelected && code != CurrentDisassembled)
+            {
+                if (!overrideFirst)
+                {
+                    DisassembleCode(code, !DisassembledYet);
+                    DisassembledYet = true;
+                }
+                else
+                    DisassembleCode(code, true);
+            }
+            if (DecompiledTab.IsSelected && code != CurrentDecompiled)
+            {
+                if (!overrideFirst)
+                {
+                    _ = DecompileCode(code, !DecompiledYet);
+                    DecompiledYet = true;
+                }
+                else
+                    _ = DecompileCode(code, true);
+            }
+        }
+
         private async void TabControl_SelectionChanged(object sender, SelectionChangedEventArgs e)
         {
             UndertaleCode code = this.DataContext as UndertaleCode;
@@ -204,20 +251,21 @@ namespace UndertaleModTool
             Directory.CreateDirectory(TempPath);
             if (code == null)
                 return;
+
             DecompiledSearchPanel.Close();
             DisassemblySearchPanel.Close();
+
             await DecompiledLostFocusBody(sender, null);
             DisassemblyEditor_LostFocus(sender, null);
-            if (DisassemblyTab.IsSelected && code != CurrentDisassembled)
+
+            if (!IsLoaded)
             {
-                DisassembleCode(code, !DisassembledYet);
-                DisassembledYet = true;
+                // If it's not loaded, then "FillInCodeViewer()" will be executed on load.
+                // This prevents a bug with freezing on code opening.
+                return;
             }
-            if (DecompiledTab.IsSelected && code != CurrentDecompiled)
-            {
-                _ = DecompileCode(code, !DecompiledYet);
-                DecompiledYet = true;
-            }
+
+            FillInCodeViewer();
         }
 
         private async void UserControl_DataContextChanged(object sender, DependencyPropertyChangedEventArgs e)
@@ -226,13 +274,14 @@ namespace UndertaleModTool
             if (code == null)
                 return;
 
+            FillObjectDicts();
+
             // compile/disassemble previously edited code (save changes)
             if (DecompiledTab.IsSelected && DecompiledFocused && DecompiledChanged &&
                 CurrentDecompiled is not null && CurrentDecompiled != code)
             {
                 DecompiledSkipped = true;
-                DecompiledEditor_LostFocus(sender, null);
-
+                await DecompiledLostFocusBody(sender, null);
             }
             else if (DisassemblyTab.IsSelected && DisassemblyFocused && DisassemblyChanged &&
                      CurrentDisassembled is not null && CurrentDisassembled != code)
@@ -241,8 +290,13 @@ namespace UndertaleModTool
                 DisassemblyEditor_LostFocus(sender, null);
             }
 
-            DecompiledEditor_LostFocus(sender, null);
+            await DecompiledLostFocusBody(sender, null);
             DisassemblyEditor_LostFocus(sender, null);
+
+            DecompiledYet = false;
+            DisassembledYet = false;
+            CurrentDecompiled = null;
+            CurrentDisassembled = null;
 
             if (MainWindow.CodeEditorDecompile != Unstated) //if opened from the code search results "link"
             {
@@ -265,16 +319,7 @@ namespace UndertaleModTool
                 MainWindow.CodeEditorDecompile = Unstated;
             }
             else
-            {
-                if (DisassemblyTab.IsSelected && code != CurrentDisassembled)
-                {
-                    DisassembleCode(code, true);
-                }
-                if (DecompiledTab.IsSelected && code != CurrentDecompiled)
-                {
-                    _ = DecompileCode(code, true);
-                }
-            }
+                FillInCodeViewer(true);
         }
 
         public static readonly RoutedEvent CtrlKEvent = EventManager.RegisterRoutedEvent(
@@ -303,11 +348,127 @@ namespace UndertaleModTool
             await CompileCommandBody(null, null);
         }
 
+        public void RestoreState(CodeTabState tabState)
+        {
+            if (tabState.IsDecompiledOpen)
+                CodeModeTabs.SelectedItem = DecompiledTab;
+            else
+                CodeModeTabs.SelectedItem = DisassemblyTab;
+
+            TextEditor textEditor = DecompiledEditor;
+            (int linePos, int columnPos, double scrollPos) = tabState.DecompiledCodePosition;
+            RestoreCaretPosition(textEditor, linePos, columnPos, scrollPos);
+
+            textEditor = DisassemblyEditor;
+            (linePos, columnPos, scrollPos) = tabState.DisassemblyCodePosition;
+            RestoreCaretPosition(textEditor, linePos, columnPos, scrollPos);
+        }
+        private static void RestoreCaretPosition(TextEditor textEditor, int linePos, int columnPos, double scrollPos)
+        {
+            if (linePos <= textEditor.LineCount)
+            {
+                int lineLen = textEditor.Document.GetLineByNumber(linePos).Length;
+                textEditor.TextArea.Caret.Line = linePos;
+                if (columnPos != -1)
+                    textEditor.TextArea.Caret.Column = columnPos;
+                else
+                    textEditor.TextArea.Caret.Column = lineLen + 1;
+
+                textEditor.ScrollToLine(linePos);
+                textEditor.ScrollToVerticalOffset(scrollPos);
+            }
+            else
+            {
+                textEditor.CaretOffset = textEditor.Text.Length;
+                textEditor.ScrollToEnd();
+            }
+        }
+        
+        private static void FillObjectDicts()
+        {
+            var data = mainWindow.Data;
+            var objLists = new IEnumerable[] {
+                data.Sounds,
+                data.Sprites,
+                data.Backgrounds,
+                data.Paths,
+                data.Scripts,
+                data.Fonts,
+                data.GameObjects,
+                data.Rooms,
+                data.Extensions,
+                data.Shaders,
+                data.Timelines,
+                data.AnimationCurves,
+                data.Sequences,
+                data.AudioGroups
+            };
+
+            NamedObjDict.Clear();
+            ScriptsDict.Clear();
+            FunctionsDict.Clear();
+            CodeDict.Clear();
+
+            foreach (var list in objLists)
+            {
+                if (list is null)
+                    continue;
+
+                foreach (var obj in list)
+                {
+                    if (obj is not UndertaleNamedResource namedObj)
+                        return;
+
+                    NamedObjDict[namedObj.Name.Content] = namedObj;
+                }
+            }
+            foreach (var scr in data.Scripts)
+            {
+                if (scr is null)
+                    continue;
+
+                ScriptsDict[scr.Name.Content] = scr;
+            }
+            foreach (var func in data.Functions)
+            {
+                if (func is null)
+                    continue;
+
+                FunctionsDict[func.Name.Content] = func;
+            }
+            foreach (var code in data.Code)
+            {
+                if (code is null)
+                    continue;
+
+                CodeDict[code.Name.Content] = code;
+            }
+        }
+
         private void DisassembleCode(UndertaleCode code, bool first)
         {
             code.UpdateAddresses();
 
             string text;
+
+            int currLine = 1;
+            int currColumn = 1;
+            double scrollPos = 0;
+            if (!first)
+            {
+                var caret = DisassemblyEditor.TextArea.Caret;
+                currLine = caret.Line;
+                currColumn = caret.Column;
+                scrollPos = DisassemblyEditor.VerticalOffset;
+            }
+            else if (OverriddenDisasmPos != default)
+            {
+                currLine = OverriddenDisasmPos.Line;
+                currColumn = OverriddenDisasmPos.Column;
+                scrollPos = OverriddenDisasmPos.ScrollPos;
+
+                OverriddenDisasmPos = default;
+            }
 
             DisassemblyEditor.TextArea.ClearSelection();
             if (code.ParentEntry != null)
@@ -319,14 +480,29 @@ namespace UndertaleModTool
             {
                 DisassemblyEditor.IsReadOnly = false;
 
-                var data = mainWindow.Data;
-                text = code.Disassemble(data.Variables, data.CodeLocals.For(code));
+                try
+                {
+                    var data = mainWindow.Data;
+                    text = code.Disassemble(data.Variables, data.CodeLocals.For(code));
 
-                CurrentLocals = new List<string>();
+                    CurrentLocals = new List<string>();
+                }
+                catch (Exception ex)
+                {
+                    DisassemblyEditor.IsReadOnly = true;
+
+                    string exStr = ex.ToString();
+                    exStr = String.Join("\n;", exStr.Split('\n'));
+                    text = $";  EXCEPTION!\n;   {exStr}\n";
+                }
             }
 
             DisassemblyEditor.Document.BeginUpdate();
             DisassemblyEditor.Document.Text = text;
+
+            if (!DisassemblyEditor.IsReadOnly)
+                RestoreCaretPosition(DisassemblyEditor, currLine, currColumn, scrollPos);
+
             DisassemblyEditor.Document.EndUpdate();
 
             if (first)
@@ -358,7 +534,7 @@ namespace UndertaleModTool
                     decompilationOutput = Decompiler.Decompile(gettextCode, new GlobalDecompileContext(null, false)).Replace("\r\n", "\n").Split('\n');
                 }
             }
-            Regex textdataRegex = new Regex("^ds_map_add\\(global\\.text_data_en, \\\"(.*)\\\", \\\"(.*)\\\"\\)");
+            Regex textdataRegex = new Regex("^ds_map_add\\(global\\.text_data_en, \\\"(.*)\\\", \\\"(.*)\\\"\\)", RegexOptions.Compiled);
             foreach (var line in decompilationOutput)
             {
                 Match m = textdataRegex.Match(line);
@@ -398,7 +574,28 @@ namespace UndertaleModTool
         private async Task DecompileCode(UndertaleCode code, bool first, LoaderDialog existingDialog = null)
         {
             DecompiledEditor.IsReadOnly = true;
+
+            int currLine = 1;
+            int currColumn = 1;
+            double scrollPos = 0;
+            if (!first)
+            {
+                var caret = DecompiledEditor.TextArea.Caret;
+                currLine = caret.Line;
+                currColumn = caret.Column;
+                scrollPos = DecompiledEditor.VerticalOffset;
+            }
+            else if (OverriddenDecompPos != default)
+            {
+                currLine = OverriddenDecompPos.Line;
+                currColumn = OverriddenDecompPos.Column;
+                scrollPos = OverriddenDecompPos.ScrollPos;
+
+                OverriddenDecompPos = default;
+            }
+
             DecompiledEditor.TextArea.ClearSelection();
+
             if (code.ParentEntry != null)
             {
                 DecompiledEditor.Text = "// This code entry is a reference to an anonymous function within " + code.ParentEntry.Name.Content + ", view it there";
@@ -411,7 +608,7 @@ namespace UndertaleModTool
                 if (existingDialog != null)
                 {
                     dialog = existingDialog;
-                    dialog.Message = "Decompiling, please wait...";
+                    dialog.Message = "Decompiling, please wait... This can take a while on complex scripts.";
                 }
                 else
                 {
@@ -453,7 +650,7 @@ namespace UndertaleModTool
                         string path = Path.Combine(TempPath, code.Name.Content + ".gml");
                         if (!SettingsWindow.ProfileModeEnabled || !File.Exists(path))
                         {
-                            decompiled = Decompiler.Decompile(code, context);
+                            decompiled = Decompiler.Decompile(code, context, (msg) => { dialog.Message = msg; });
                         }
                         else
                             decompiled = File.ReadAllText(path);
@@ -539,6 +736,8 @@ namespace UndertaleModTool
                                     CurrentLocals.Add(local.Name.Content);
                             }
 
+                            RestoreCaretPosition(DecompiledEditor, currLine, currColumn, scrollPos);
+
                             if (existingDialog is not null)                      //if code was edited (and compiles after it)
                             {
                                 dataa.GMLCacheChanged.Add(code.Name.Content);
@@ -547,10 +746,12 @@ namespace UndertaleModTool
                                 openSaveDialog = mainWindow.IsSaving;
                             }
                         }
+
                         DecompiledEditor.Document.EndUpdate();
                         DecompiledEditor.IsReadOnly = false;
                         if (first)
                             DecompiledEditor.Document.UndoStack.ClearAll();
+
                         DecompiledChanged = false;
 
                         CurrentDecompiled = code;
@@ -636,7 +837,18 @@ namespace UndertaleModTool
             var dispatcher = Dispatcher;
             Task t = Task.Run(() =>
             {
-                compileContext = Compiler.CompileGMLText(text, data, code, (f) => { dispatcher.Invoke(() => f()); });
+                try
+                {
+                    compileContext = Compiler.CompileGMLText(text, data, code, (f) => { dispatcher.Invoke(() => f()); });
+                }
+                catch (Exception ex)
+                {
+                    compileContext = new(data, code)
+                    {
+                        HasError = true,
+                        ResultError = ex.ToString()
+                    };
+                }
             });
             await t;
 
@@ -662,30 +874,26 @@ namespace UndertaleModTool
             }
 
             code.Replace(compileContext.ResultAssembly);
-
-            if (!mainWindow.Data.GMS2_3)
+            try
             {
-                try
+                string path = Path.Combine(TempPath, code.Name.Content + ".gml");
+                if (SettingsWindow.ProfileModeEnabled)
                 {
-                    string path = Path.Combine(TempPath, code.Name.Content + ".gml");
-                    if (SettingsWindow.ProfileModeEnabled)
-                    {
-                        // Write text, only if in the profile mode.
-                        File.WriteAllText(path, DecompiledEditor.Text);
-                    }
-                    else
-                    {
-                        // Destroy file with comments if it's been edited outside the profile mode.
-                        // We're dealing with the decompiled code only, it has to happen.
-                        // Otherwise it will cause a desync, which is more important to prevent.
-                        if (File.Exists(path))
-                            File.Delete(path);
-                    }
+                    // Write text, only if in the profile mode.
+                    File.WriteAllText(path, DecompiledEditor.Text);
                 }
-                catch (Exception exc)
+                else
                 {
-                    mainWindow.ShowError("Error during writing of GML code to profile:\n" + exc);
+                    // Destroy file with comments if it's been edited outside the profile mode.
+                    // We're dealing with the decompiled code only, it has to happen.
+                    // Otherwise it will cause a desync, which is more important to prevent.
+                    if (File.Exists(path))
+                        File.Delete(path);
                 }
+            }
+            catch (Exception exc)
+            {
+                mainWindow.ShowError("Error during writing of GML code to profile:\n" + exc);
             }
 
             // Invalidate gettext if necessary
@@ -793,22 +1001,51 @@ namespace UndertaleModTool
             }
         }
 
-        // Based on https://stackoverflow.com/questions/28379206/custom-hyperlinks-using-avalonedit
         public class NumberGenerator : VisualLineElementGenerator
         {
-            readonly static Regex regex = new Regex(@"-?\d+\.?");
+            private readonly IHighlighter highlighterInst;
+            private readonly UndertaleCodeEditor codeEditorInst;
 
-            public NumberGenerator()
+            // <offset, length>
+            private readonly Dictionary<int, int> lineNumberSections = new();
+
+            public NumberGenerator(UndertaleCodeEditor codeEditorInst, TextArea textAreaInst)
             {
+                this.codeEditorInst = codeEditorInst;
+
+                highlighterInst = textAreaInst.GetService(typeof(IHighlighter)) as IHighlighter;
             }
 
-            Match FindMatch(int startOffset, Regex r)
+            public override void StartGeneration(ITextRunConstructionContext context)
             {
-                // fetch the end offset of the VisualLine being generated
-                int endOffset = CurrentContext.VisualLine.LastDocumentLine.EndOffset;
-                TextDocument document = CurrentContext.Document;
-                string relevantText = document.GetText(startOffset, endOffset - startOffset);
-                return r.Match(relevantText);
+                lineNumberSections.Clear();
+
+                var docLine = context.VisualLine.FirstDocumentLine;
+                if (docLine.Length != 0)
+                {
+                    int line = docLine.LineNumber;
+                    var highlighter = highlighterInst;
+                    
+                    HighlightedLine highlighted;
+                    try
+                    {
+                        highlighted = highlighter.HighlightLine(line);
+                    }
+                    catch
+                    {
+                        Debug.WriteLine($"(NumberGenerator) Code editor line {line} highlight error.");
+                        base.StartGeneration(context);
+                        return;
+                    }
+
+                    foreach (var section in highlighted.Sections)
+                    {
+                        if (section.Color.Name == "Number")
+                            lineNumberSections[section.Offset] = section.Length;
+                    }
+                }
+
+                base.StartGeneration(context);
             }
 
             /// Gets the first offset >= startOffset where the generator wants to construct
@@ -816,39 +1053,10 @@ namespace UndertaleModTool
             /// Return -1 to signal no interest.
             public override int GetFirstInterestedOffset(int startOffset)
             {
-                Match m = FindMatch(startOffset, regex);
-
-                var textArea = CurrentContext.TextView.GetService(typeof(TextArea)) as TextArea;
-                var highlighter = textArea.GetService(typeof(IHighlighter)) as IHighlighter;
-                int line = CurrentContext.Document.GetLocation(startOffset).Line;
-                HighlightedLine highlighted = null;
-                try
+                foreach (var section in lineNumberSections)
                 {
-                    highlighted = highlighter.HighlightLine(line);
-                }
-                catch
-                {
-                }
-
-                while (m.Success)
-                {
-                    int res = startOffset + m.Index;
-                    int currLine = CurrentContext.Document.GetLocation(res).Line;
-                    if (currLine != line)
-                    {
-                        line = currLine;
-                        highlighted = highlighter.HighlightLine(line);
-                    }
-
-                    foreach (var section in highlighted.Sections)
-                    {
-                        if (section.Color.Name == "Number" &&
-                            section.Offset == res)
-                            return res;
-                    }
-
-                    startOffset += m.Length;
-                    m = FindMatch(startOffset, regex);
+                    if (startOffset <= section.Key)
+                        return section.Key;
                 }
 
                 return -1;
@@ -858,132 +1066,192 @@ namespace UndertaleModTool
             /// May return null if no element should be constructed.
             public override VisualLineElement ConstructElement(int offset)
             {
-                Match m = FindMatch(offset, regex);
+                int numLength = -1;
+                if (!lineNumberSections.TryGetValue(offset, out numLength))
+                    return null;
 
-                if (m.Success && m.Index == 0)
+                var doc = CurrentContext.Document;
+                string numText = doc.GetText(offset, numLength); 
+
+                var line = new ClickVisualLineText(numText, CurrentContext.VisualLine, numLength);
+                
+                line.Clicked += (text, inNewTab) =>
                 {
-                    var line = new ClickVisualLineText(m.Value, CurrentContext.VisualLine, m.Length);
-                    var doc = CurrentContext.Document;
-                    var textArea = CurrentContext.TextView.GetService(typeof(TextArea)) as TextArea;
-                    var editor = textArea.GetService(typeof(TextEditor)) as TextEditor;
-                    var parent = VisualTreeHelper.GetParent(editor);
-                    do
+                    if (int.TryParse(text, out int id))
                     {
-                        if ((parent as FrameworkElement) is UserControl)
-                            break;
-                        parent = VisualTreeHelper.GetParent(parent);
-                    } while (parent != null);
-                    line.Clicked += (text) =>
-                    {
-                        if (text.EndsWith("."))
-                            return;
-                        if (int.TryParse(text, out int id))
+                        codeEditorInst.DecompiledFocused = true;
+                        UndertaleData data = mainWindow.Data;
+
+                        List<UndertaleObject> possibleObjects = new List<UndertaleObject>();
+                        if (id >= 0)
                         {
-                            (parent as UndertaleCodeEditor).DecompiledFocused = true;
-                            UndertaleData data = mainWindow.Data;
-
-                            List<UndertaleObject> possibleObjects = new List<UndertaleObject>();
-                            if (id >= 0)
-                            {
-                                if (id < data.Sprites.Count)
-                                    possibleObjects.Add(data.Sprites[id]);
-                                if (id < data.Rooms.Count)
-                                    possibleObjects.Add(data.Rooms[id]);
-                                if (id < data.GameObjects.Count)
-                                    possibleObjects.Add(data.GameObjects[id]);
-                                if (id < data.Backgrounds.Count)
-                                    possibleObjects.Add(data.Backgrounds[id]);
-                                if (id < data.Scripts.Count)
-                                    possibleObjects.Add(data.Scripts[id]);
-                                if (id < data.Paths.Count)
-                                    possibleObjects.Add(data.Paths[id]);
-                                if (id < data.Fonts.Count)
-                                    possibleObjects.Add(data.Fonts[id]);
-                                if (id < data.Sounds.Count)
-                                    possibleObjects.Add(data.Sounds[id]);
-                                if (id < data.Shaders.Count)
-                                    possibleObjects.Add(data.Shaders[id]);
-                                if (id < data.Timelines.Count)
-                                    possibleObjects.Add(data.Timelines[id]);
-                            }
-
-                            ContextMenu contextMenu = new ContextMenu();
-                            foreach (UndertaleObject obj in possibleObjects)
-                            {
-                                MenuItem item = new MenuItem();
-                                item.Header = obj.ToString().Replace("_", "__");
-                                item.Click += (sender2, ev2) =>
-                                {
-                                    if ((Keyboard.Modifiers & ModifierKeys.Shift) == ModifierKeys.Shift)
-                                        mainWindow.ChangeSelection(obj);
-                                    else
-                                    {
-                                        doc.Replace(line.ParentVisualLine.StartOffset + line.RelativeTextOffset,
-                                                    text.Length, (obj as UndertaleNamedResource).Name.Content, null);
-                                        (parent as UndertaleCodeEditor).DecompiledChanged = true;
-                                    }
-                                };
-                                contextMenu.Items.Add(item);
-                            }
-                            if (id > 0x00050000)
-                            {
-                                MenuItem item = new MenuItem();
-                                item.Header = "0x" + id.ToString("X6") + " (color)";
-                                item.Click += (sender2, ev2) =>
-                                {
-                                    if (!((Keyboard.Modifiers & ModifierKeys.Shift) == ModifierKeys.Shift))
-                                    {
-                                        doc.Replace(line.ParentVisualLine.StartOffset + line.RelativeTextOffset,
-                                                    text.Length, "0x" + id.ToString("X6"), null);
-                                        (parent as UndertaleCodeEditor).DecompiledChanged = true;
-                                    }
-                                };
-                                contextMenu.Items.Add(item);
-                            }
-                            BuiltinList list = mainWindow.Data.BuiltinList;
-                            var myKey = list.Constants.FirstOrDefault(x => x.Value == (double)id).Key;
-                            if (myKey != null)
-                            {
-                                MenuItem item = new MenuItem();
-                                item.Header = myKey.Replace("_", "__") + " (constant)";
-                                item.Click += (sender2, ev2) =>
-                                {
-                                    if (!((Keyboard.Modifiers & ModifierKeys.Shift) == ModifierKeys.Shift))
-                                    {
-                                        doc.Replace(line.ParentVisualLine.StartOffset + line.RelativeTextOffset,
-                                                    text.Length, myKey, null);
-                                        (parent as UndertaleCodeEditor).DecompiledChanged = true;
-                                    }
-                                };
-                                contextMenu.Items.Add(item);
-                            }
-                            contextMenu.Items.Add(new MenuItem() { Header = id + " (number)", IsEnabled = false });
-
-                            contextMenu.IsOpen = true;
+                            if (id < data.Sprites.Count)
+                                possibleObjects.Add(data.Sprites[id]);
+                            if (id < data.Rooms.Count)
+                                possibleObjects.Add(data.Rooms[id]);
+                            if (id < data.GameObjects.Count)
+                                possibleObjects.Add(data.GameObjects[id]);
+                            if (id < data.Backgrounds.Count)
+                                possibleObjects.Add(data.Backgrounds[id]);
+                            if (id < data.Scripts.Count)
+                                possibleObjects.Add(data.Scripts[id]);
+                            if (id < data.Paths.Count)
+                                possibleObjects.Add(data.Paths[id]);
+                            if (id < data.Fonts.Count)
+                                possibleObjects.Add(data.Fonts[id]);
+                            if (id < data.Sounds.Count)
+                                possibleObjects.Add(data.Sounds[id]);
+                            if (id < data.Shaders.Count)
+                                possibleObjects.Add(data.Shaders[id]);
+                            if (id < data.Timelines.Count)
+                                possibleObjects.Add(data.Timelines[id]);
+                            if (id < (data.AnimationCurves?.Count ?? 0))
+                                possibleObjects.Add(data.AnimationCurves[id]);
+                            if (id < (data.Sequences?.Count ?? 0))
+                                possibleObjects.Add(data.Sequences[id]);
+                            if (id < (data.ParticleSystems?.Count ?? 0))
+                                possibleObjects.Add(data.ParticleSystems[id]);
                         }
-                    };
-                    return line;
-                }
 
-                return null;
+                        ContextMenuDark contextMenu = new();
+                        foreach (UndertaleObject obj in possibleObjects)
+                        {
+                            MenuItemDark item = new();
+                            item.Header = obj.ToString().Replace("_", "__");
+                            item.PreviewMouseDown += (sender2, ev2) =>
+                            {
+                                if (ev2.ChangedButton != Input.MouseButton.Left
+                                    && ev2.ChangedButton != Input.MouseButton.Middle)
+                                    return;
+
+                                if (ev2.ChangedButton == Input.MouseButton.Middle)
+                                {
+                                    mainWindow.Focus();
+                                    mainWindow.ChangeSelection(obj, true);
+                                    
+                                }
+                                else if ((Keyboard.Modifiers & ModifierKeys.Shift) == ModifierKeys.Shift)
+                                    mainWindow.ChangeSelection(obj);
+                                else
+                                {
+                                    doc.Replace(line.ParentVisualLine.StartOffset + line.RelativeTextOffset,
+                                                text.Length, (obj as UndertaleNamedResource).Name.Content, null);
+                                    codeEditorInst.DecompiledChanged = true;
+                                }
+                            };
+                            contextMenu.Items.Add(item);
+                        }
+                        if (id > 0x00050000)
+                        {
+                            MenuItemDark item = new();
+                            item.Header = "0x" + id.ToString("X6") + " (color)";
+                            item.Click += (sender2, ev2) =>
+                            {
+                                if (!((Keyboard.Modifiers & ModifierKeys.Shift) == ModifierKeys.Shift))
+                                {
+                                    doc.Replace(line.ParentVisualLine.StartOffset + line.RelativeTextOffset,
+                                                text.Length, "0x" + id.ToString("X6"), null);
+                                    codeEditorInst.DecompiledChanged = true;
+                                }
+                            };
+                            contextMenu.Items.Add(item);
+                        }
+                        BuiltinList list = mainWindow.Data.BuiltinList;
+                        var myKey = list.Constants.FirstOrDefault(x => x.Value == (double)id).Key;
+                        if (myKey != null)
+                        {
+                            MenuItemDark item = new();
+                            item.Header = myKey.Replace("_", "__") + " (constant)";
+                            item.Click += (sender2, ev2) =>
+                            {
+                                if (!((Keyboard.Modifiers & ModifierKeys.Shift) == ModifierKeys.Shift))
+                                {
+                                    doc.Replace(line.ParentVisualLine.StartOffset + line.RelativeTextOffset,
+                                                text.Length, myKey, null);
+                                    codeEditorInst.DecompiledChanged = true;
+                                }
+                            };
+                            contextMenu.Items.Add(item);
+                        }
+                        contextMenu.Items.Add(new MenuItemDark() { Header = id + " (number)", IsEnabled = false });
+
+                        contextMenu.IsOpen = true;
+                    }
+                };
+
+                return line;
             }
         }
 
         public class NameGenerator : VisualLineElementGenerator
         {
-            readonly static Regex regex = new Regex(@"[_a-zA-Z][_a-zA-Z0-9]*");
+            private readonly IHighlighter highlighterInst;
+            private readonly TextEditor textEditorInst;
+            private readonly UndertaleCodeEditor codeEditorInst;
 
-            public NameGenerator()
+            private static readonly SolidColorBrush FunctionBrush = new(Color.FromRgb(0xFF, 0xB8, 0x71));
+            private static readonly SolidColorBrush GlobalBrush = new(Color.FromRgb(0xF9, 0x7B, 0xF9));
+            private static readonly SolidColorBrush ConstantBrush = new(Color.FromRgb(0xFF, 0x80, 0x80));
+            private static readonly SolidColorBrush InstanceBrush = new(Color.FromRgb(0x58, 0xE3, 0x5A));
+            private static readonly SolidColorBrush LocalBrush = new(Color.FromRgb(0xFF, 0xF8, 0x99));
+
+            private static ContextMenuDark contextMenu;
+
+            // <offset, length>
+            private readonly Dictionary<int, int> lineNameSections = new();
+
+            public NameGenerator(UndertaleCodeEditor codeEditorInst, TextArea textAreaInst)
             {
+                this.codeEditorInst = codeEditorInst;
+
+                highlighterInst = textAreaInst.GetService(typeof(IHighlighter)) as IHighlighter;
+                textEditorInst = textAreaInst.GetService(typeof(TextEditor)) as TextEditor;
+
+                var menuItem = new MenuItemDark()
+                {
+                    Header = "Open in new tab"
+                };
+                menuItem.Click += (sender, _) =>
+                {
+                    mainWindow.ChangeSelection((sender as FrameworkElement).DataContext, true);
+                };
+                contextMenu = new()
+                {
+                    Items = { menuItem },
+                    Placement = PlacementMode.MousePoint
+                };
             }
 
-            Match FindMatch(int startOffset, Regex r)
+            public override void StartGeneration(ITextRunConstructionContext context)
             {
-                // fetch the end offset of the VisualLine being generated
-                int endOffset = CurrentContext.VisualLine.LastDocumentLine.EndOffset;
-                TextDocument document = CurrentContext.Document;
-                string relevantText = document.GetText(startOffset, endOffset - startOffset);
-                return r.Match(relevantText);
+                lineNameSections.Clear();
+
+                var docLine = context.VisualLine.FirstDocumentLine;
+                if (docLine.Length != 0)
+                {
+                    int line = docLine.LineNumber;
+                    var highlighter = highlighterInst;
+
+                    HighlightedLine highlighted;
+                    try
+                    {
+                        highlighted = highlighter.HighlightLine(line);
+                    }
+                    catch
+                    {
+                        Debug.WriteLine($"(NameGenerator) Code editor line {line} highlight error.");
+                        base.StartGeneration(context);
+                        return;
+                    }
+
+                    foreach (var section in highlighted.Sections)
+                    {
+                        if (section.Color.Name == "Identifier" || section.Color.Name == "Function")
+                            lineNameSections[section.Offset] = section.Length;
+                    }
+                }
+
+                base.StartGeneration(context);
             }
 
             /// Gets the first offset >= startOffset where the generator wants to construct
@@ -991,42 +1259,12 @@ namespace UndertaleModTool
             /// Return -1 to signal no interest.
             public override int GetFirstInterestedOffset(int startOffset)
             {
-                Match m = FindMatch(startOffset, regex);
-
-                var textArea = CurrentContext.TextView.GetService(typeof(TextArea)) as TextArea;
-                var highlighter = textArea.GetService(typeof(IHighlighter)) as IHighlighter;
-                int line = CurrentContext.Document.GetLocation(startOffset).Line;
-                HighlightedLine highlighted = null;
-                try
+                foreach (var section in lineNameSections)
                 {
-                    highlighted = highlighter.HighlightLine(line);
-                }
-                catch
-                {
+                    if (startOffset <= section.Key)
+                        return section.Key;
                 }
 
-                while (m.Success)
-                {
-                    int res = startOffset + m.Index;
-                    int currLine = CurrentContext.Document.GetLocation(res).Line;
-                    if (currLine != line)
-                    {
-                        line = currLine;
-                        highlighted = highlighter.HighlightLine(line);
-                    }
-
-                    foreach (var section in highlighted.Sections)
-                    {
-                        if (section.Color.Name == "Identifier" || section.Color.Name == "Function")
-                        {
-                            if (section.Offset == res)
-                                return res;
-                        }
-                    }
-
-                    startOffset += m.Length;
-                    m = FindMatch(startOffset, regex);
-                }
                 return -1;
             }
 
@@ -1034,106 +1272,106 @@ namespace UndertaleModTool
             /// May return null if no element should be constructed.
             public override VisualLineElement ConstructElement(int offset)
             {
-                Match m = FindMatch(offset, regex);
+                int nameLength = -1;
+                if (!lineNameSections.TryGetValue(offset, out nameLength))
+                    return null;
 
-                if (m.Success && m.Index == 0)
+                var doc = CurrentContext.Document;
+                string nameText = doc.GetText(offset, nameLength);
+
+                UndertaleData data = mainWindow.Data;
+                bool func = (offset + nameLength + 1 < CurrentContext.VisualLine.LastDocumentLine.EndOffset) &&
+                            (doc.GetCharAt(offset + nameLength) == '(');
+                UndertaleNamedResource val = null;
+
+                var editor = textEditorInst;
+
+                // Process the content of this identifier/function
+                if (func)
                 {
-                    UndertaleData data = mainWindow.Data;
-                    bool func = (offset + m.Length + 1 < CurrentContext.VisualLine.LastDocumentLine.EndOffset) &&
-                                (CurrentContext.Document.GetCharAt(offset + m.Length) == '(');
-                    UndertaleNamedResource val = null;
-
-                    var doc = CurrentContext.Document;
-                    var textArea = CurrentContext.TextView.GetService(typeof(TextArea)) as TextArea;
-                    var editor = textArea.GetService(typeof(TextEditor)) as TextEditor;
-                    var parent = VisualTreeHelper.GetParent(editor);
-                    do
+                    val = null;
+                    if (!data.IsVersionAtLeast(2, 3)) // in GMS2.3 every custom "function" is in fact a member variable and scripts are never referenced directly
+                        ScriptsDict.TryGetValue(nameText, out val);
+                    if (val == null)
                     {
-                        if ((parent as FrameworkElement) is UserControl)
-                            break;
-                        parent = VisualTreeHelper.GetParent(parent);
-                    } while (parent != null);
-
-                    // Process the content of this identifier/function
-                    if (func)
-                    {
-                        val = null;
-                        if (!data.GMS2_3) // in GMS2.3 every custom "function" is in fact a member variable and scripts are never referenced directly
-                            val = data.Scripts.ByName(m.Value);
-                        if (val == null)
+                        FunctionsDict.TryGetValue(nameText, out val);
+                        if (data.IsVersionAtLeast(2, 3))
                         {
-                            val = data.Functions.ByName(m.Value);
-                            if (data.GMS2_3)
+                            if (val != null)
                             {
-                                if (val != null)
+                                if (CodeDict.TryGetValue(val.Name.Content, out _))
+                                    val = null; // in GMS2.3 every custom "function" is in fact a member variable, and the names in functions make no sense (they have the gml_Script_ prefix)
+                            }
+                            else
+                            {
+                                // Resolve 2.3 sub-functions for their parent entry
+                                if (data.KnownSubFunctions?.TryGetValue(nameText, out UndertaleFunction f) == true)
                                 {
-                                    if (data.Code.ByName(val.Name.Content) != null)
-                                        val = null; // in GMS2.3 every custom "function" is in fact a member variable, and the names in functions make no sense (they have the gml_Script_ prefix)
-                                }
-                                else
-                                {
-                                    // Resolve 2.3 sub-functions for their parent entry
-                                    UndertaleFunction f = null;
-                                    if (data.KnownSubFunctions?.TryGetValue(m.Value, out f) == true)
-                                        val = data.Scripts.ByName(f.Name.Content).Code?.ParentEntry;
+                                    ScriptsDict.TryGetValue(f.Name.Content, out val);
+                                    val = (val as UndertaleScript)?.Code?.ParentEntry;
                                 }
                             }
                         }
-                        if (val == null)
-                        {
-                            if (data.BuiltinList.Functions.ContainsKey(m.Value))
-                            {
-                                var res = new ColorVisualLineText(m.Value, CurrentContext.VisualLine, m.Length,
-                                                                  new SolidColorBrush(Color.FromRgb(0xFF, 0xB8, 0x71)));
-                                res.Bold = true;
-                                return res;
-                            }
-                        }
-                    }
-                    else
-                    {
-                        val = data.ByName(m.Value);
-                        if (data.GMS2_3 & val is UndertaleScript)
-                            val = null; // in GMS2.3 scripts are never referenced directly
                     }
                     if (val == null)
                     {
-                        if (offset >= 7)
+                        if (data.BuiltinList.Functions.ContainsKey(nameText))
                         {
-                            if (CurrentContext.Document.GetText(offset - 7, 7) == "global.")
-                            {
-                                return new ColorVisualLineText(m.Value, CurrentContext.VisualLine, m.Length,
-                                                                new SolidColorBrush(Color.FromRgb(0xF9, 0x7B, 0xF9)));
-                            }
+                            var res = new ColorVisualLineText(nameText, CurrentContext.VisualLine, nameLength,
+                                                              FunctionBrush);
+                            res.Bold = true;
+                            return res;
                         }
-                        if (data.BuiltinList.Constants.ContainsKey(m.Value))
-                            return new ColorVisualLineText(m.Value, CurrentContext.VisualLine, m.Length,
-                                                            new SolidColorBrush(Color.FromRgb(0xFF, 0x80, 0x80)));
-                        if (data.BuiltinList.GlobalNotArray.ContainsKey(m.Value) ||
-                            data.BuiltinList.Instance.ContainsKey(m.Value) ||
-                            data.BuiltinList.GlobalArray.ContainsKey(m.Value))
-                            return new ColorVisualLineText(m.Value, CurrentContext.VisualLine, m.Length,
-                                                            new SolidColorBrush(Color.FromRgb(0x58, 0xE3, 0x5A)));
-                        if ((parent as UndertaleCodeEditor).CurrentLocals.Contains(m.Value))
-                            return new ColorVisualLineText(m.Value, CurrentContext.VisualLine, m.Length,
-                                                            new SolidColorBrush(Color.FromRgb(0xFF, 0xF8, 0x99)));
-                        return null;
                     }
-
-                    var line = new ClickVisualLineText(m.Value, CurrentContext.VisualLine, m.Length,
-                                                        func ? new SolidColorBrush(Color.FromRgb(0xFF, 0xB8, 0x71)) :
-                                                               new SolidColorBrush(Color.FromRgb(0xFF, 0x80, 0x80)));
-                    if (func)
-                        line.Bold = true;
-                    line.Clicked += (text) =>
+                }
+                else
+                {
+                    NamedObjDict.TryGetValue(nameText, out val);
+                    if (data.IsVersionAtLeast(2, 3) & val is UndertaleScript)
+                        val = null; // in GMS2.3 scripts are never referenced directly
+                }
+                if (val == null)
+                {
+                    if (offset >= 7)
                     {
-                        mainWindow.ChangeSelection(val);
-                    };
-
-                    return line;
+                        if (doc.GetText(offset - 7, 7) == "global.")
+                        {
+                            return new ColorVisualLineText(nameText, CurrentContext.VisualLine, nameLength,
+                                                           GlobalBrush);
+                        }
+                    }
+                    if (data.BuiltinList.Constants.ContainsKey(nameText))
+                        return new ColorVisualLineText(nameText, CurrentContext.VisualLine, nameLength,
+                                                       ConstantBrush);
+                    if (data.BuiltinList.GlobalNotArray.ContainsKey(nameText) ||
+                        data.BuiltinList.Instance.ContainsKey(nameText) ||
+                        data.BuiltinList.GlobalArray.ContainsKey(nameText))
+                        return new ColorVisualLineText(nameText, CurrentContext.VisualLine, nameLength,
+                                                       InstanceBrush);
+                    if (codeEditorInst.CurrentLocals.Contains(nameText) == true)
+                        return new ColorVisualLineText(nameText, CurrentContext.VisualLine, nameLength,
+                                                       LocalBrush);
+                    return null;
                 }
 
-                return null;
+                var line = new ClickVisualLineText(nameText, CurrentContext.VisualLine, nameLength,
+                                                   func ? FunctionBrush : ConstantBrush);
+                if (func)
+                    line.Bold = true;
+                line.Clicked += async (text, button) =>
+                {
+                    await codeEditorInst?.SaveChanges();
+
+                    if (button == Input.MouseButton.Right)
+                    {
+                        contextMenu.DataContext = val;
+                        contextMenu.IsOpen = true;
+                    }
+                    else
+                        mainWindow.ChangeSelection(val, button == Input.MouseButton.Middle);
+                };
+
+                return line;
             }
         }
 
@@ -1174,7 +1412,7 @@ namespace UndertaleModTool
         public class ClickVisualLineText : VisualLineText
         {
 
-            public delegate void ClickHandler(string text);
+            public delegate void ClickHandler(string text, Input.MouseButton button);
 
             public event ClickHandler Clicked;
 
@@ -1226,12 +1464,12 @@ namespace UndertaleModTool
             {
                 if (e.Handled)
                     return;
-                if ((e.ChangedButton == System.Windows.Input.MouseButton.Left && LinkIsClickable()) ||
-                     e.ChangedButton == System.Windows.Input.MouseButton.Middle)
+                if ((e.ChangedButton == Input.MouseButton.Left && LinkIsClickable())
+                    || e.ChangedButton == Input.MouseButton.Middle || e.ChangedButton == Input.MouseButton.Right)
                 {
                     if (Clicked != null)
                     {
-                        Clicked(Text);
+                        Clicked(Text, e.ChangedButton);
                         e.Handled = true;
                     }
                 }
